% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_taxa--taxmap--internal.R
\name{all_functions}
\alias{all_functions}
\title{Get list of usable functions}
\usage{
all_functions()
}
\value{
vector
}
\description{
Returns the names of all functions that can be called from any environment
}
\keyword{internal}
