% Generated by roxygen2 (4.0.2): do not edit by hand
\name{append_to_each}
\alias{append_to_each}
\title{Append each data.frame in list}
\usage{
append_to_each(my_list, data)
}
\arguments{
\item{my_list}{(\code{list} of \code{data.frame})}

\item{data}{(named \code{list}) The column content to add. The name of each element
should match a column in the \code{data.frame}s in \code{my_list}. Each element should consist
of as many values as their are \code{data.frame}s in \code{my_list}.}
}
\description{
Add a row to each \code{data.frame} in a list.
}

