% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_plotting.R
\name{apply_color_scale}
\alias{apply_color_scale}
\title{Covert numbers to colors}
\usage{
apply_color_scale(values, color_series, no_color_in_palette = 1000)
}
\arguments{
\item{values}{(\code{numeric}) The numbers to represent as colors}

\item{color_series}{(\code{character}) Hex values or a character in \code{colors}}

\item{no_color_in_palette}{(\code{numeric} of length 1) The number of distinct colors to use.}
}
\value{
\code{character} Hex color codes.
}
\description{
Convert numbers to colors.
If colors are already supplied, return the input
}
\keyword{internal}

