% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sequence_download.R
\name{as.data.frame.qaw}
\alias{as.data.frame.qaw}
\title{Coerce seqinr::qaw to a Data Frame}
\usage{
\method{as.data.frame}{qaw}(x, ...)
}
\arguments{
\item{x}{An object of class \code{\link[seqinr]{qaw}}}
}
\value{
A data frame with columns corresponding to \code{\link[seqinr]{SeqAcnucWeb}} attributes.
}
\description{
Coerce an object of class \code{\link[seqinr]{qaw}} to a Data Frame.
}

