% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_taxa--taxmap--internal.R
\name{as_id}
\alias{as_id}
\title{Convert a vector to database IDs}
\usage{
as_id(ids, database, ...)
}
\arguments{
\item{ids}{The character or numeric vector of raw taxon IDs.}

\item{database}{The database format to convert the IDs to. Either ncbi,
itis, eol, col, tropicos, gbif, nbn, worms, natserv, bold, or wiki}

\item{...}{Passed to \code{as.*id} function.}
}
\description{
This is a convenience function to convert to identifiers of various  data
sources. It wraps the \code{as.*id} functions in \code{\link[taxize]{taxize}}
}
\keyword{internal}
