% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxmap--column_generators.R
\name{assigned}
\alias{assigned}
\title{Count observation assigned in \code{\link{taxmap}}}
\usage{
assigned(obj, subset = obj$taxon_data$taxon_ids)
}
\arguments{
\item{obj}{(\code{\link{taxmap}})}

\item{subset}{(\code{character}) The \code{taxon_ids}s to get counts for.}
}
\value{
\code{numeric}
}
\description{
Count observations assigned to a specific taxon in \code{\link{taxmap}}
}

