% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_documentation.R
\docType{data}
\name{bryophytes_ex_data}
\alias{bryophytes_ex_data}
\title{Example dataset of bryophytes}
\format{An object of type \code{\link{taxmap}}}
\usage{
bryophytes_ex_data
}
\description{
A dataset containing information from 171 bryophytes family names scraped from \url{http://www.theplantlist.org/1.1/browse/B/}:
}
\examples{
\dontrun{

library(XML)
library(taxa) # The parsers in taxa are used
taxon_names <- XML::htmlTreeParse("http://www.theplantlist.org/1.1/browse/B/") \%>\% 
  xmlRoot() \%>\%
  getNodeSet("//ul[@id='nametree']/li/a/i") \%>\%
  sapply(xmlValue)

bryophytes_ex_data <- lookup_tax_data(taxon_names,
                                      type = "taxon_name", 
                                      database = "itis")
}

}
\seealso{
Other example_datasets: \code{\link{contaminants}},
  \code{\link{genbank_ex_data}},
  \code{\link{its1_ex_data}}, \code{\link{pr2_ex_data}},
  \code{\link{silva_ex_data}},
  \code{\link{unite_ex_data_1}},
  \code{\link{unite_ex_data_2}}
}
\keyword{datasets}
