% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amplicon_extraction.R
\name{calc_amplicons}
\alias{calc_amplicons}
\title{Extract expected amplicons}
\usage{
calc_amplicons(seqs, forward, reverse, min_coverage = 0.9, ...)
}
\arguments{
\item{seqs}{The DNA sequences of the target to be amplified. Letter case is ignored but
preserved.}

\item{forward}{The DNA sequence of the forward primer in the 5' to 3' orientation. IUPAC codes
can be present.}

\item{reverse}{The DNA sequence of the forward primer in the 5' to 3' orientation. IUPAC codes
can be present.}

\item{max_mismatch}{The maximum number of mismatching bases allowed. See
[Biostrings::matchPattern] for more information.}

\item{min_mismatch}{The minimum number of mismatching bases allowed. See
[Biostrings::matchPattern] for more information.}

\item{allow_indels}{If TRUE then indels are allowed. See [Biostrings::matchPattern] for more
information.}

\item{algorithm}{One of: "auto", "naive-exact", "naive-inexact", "boyer-moore" or "shift-or". See
[Biostrings::matchPattern] for more information.}

\item{infer_amps}{If TRUE, performe an additional step to find amplicons in sequences that might
not contain the primer binding sites due to being produced with similar primers. This is done
by aligning amplicons found using primers to sequences not found to contain the primers. For
each sequence, the region aligned to the best-matching amplicon will be included as an amplion
if:
*  The aligned region extends to the boundry of the sequence, suggesting the missing primer site is not in the sequence. 
*  If the aligned region covers the minimum proportion of the best matching amplicon defined by `infer_thresh_cov`.
*  If the aligned region has the minimum percent identity to the best matching amplicon defined by `infer_thresh_pid`.}

\item{infer_thresh_cov}{See the `infer_amps` option description.}

\item{infer_thresh_pid}{See the `infer_amps` option description.}
}
\description{
Extract expected amplicons from sequences using primer sequences.
}
