% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{calc_obs_props}
\alias{calc_obs_props}
\title{Calculate proportions from observation counts}
\usage{
calc_obs_props(obj, dataset, cols = NULL, keep_other_cols = TRUE)
}
\arguments{
\item{obj}{A taxmap object}

\item{cols}{The names/indexes of columns in \code{data} that have counts. By Default,
all numeric columns in \code{data} are used.}

\item{keep_other_cols}{If \code{TRUE}, keep non-count cols in the input data.
The "taxon_id" column will always be preserved.}

\item{data}{The name of a table in \code{obj} that contains counts.}
}
\value{
A tibble
}
\description{
For a given table in a taxmap object, convert one or more columns containing
counts to proportions. This is meant to be used with counts associated with
observations (e.g. OTUs), as opposed to counts that have already been summed
per taxon. If counts have already been summed per taxon, use
[calc_tax_props()] instead.
}
