% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{calc_taxon_abund}
\alias{calc_taxon_abund}
\title{Sum observation values for each taxon}
\usage{
calc_taxon_abund(obj, dataset, cols = NULL, groups = NULL,
  out_names = NULL)
}
\arguments{
\item{obj}{A taxmap object}

\item{dataset}{The name of a table in \code{obj} that contains counts.}

\item{cols}{The names/indexes of columns in \code{data} that have counts. By 
Default, all numeric columns in \code{data} are used.}

\item{groups}{Group counts of multiple columns per treatment/group. This
should be a vector of group IDs (e.g. character, integer) the same
length as \code{cols} that defines which samples go in which group. When
used, there will be one column in the output for each unique value in
\code{groups}.}

\item{out_names}{The names of count columns in the output. Must be the same
length as \code{cols} (or \code{unique(groups)}, if used).}
}
\value{
A tibble
}
\description{
For a given table in a taxmap object, sum the values in each column for each 
taxon. This is useful to convert per-observation counts (e.g. OTU counts) to 
per-taxon counts.
}
\seealso{
Other calculations: \code{\link{calc_n_samples}},
  \code{\link{calc_obs_props}},
  \code{\link{compare_treatments}},
  \code{\link{rarefy_obs}}, \code{\link{zero_low_counts}}
}
