% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{can_be_num}
\alias{can_be_num}
\title{Test if characters can be converted to numbers}
\usage{
can_be_num(input)
}
\arguments{
\item{input}{A character vector}
}
\description{
Makes TRUE/FALSE vector
}
\keyword{internal}
