% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_taxa--taxmap--internal.R
\name{can_be_used_in_taxmap}
\alias{can_be_used_in_taxmap}
\title{Check that a unknown object can be used with taxmap}
\usage{
can_be_used_in_taxmap(obj)
}
\arguments{
\item{obj}{}
}
\value{
TRUE/FALSE
}
\description{
Check that a unknown object can be assigned taxon IDs and filtered.
}
\keyword{internal}
