% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_taxa--taxmap--parsers.R
\name{check_class_col}
\alias{check_class_col}
\title{Check for name/index in input data}
\usage{
check_class_col(tax_data, column)
}
\arguments{
\item{tax_data}{A table, list, or vector that contain sequence IDs, taxon
IDs, or taxon names.
* tables: The `column` option must be used to specify which column
contains the sequence IDs, taxon IDs, or taxon names.
* lists: There must be only one item per list entry unless the `column`
option is used to specify what item to use in each list entry.
* vectors: simply a vector of sequence IDs, taxon IDs, or taxon names.}

\item{column}{(`character` or `integer`) The name or index of the column that
contains information used to lookup classifications. This only applies when
a table or list is supplied to `tax_data`.}
}
\description{
Used by parse_tax_data and lookup_tax_data to check that columm/class_col is valid for the input data
}
\keyword{internal}
