% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new--heat_tree--validation.R
\name{check_color}
\alias{check_color}
\title{Check color arguments}
\usage{
check_color(args)
}
\arguments{
\item{args}{A list of arguments}
}
\value{
A named list of argument values, potentially modified
}
\description{
Check that arguments specifying color make sense. They must be either numbers or a character
representing a hex color code and a color name in colors()
}
\keyword{internal}
