% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tree_plotting.R
\name{check_element_length}
\alias{check_element_length}
\title{Check length of graph attributes}
\usage{
check_element_length(args)
}
\description{
Length should divind evenly into the number of taxon/parent IDs
}

