% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new--heat_tree--validation.R
\name{check_for_na}
\alias{check_for_na}
\title{Check for NAs}
\usage{
check_for_na(args, warn)
}
\arguments{
\item{args}{A list of arguments}

\item{warn}{If TRUE, warn the user about NAs in argument values}
}
\value{
A named list of argument values, potentially modified
}
\description{
Check for NAs in the values and IDs of parameters
}
\keyword{internal}
