% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new--heat_tree--validation.R
\name{check_intervals}
\alias{check_intervals}
\title{Verify interval parameters}
\usage{
check_intervals(args)
}
\arguments{
\item{args}{A list of arguments}
}
\value{
A named list of argument values, potentially modified
}
\description{
Verify interval parameters
}
\keyword{internal}
