% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new--heat_tree--validation.R
\name{check_labels}
\alias{check_labels}
\title{Verify label parameters}
\usage{
check_labels(args)
}
\arguments{
\item{args}{A list of arguments}
}
\value{
A named list of argument values, potentially modified
}
\description{
Verify label parameters
}
\keyword{internal}
