% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new--heat_tree--validation.R
\name{check_length_or_id}
\alias{check_length_or_id}
\title{Check for length or taxon id}
\usage{
check_length_or_id(obj, args)
}
\arguments{
\item{obj}{taxmap object}

\item{args}{A list of arguments}
}
\description{
Check that inputs are named by taxon ID or are the same length as the number of taxa.
Print a message if an argument does not have a taxon ID but is same length as the number of taxa.
}
\keyword{internal}
