% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations--internal.R
\name{check_option_groups}
\alias{check_option_groups}
\title{Check option: groups}
\usage{
check_option_groups(groups, cols = NULL)
}
\arguments{
\item{groups}{The groups option to check}

\item{cols}{The cols option, if applicable}
}
\description{
This option is used in a few of the calculation functions
}
\keyword{internal}
