% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new--heat_tree--validation.R
\name{check_shape}
\alias{check_shape}
\title{Check shape arguments}
\usage{
check_shape(shape_args, args, default_shapes)
}
\arguments{
\item{shape_args}{A list of arguments specifying shapes}

\item{args}{All args}

\item{default_shapes}{A character vector of built-in function names for shapes}
}
\value{
A named list of argument values, potentially modified
}
\description{
Check that arguments specifying shapes make sense.
}
\keyword{internal}
