% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new--heat_tree--validation.R
\name{check_shape_range}
\alias{check_shape_range}
\title{Verify shape range parameters}
\usage{
check_shape_range(args, valid_shapes)
}
\arguments{
\item{args}{A list of arguments}
}
\value{
A named list of argument values, potentially modified
}
\description{
Verify shape range parameters. These should be either character or named functions in a list.
}
\keyword{internal}
