% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new--heat_tree--validation.R
\name{check_size}
\alias{check_size}
\title{Check size arguments}
\usage{
check_size(args)
}
\arguments{
\item{args}{A list of arguments}
}
\value{
A named list of argument values, potentially modified
}
\description{
Check that arguments specifying size make sense. Parmeters specifying size must be numeric and
not NA and at least one must be finite. Infinite values will be converted to the max/min of finite values.
}
\keyword{internal}
