% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new--heat_tree--validation.R
\name{check_size_range}
\alias{check_size_range}
\title{Verify size range parameters}
\usage{
check_size_range(args)
}
\arguments{
\item{args}{A list of arguments}
}
\value{
A named list of argument values, potentially modified
}
\description{
Verify size range parameters
}
\keyword{internal}
