% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new--heat_tree--validation.R
\name{check_trans}
\alias{check_trans}
\title{Verify transformation parameters}
\usage{
check_trans(args)
}
\arguments{
\item{args}{A list of arguments}
}
\value{
A named list of argument values, potentially modified
}
\description{
Verify transformation parameters
}
\keyword{internal}
