% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_taxonomy--parsers.R
\name{class_from_name}
\alias{class_from_name}
\title{Retrieve classifications from taxon names}
\usage{
class_from_name(name, database, ...)
}
\arguments{
\item{name}{(\code{character})
The name of a taxon. Not necessarily unique, but are interpretable by a particular \code{database}.
Requires an internet connection.}

\item{database}{(\code{character} of length 1)
The name of the database that patterns given in  \code{parser} will apply to.
Valid databases include "ncbi", "itis", "eol", "col", "tropicos",
"nbn", and "none". \code{"none"} will cause no database to be quired; use this if you want to not use the
internet. NOTE: Only \code{"ncbi"} has been tested so far.}

\item{...}{Not used}
}
\value{
\code{list} of \code{data.frame}
}
\description{
Retrieve taxonomic classifications from taxon names using a specified database.
}
\keyword{internal}
