% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_taxonomy--parsers.R
\name{class_from_obs_id}
\alias{class_from_obs_id}
\title{Retrieve classifications from observation IDs}
\usage{
class_from_obs_id(obs_id, database = c("ncbi", "none"), batch_size = 100,
  ...)
}
\arguments{
\item{obs_id}{(\code{character})
An unique observation (e.g. sequence) identifier for a particular \code{database}.
Requires an internet connection.}

\item{database}{(\code{character} of length 1)
The name of the database that patterns given in  \code{parser} will apply to.
Valid databases include "ncbi", "itis", "eol", "col", "tropicos",
"nbn", and "none". \code{"none"} will cause no database to be quired; use this if you want to not use the
internet. NOTE: Only \code{"ncbi"} has been tested so far.}

\item{...}{Not used}
}
\value{
\code{list} of \code{data.frame}
}
\description{
Retrieve taxonomic classifications from observation (e.g. sequence) IDs using a specified database.
}
\keyword{internal}

