% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_taxonomy--parsers.R
\name{class_from_taxon_id}
\alias{class_from_taxon_id}
\title{Retrieve classifications from taxon IDs}
\usage{
class_from_taxon_id(taxon_id, database, ...)
}
\arguments{
\item{database}{(\code{character} of length 1)
The name of the database that patterns given in  \code{parser} will apply to.
Valid databases include "ncbi", "itis", "eol", "col", "tropicos",
"nbn", and "none". \code{"none"} will cause no database to be quired; use this if you want to not use the
internet. NOTE: Only \code{"ncbi"} has been tested so far.}

\item{...}{Not used}

\item{name}{(\code{character})
The name of a taxon. Not necessarily unique, but are interpretable by a particular \code{database}.
Requires an internet connection.}
}
\value{
\code{list} of \code{data.frame}
}
\description{
Retrieve taxonomic classifications from taxon IDs using a specified database.
}
\keyword{internal}

