% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classified--class.R
\name{classified}
\alias{classified}
\title{Create an instance of \code{classified}}
\usage{
classified(taxon_ids, parent_ids, taxa = taxon_ids,
  item_taxon_ids = numeric(0), taxon_data = NULL, item_data = NULL,
  taxon_funcs = list(item_counts = item_counts, taxon_ranks = taxon_ranks,
  classifications = classifications), item_funcs = list())
}
\arguments{
\item{taxon_ids}{(\code{character}) 
These are unique identifiers for taxa.
They will be coerced into characters.}

\item{parent_ids}{(\code{character} OR (\code{numeric}))
Parent taxa (i.e. supertaxa) of \code{taxa}.
If a \code{character} vector, then these should be in the same format as \code{taxon_ids}.
If a \code{numeric} vector, then it is interpreted as the indexes of \code{taxon_ids}.
Taxa without parents should be \code{NA}.}

\item{taxa}{(\code{character})
Objects representing taxa.
Currently, these can be anything, but this might change in the future.}

\item{item_taxon_ids}{(\code{character} OR (\code{numeric}))
Taxon assignments of items.
Parent taxa (i.e. supertaxa) of \code{taxa}.
If a \code{character} vector, then these should be in the same format as \code{taxon_ids}.
If a \code{numeric} vector, then it is interpreted as the indexes of \code{taxon_ids}.}

\item{taxon_data}{(\code{data.frame})
A table with rows pretaining to \code{taxa}}

\item{item_data}{A (\code{data.frame})
A table with rows pretaining to \code{item_taxa}}

\item{taxon_funcs}{(\code{list} of named \code{function}s)
These the values produced by these functions will be accessible as a column in \code{taxon_data}.
The first parameter of each function should be a single \code{classified} object.}

\item{item_funcs}{(\code{list} of named \code{function}s)
These the values produced by these functions will be accessible as a column in \code{item_data}.
The first parameter of each function should be a single \code{classified} object.}
}
\value{
An object of type \code{classified}
}
\description{
Create an instance of \code{classified} containing items classified by a taxonomy.
}

