% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{color_legend}
\alias{color_legend}
\title{Make legend}
\usage{
color_legend(colors, min_value, max_value, width, height,
  tick_height = height/100, tick_width = width/10, label_count = 7,
  color_count = 20)
}
\arguments{
\item{colors}{(\code{character}) Colors used in the gradient. Must be interpretable by
\code{\link{colorRampPalette}}.}

\item{min_value}{(\code{numeric}; length = 1) The minimum value on the scale that the colors
represent}

\item{max_value}{(\code{numeric}; length = 1) The maximum value on the scale that the colors
represent}

\item{width}{(\code{numeric}; length = 1) The width of the legend color gradient, not including
the ticks or labels.}

\item{tick_height}{(\code{numeric}; length = 1) The width of the ticks}

\item{tick_width}{(\code{numeric}; length = 1) The height (i.e. thickness) of the ticks}

\item{label_count}{(\code{numeric}; length = 1) The number of labels}

\item{color_count}{(\code{numeric}; length = 1) The number of distinct colors in the gradient}

\item{width}{(\code{numeric}; length = 1) The height of the legend color gradient}
}
\value{
A list of 1) a \code{data.frame} of color block coordinants that can be plotted with
ggplot2::geom_polygon, 2) label y, 3) label x.
}
\description{
Make legend coordinants for color gradient
}

