% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{compare_treatments}
\alias{compare_treatments}
\title{Compare treatments}
\usage{
compare_treatments(obj, dataset, sample_ids, treatments, func = NULL,
  combinations = NULL, keep_cols = TRUE)
}
\arguments{
\item{obj}{A taxmap object}

\item{dataset}{The name of a table in \code{obj} that contains data for each 
sample in columns.}

\item{sample_ids}{The names of sample columns in \code{dataset}}

\item{treatments}{The treatment associated with each sample. Must be the same
order and length as \code{sample_ids}.}

\item{func}{The function to apply for each comparison. For each row in 
\code{dataset}, for each combination of treatments, this function will 
recieve the data for each treatment, passed a two character vecotors.
Therefor the function must take at least 2 arguments corresponding to the
two treatments compared. The function should return a vector or list or
results of a fixed length. If named, the names will be used in the output.
The names should be consistent as well. A simple example is
\code{function(x, y) mean(x) - mean(y)}. By default, the following function
is used:
\preformatted{
function(abund_1, abund_2) {
  log_ratio <- log2(median(abund_1) / median(abund_2))
  if (is.nan(log_ratio)) {
    log_ratio <- 0
  }
  list(log2_median_ratio = log_ratio,
       median_diff = median(abund_1) - median(abund_2),
       mean_diff = mean(abund_1) - mean(abund_2),
       wilcox_p_value = wilcox.test(abund_1, abund_2)$p.value)
}
}}

\item{combinations}{Which combinations of treatments to use. Must be a list 
of vectors, each containing the names of 2 treatments to compare. By 
default, all pairwise combinations of treatments are compared.}

\item{keep_cols}{If \code{TRUE}, preserve all columns not in 
\code{sample_ids} in the output. If \code{FALSE}, dont keep other columns. 
If a column names or indexes are supplied, only preserve those columns.}
}
\value{
A tibble
}
\description{
Apply a function to compare data, usually abundance, from pairs of 
treatments. By default, every pairwise combination of treatments are 
compared. A custom function can be supplied to perform the comparison.
}
