% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{convert_base}
\alias{convert_base}
\title{Converts decimal numbers to other bases}
\usage{
convert_base(
  numbers,
  symbols = letters,
  base = length(symbols),
  min_length = 0
)
}
\arguments{
\item{numbers}{One or more numbers to convert.}

\item{symbols}{The set of symbols to use for the new base.}

\item{base}{The base to convert to.}

\item{min_length}{The minimum number of symbols in each result.}
}
\value{
character vector
}
\description{
Converts from base 10 to other bases represented by a given set of symbols.
}
\keyword{internal}
