% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_taxonomy--internal.R
\name{convert_numeric_cols}
\alias{convert_numeric_cols}
\title{Convert columns to numeric if appropriate}
\usage{
convert_numeric_cols(table, columns = colnames(table))
}
\arguments{
\item{table}{(\code{data.frame})}

\item{columns}{(\code{character} or \code{numeric}) A subset of columns to attempt to convert.}
}
\value{
\code{data.frame}
}
\description{
Converts columns in a \code{data.frame} to numeric is they can be converted without producing \code{NA}s.
}
\keyword{internal}
