% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_taxa--taxmap--parsers.R
\name{correct_taxon_names}
\alias{correct_taxon_names}
\title{Look up official names from potentially misspelled names}
\usage{
correct_taxon_names(names, database = "ncbi")
}
\arguments{
\item{names}{Potentially misspelled taxon names}

\item{database}{The database the names are being looked up for. If `NULL`, do
not consider database.}
}
\value{
vector of names
}
\description{
Look up official names from potentially misspelled names using Global Names
Resolver (GNR). If a result from the chosen database is present, then it is
used, otherwise the NCBI result is used and if that does not exist, then the
first result is used. Names with no match will return NA.
}
\keyword{internal}
