% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_taxa--taxmap--parsers.R
\name{count_capture_groups}
\alias{count_capture_groups}
\title{Count capture groups}
\source{
http://stackoverflow.com/questions/16046620/regex-to-count-the-number-of-capturing-groups-in-a-regex
}
\usage{
count_capture_groups(regex)
}
\arguments{
\item{regex}{(\code{character} of length 1)}
}
\value{
\code{numeric} of length 1
}
\description{
Count the number of capture groups in a regular expression.
}
\keyword{internal}
