% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_taxon_class.R
\name{count_items}
\alias{count_items}
\title{Count items in \code{\link{classified}}}
\usage{
count_items(obj)
}
\arguments{
\item{obj}{(\code{\link{classified}})}
}
\value{
\code{numeric} of length 1
}
\description{
Count items in \code{\link{classified}}
}

