% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_taxa--taxon_database.R
\docType{data}
\name{database_list}
\alias{database_list}
\title{Database list}
\format{
An object of class \code{list} of length 8.
}
\usage{
database_list
}
\description{
The list of known databases. Not currently used much, but will be when we add
more check for taxon IDs and taxon ranks from particular databases.
}
\details{
List of databases with pre-filled details, where each has the
format:
\itemize{
 \item url: A base URL for the database source.
 \item description: Description of the database source.
 \item id regex: identifier regex.
}
}
\examples{
database_list
database_list$ncbi
database_list$ncbi$name
database_list$ncbi$description
database_list$ncbi$url
}
\seealso{
[taxon_database]
}
\keyword{datasets}
