% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/internal.R
\name{delete_vetices_and_children}
\alias{delete_vetices_and_children}
\title{delete_vetices_and_children}
\usage{
delete_vetices_and_children(graph, vertices)
}
\description{
delete_vetices_and_children
}

