% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_taxa--taxmap--printers.R
\name{desc_font}
\alias{desc_font}
\title{Description formatting in print methods}
\usage{
desc_font(text)
}
\arguments{
\item{text}{What to print}
}
\description{
A simple wrapper to make changing the formatting of text printed easier.
This is used for non-data, formatting characters
}
\seealso{
Other printer fonts: 
\code{\link{error_font}()},
\code{\link{name_font}()},
\code{\link{punc_font}()},
\code{\link{tid_font}()}
}
\concept{printer fonts}
\keyword{internal}
