% Generated by roxygen2 (4.0.2): do not edit by hand
\name{DNAbin_to_char}
\alias{DNAbin_to_char}
\title{Converts DNAbin to a named character vector}
\usage{
DNAbin_to_char(dna_bin)
}
\arguments{
\item{dna_bin}{(\code{DNAbin} of length 1) the input.}
}
\description{
Converts an object of class DNAbin (as produced by ape) to a named character vector.
}

