% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sequence_download.R
\name{download_gb_taxon}
\alias{download_gb_taxon}
\title{Downloads sequences that result from a taxon and keyword search}
\usage{
download_gb_taxon(taxon, key, type, seq_length = c(1, 10000),
  max_count = 100, subsample = c("random", "head", "tail"),
  standardize = TRUE, separate = FALSE, use_acnuc = FALSE)
}
\arguments{
\item{taxon}{A character vector of taxa.}

\item{key}{A character vector of keywords. All keywords must be present for a given
sequence to be returned.}

\item{type}{The type of the sequence to return (e.g. RRNA). Use `getType()` to see options.}

\item{seq_length}{A numeric vector of length 2. The range of sequence lengths to allow.}

\item{max_count}{The maximum number of sequences to download. See note below for additional
considerations.}

\item{subsample}{A character vector of length 1. Specifies how to subsample if more search
results are found than `max_count`. "subsample": randomly select using `sample`; "head": use
first results; "tail": use last results.}

\item{standardize}{If TRUE, validate binomial taxon names using `taxize`.}

\item{separate}{If TRUE, each taxon supplied is searched for separately and the `max_count`
option is applied to each separately, therefore the resulting max count of all results would be
\code{max_count * length(taxon)}.}

\item{use_acnuc}{If TRUE, sequences are downloaded using tools from `sequinr`. This is typically
slower.}
}
\description{
This function is meant to download sequences associated with a taxon and a set of keywords from
Genbank using a simple interface. More complicated queries should be done using the tools of the
`sequnir` and `taxize` packages, as this function does.
}
\details{
This function first searches genbank via ACNUC and finds all sequences of a given taxon
with given keywords (e.g. genes). It then subsamples these search results if there are too many
and downloads the subsample sequences. The results are compiled into a data frame and
returned.
}
\note{
When `use_acnuc = FALSE`, some sequences that are found during searching with `seqinr` are
  not found when downloading using `taxize`. This is because `seqinr::query` returns the ACNUC
  id, which appears to be the genbank "locus" field, whereas `taxize::ncbi_getbyid` uses the
  "accession" or "GI" field. Usually the "locus" and "accession" field are the same, but when
  they are differnt results of the search are not downloaded. Therefore the maximum count of
  sequences will not always be returned even if more than the maximum are found.
}
\examples{
\dontrun{
x <- download_gb_taxon(c("phytophthora", "pythium"),
                       c("@18S@", "@28S@"),
                       "RRNA")}
}
\seealso{
\code{\link{download_gb_query}} \code{\link{query_taxon}}
}

