% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{edge_list_depth}
\alias{edge_list_depth}
\title{Get distance from root of edgelist items}
\usage{
edge_list_depth(taxa, parents)
}
\arguments{
\item{taxa}{(\code{character}) Unique taxon IDs for every possible taxon.}

\item{parents}{(\code{character}) Unique taxon IDs for the supertaxa of every possible taxon.
Root taxa should have \code{NA} in this column.}
}
\description{
Gets the number of ancestors/supergroups for items of an edge/adjacency list
}

