% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new--heat_tree--shapes.R
\name{edge_shape_default}
\alias{edge_shape_default}
\title{Edge shape function: circle}
\usage{
edge_shape_default(edge_color = "grey")
}
\description{
A function used to plot edges in a heat tree. This function plots the default lines in a solid color.
}
