% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new--heat_tree--shapes.R
\name{edge_shape_functions}
\alias{edge_shape_functions}
\title{All built-in node shape functions}
\usage{
edge_shape_functions(simple = TRUE)
}
\arguments{
\item{simple}{If TRUE, only return shape function for simple geometric shapes}
}
\value{
named list of functions
}
\description{
Returns a named list of all built-in node shape functions.
}
