% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_taxa--dataset_documentation.R
\name{ex_hierarchies}
\alias{ex_hierarchies}
\title{An example hierarchies object}
\format{
A [hierarchies()] object.
}
\source{
Created from the example code in the [hierarchies()]
  documentation.
}
\description{
An example hierarchies object built from the ground up.
}
\seealso{
Other taxa-datasets: 
\code{\link{ex_hierarchy1}},
\code{\link{ex_hierarchy2}},
\code{\link{ex_hierarchy3}},
\code{\link{ex_taxmap}}
}
\concept{taxa-datasets}
\keyword{data}
