% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_taxa--dataset_documentation.R
\name{ex_hierarchy1}
\alias{ex_hierarchy1}
\title{An example Hierarchy object}
\format{
A [hierarchy()] object with
\itemize{
 \item name: Poaceae / rank: family / id: 4479
 \item name: Poa / rank: genus / id: 4544
 \item name: Poa annua / rank: species / id: 93036
}
Based on NCBI taxonomic classification
}
\source{
Created from the example code in the [hierarchy()]
  documentation.
}
\description{
An example Hierarchy object built from the ground up.
}
\seealso{
Other taxa-datasets: 
\code{\link{ex_hierarchies}},
\code{\link{ex_hierarchy2}},
\code{\link{ex_hierarchy3}},
\code{\link{ex_taxmap}}
}
\concept{taxa-datasets}
\keyword{data}
