% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_taxa--dataset_documentation.R
\name{ex_hierarchy2}
\alias{ex_hierarchy2}
\title{An example Hierarchy object}
\format{
A [hierarchy()] object with
\itemize{
 \item name: Felidae / rank: family / id: 9681
 \item name: Puma / rank: genus / id: 146712
 \item name: Puma concolor / rank: species / id: 9696
}
Based on NCBI taxonomic classification
}
\source{
Created from the example code in the [hierarchy()]
  documentation.
}
\description{
An example Hierarchy object built from the ground up.
}
\seealso{
Other taxa-datasets: 
\code{\link{ex_hierarchies}},
\code{\link{ex_hierarchy1}},
\code{\link{ex_hierarchy3}},
\code{\link{ex_taxmap}}
}
\concept{taxa-datasets}
\keyword{data}
