% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_taxa--dataset_documentation.R
\name{ex_hierarchy3}
\alias{ex_hierarchy3}
\title{An example Hierarchy object}
\format{
A [hierarchy()] object with
\itemize{
 \item name: Chordata / rank: phylum / id: 158852
 \item name: Vertebrata / rank: subphylum / id: 331030
 \item name: Teleostei / rank: class / id: 161105
 \item name: Salmonidae / rank: family / id: 161931
 \item name: Salmo / rank: genus / id: 161994
 \item name: Salmo salar / rank: species / id: 161996
}
Based on ITIS taxonomic classification
}
\source{
Created from the example code in the [hierarchy()]
  documentation.
}
\description{
An example Hierarchy object built from the ground up.
}
\seealso{
Other taxa-datasets: 
\code{\link{ex_hierarchies}},
\code{\link{ex_hierarchy1}},
\code{\link{ex_hierarchy2}},
\code{\link{ex_taxmap}}
}
\concept{taxa-datasets}
\keyword{data}
