% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_taxa--dataset_documentation.R
\name{ex_taxmap}
\alias{ex_taxmap}
\title{An example taxmap object}
\format{
A [taxmap()] object.
}
\source{
Created from the example code in the [taxmap()]
  documentation.
}
\description{
An example taxmap object built from the ground up. Typically, data stored in
taxmap would be parsed from an input file, but this data set is just for
demonstration purposes.
}
\seealso{
Other taxa-datasets: 
\code{\link{ex_hierarchies}},
\code{\link{ex_hierarchy1}},
\code{\link{ex_hierarchy2}},
\code{\link{ex_hierarchy3}}
}
\concept{taxa-datasets}
\keyword{data}
