% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_taxa--dataset_documentation.R
\name{ex_taxonomy}
\alias{ex_taxonomy}
\title{An example Taxonomy object}
\format{
A [taxonomy()] object.
}
\source{
Created from the example code in the [taxonomy()]
  documentation.
}
\description{
An example Taxonomy object built from the ground up.
}
\seealso{
Other taxa-datasets: 
\code{\link{ex_hierarchies}},
\code{\link{ex_hierarchy1}},
\code{\link{ex_hierarchy2}},
\code{\link{ex_hierarchy3}},
\code{\link{ex_taxmap}}
}
\concept{taxa-datasets}
\keyword{data}
