% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sequence_download.R
\name{extract_description}
\alias{extract_description}
\title{Extract the description from genbank annotations.}
\usage{
extract_description(annotation)
}
\description{
Takes a genebank annotation as a character vector with one line per element and returns the
description.
}

