% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sequence_download.R
\name{extract_gi}
\alias{extract_gi}
\title{Extract the gi from genbank annotations.}
\usage{
extract_gi(annotation)
}
\description{
Takes a genebank annotation as a character vector with one line per element and returns the
gi.
}

