% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sequence_download.R
\name{extract_organism}
\alias{extract_organism}
\title{Extract the binomial organism name from genbank annotations.}
\usage{
extract_organism(annotation)
}
\description{
Takes a genebank annotation as a character vector with one line per element and returns the
organism name
}

