% Generated by roxygen2 (4.0.2): do not edit by hand
\name{extract_taxonomy}
\alias{extract_taxonomy}
\title{Extract taxonomy information from sequence headers}
\usage{
extract_taxonomy(input, parser, database = "ncbi")
}
\arguments{
\item{input}{(\code{character}) A vector from which to extract taxonomy information.}

\item{parser}{(\code{character; length == 1}) A regular expression with custom replacement
patterns indicating the locations of relevant information. Supported patterns include:
\describe{
  \item{\code{\%\%tax_id\%\%}}{A unique numeric id for a taxon.}
  \item{\code{\%\%name\%\%}}{The name of a taxon. Not necessarily unique.}
  \item{\code{\%\%lineage<sep1><sep2>\%\%}}{A list of taxa names that consitute the full taxonomic classification
from broad to specific. Individual names are not necessarily unique. Individual names are separated
by the \code{<sep>} terms. \code{<sep1>}
is required, as it defines the characters the deliminate taxa in the lineage. \code{<sep2>} is optional,
and further splits each taxon into its name and its level. For example:
\href{http://unite.ut.ee/repository.php}{UNITE} uses a format for taxonomy like
"k__Fungi;p__Ascomycota;c__Dothideomycetes...". For this format use \code{\%\%lineage<;><__>\%\%}}
  \item{\code{\%\%lineage_id<sep1><sep2>\%\%}}{A list of taxa unique ids that consitute the full taxonomic
classification from broad to specific. Same usage as \code{\%\%lineage<sep1><sep2>\%\%}}
  \item{\code{\%\%seq_id\%\%}}{An unique sequence identifier. The taxonomy information will be
looked up if available. Requires an internet connection.}
}}

\item{database}{(\code{character}): The name of the database that patterns given in
\code{parser} will apply to. Currently, only \code{ncbi} is being supported.}
}
\value{
Returns a list of two elements:
 \describe{
   \item{\code{taxonomy}}{An type of \href{http://en.wikipedia.org/wiki/Adjacency_list}{adjacency list}
   represented as a data.frame with 3  columns. The first column is the taxon id and the second two
   are ids of a taxon and its subtaxon respectivly. If a taxon has more than subtaxon, it will be
   present multiple times in the 2nd "parent taxon" column. However, a taxon can only be
   present once in the 3rd "sub taxon" column; otherwise, it would imply the taxon has more than
   one "parent". This data structure can be converted to more intuitive ones with ... TO BE CONTINUED.}
   \item{\code{taxon_info}}{A data.frame with one row per taxon, containing available information
   on each taxon. The number and nature of columns depend on the input data. Typically, a column
   of taxon names is present. }
   \item{\code{item_info}}{A data.frame corresponding to input items (typically sequences)
   containing their taxon ids. In addition, the content of any
   \href{http://en.wikipedia.org/wiki/Regular_expression}{regex} capturing groups (i.e. parts bound
   by parenthesis) in the \code{parser} argument will be present as columns of type
   \code{character}.}
 }
}
\description{
Extracts the taxonomy used by a set of sequences based on their header information. A data
structure representing the heirerarchical nature of the taxonomy as well as a vector
identifing the taxon of each sequence is returned. Taxa are translated into unique codes if they
are not already encoded this way.
}

