% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fasta.R
\name{fasta_headers}
\alias{fasta_headers}
\title{Get line numbers of FASTA headers}
\usage{
fasta_headers(file_path, buffer_size = 1000, return_headers = TRUE)
}
\arguments{
\item{file_path}{(\code{character} of length 1) The path to a file to read.}

\item{buffer_size}{(\code{numeric} of length 1) The number of lines in each chunk.}

\item{return_headers}{(\code{logical} of length 1) If \code{TRUE}, name the result with the headers.}
}
\value{
\code{numeric}
}
\description{
Get line numbers of FASTA headers without reading whole fasta file into RAM.
}
\keyword{internal}
