% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classified--dplyr.R
\name{filter.classified}
\alias{filter.classified}
\title{Filter \code{\link{classified}} on a list of conditions}
\usage{
\method{filter}{classified}(obj, ..., subtaxa = TRUE, supertaxa = FALSE,
  itemless = TRUE, taxonless = FALSE)
}
\arguments{
\item{obj}{\code{\link{classified}}}

\item{...}{Filtering conditions.
Each condition of must contain the name of at least one column from \code{obj$taxon_data} or \code{obj$taxon_data}.
To filter by index or \code{TRUE}/\code{FALSE} vector, use \code{\link{filter_taxa}} or \code{\link{filter_items}}.}

\item{subtaxa}{(\code{logical} of length 1)
If \code{TRUE}, return subtaxa of specified taxa.}

\item{supertaxa}{(\code{logical} of length 1)
If \code{TRUE}, return supertaxa of specified taxa.}

\item{itemless}{(\code{logical} of length 1)
If \code{TRUE}, return taxa even if they have no items assigned to them.}

\item{taxonless}{(\code{logical} of length 1)
If \code{TRUE}, return items even if they are not assigned to taxa.}
}
\value{
\code{\link{classified}}
}
\description{
Create a subset of a \code{\link{classified}} object.
Can filter based on columns in \code{obj$taxon_data} or \code{obj$taxon_data}.
}

