% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classified--dplyr.R
\name{filter_.classified}
\alias{filter_.classified}
\title{Standard evaluation version of \code{\link{filter}}}
\usage{
filter_.classified(obj, ..., .dots, subtaxa = TRUE, supertaxa = FALSE,
  itemless = TRUE, taxonless = FALSE)
}
\description{
Standard evaluation version of \code{\link{filter}}
}

