% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_ambiguous.R
\name{filter_ambiguous_taxa}
\alias{filter_ambiguous_taxa}
\title{Filter ambiguous taxa}
\usage{
filter_ambiguous_taxa(obj, unknown = TRUE, uncultured = TRUE,
  name_regex = ".", ...)
}
\arguments{
\item{obj}{A taxmap object}

\item{unknown}{If \code{TRUE}, Remove taxa with names the suggest they are
placeholders for unknown taxa (e.g. "unknown ...").}

\item{uncultured}{If \code{TRUE}, Remove taxa with names the suggest they are
assinged to uncultured organisms (e.g. "uncultured ...").}

\item{name_regex}{The regex code to match a valid character in a taxon name.
For example, "[a-z]" would mean taxon names can only be lower case letters.}

\item{...}{Passed to \code{\link[taxa](filter_taxa)}}
}
\description{
Filter out taxa with ambiguous names, such as "unknown" or "uncultured". If
you encounter a taxon name that represents an ambiguous taxon that is not
filtered out by this function, let us know and we will add it.
}
