% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classified--dplyr.R
\name{filter_taxa}
\alias{filter_taxa}
\title{UNDER CONSTRUCTION}
\usage{
filter_taxa(.data, ...)
}
\arguments{
\item{.data}{\code{\link{classified}}}

\item{...}{Filtering conditions.}
}
\description{
UNDER CONSTRUCTION
}

