% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classified--internal.R
\name{format_taxon_subset}
\alias{format_taxon_subset}
\title{Format taxon subset value}
\usage{
format_taxon_subset(obj, index)
}
\arguments{
\item{obj}{(\code{classified})
The \code{classified} object containing taxon information to be queried.}

\item{index}{}
}
\value{
\code{character}
}
\description{
Format an input to a \code{subset} option on functions like \code{subset}.
Converts logical and numeric vectors into taxon ids
}
\keyword{internal}

