% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_documentation.R
\docType{data}
\name{genbank_ex_data}
\alias{genbank_ex_data}
\title{Fungal ITS Genbank refseq}
\format{An object of type \code{\link{classified}}}
\source{
\url{http://www.ncbi.nlm.nih.gov/nuccore}
}
\usage{
genbank_ex_data
}
\description{
A dataset containing information for 605 sequences obtained form NCBI using the following query:
}
\details{
\code{
(18s[All Fields] AND 28s[All Fields]) AND "basidiomycetes"[porgn] AND (refseq[filter] AND ("1"[SLEN] : "1000"[SLEN]))
}
}
\examples{
\dontrun{
file_path <- system.file("extdata", "ncbi_basidiomycetes.fasta", package = "metacoder")
sequences <- ape::read.FASTA(file_path)
genbank_ex_data <- extract_taxonomy(sequences,
                                    regex = "^.*\\\\|(.*)\\\\|.*\\\\|(.*)\\\\|(.*)$",
                                    key = c("item_id", acc_no = "item_info", desc = "item_info"))
}

}
\keyword{datasets}

