% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new--heat_tree--internal.R
\name{get_characters}
\alias{get_characters}
\title{Return character values from a list or vector}
\usage{
get_characters(x)
}
\arguments{
\item{x}{A list or vector}
}
\value{
a character vector
}
\description{
Return character values from a list or vector
}
\keyword{internal}
