% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_taxa--internal.R
\name{get_database_name}
\alias{get_database_name}
\title{Return name of database}
\usage{
get_database_name(input)
}
\arguments{
\item{input}{Either a character vector or `TaxonDatabase` class}
}
\value{
The name of the database
}
\description{
This is meant to return the name of a database when it is not known if the
input is a `TaxonDatabase` object or a simple character vector.
}
\keyword{internal}
