% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_taxa--internal.R
\name{get_dots_or_list}
\alias{get_dots_or_list}
\title{Get input from dots or list}
\usage{
get_dots_or_list(..., .list = NULL)
}
\arguments{
\item{...}{Dots input}

\item{.list}{List input}
}
\value{
A list of inputs
}
\description{
Get input from dots or list, but not both.
Throws an error if both are supplied.
}
\keyword{internal}
