% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_phyloseq.R
\name{get_expected_data}
\alias{get_expected_data}
\title{Get a data set in as_phyloseq}
\usage{
get_expected_data(obj, input, default, expected_class)
}
\arguments{
\item{obj}{The taxmap object}

\item{input}{The input to as_phyloseq options.}

\item{default}{The default name of the data set.}

\item{expected_class}{What the dataset is expected to be.}
}
\description{
Get a data set in as_phyloseq
}
\keyword{internal}
