% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new--heat_tree--internal.R
\name{get_factors}
\alias{get_factors}
\title{Return factors from a list or vector}
\usage{
get_factors(x)
}
\arguments{
\item{x}{A list or vector}
}
\value{
a factor vector
}
\description{
Return factors from a list or vector
}
\keyword{internal}
