% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_taxonomy--parsers.R
\name{get_id_from_name}
\alias{get_id_from_name}
\title{Get taxon ID from name}
\usage{
get_id_from_name(name, database)
}
\arguments{
\item{name}{(\code{character})
Names to look up IDs for}

\item{database}{(\code{character} of length 1)
Database to use to look up id.}
}
\description{
Get a taxon's ID from its name
}

