% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{get_invalid_cols}
\alias{get_invalid_cols}
\title{Return invalid column names/indexes}
\usage{
get_invalid_cols(table, cols)
}
\arguments{
\item{table}{Table to check}

\item{cols}{he names/indexes of columns}
}
\description{
Return invalid column names/indexes
}
\keyword{internal}
