% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_taxonomy--parsers.R
\name{get_name_from_id}
\alias{get_name_from_id}
\title{Get taxon name from ID}
\usage{
get_name_from_id(id, database)
}
\arguments{
\item{id}{(\code{character})
Id to look up names for}

\item{database}{(\code{character} of length 1)
Database to use to look up id.}
}
\description{
Get a taxon's name from its ID
}
\keyword{internal}

