% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R, R/new--heat_tree--internal.R
\name{get_numerics}
\alias{get_numerics}
\title{Return numeric values in a character}
\usage{
get_numerics(x)

get_numerics(x)
}
\arguments{
\item{x}{A list or vector}

\item{input}{}
}
\value{
a numeric vector
}
\description{
Returns just valid numeric values and ignores others.

Return numeric values from a list or vector
}
\keyword{internal}
