% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/internal.R
\name{get_optimal_range}
\alias{get_optimal_range}
\title{Find optimal range}
\usage{
get_optimal_range(max_range, min_range, resolution, opt_crit, minimize = TRUE)
}
\arguments{
\item{max_range}{(\code{numeric} of length 2) The min and max boundries to the search space for
the optimal maximum value.}

\item{min_range}{(\code{numeric} of length 2) The min and max boundries to the search space for
the optimal minimum value.}

\item{resolution}{(\code{numeric} of length 2) The number of increments in each dimension.}

\item{opt_crit}{(\code{function}) A function that takes two arguments, the max and min, and
returns the optimality statistic.}

\item{minimize}{(\code{logical} of length 1) If \code{TRUE}, the smallest optimality statistic
is considered the best. By default, larger is considered better.}
}
\value{
A \code{numeric} vector of length 2. The optimal max and min values.
}
\description{
Finds optimal max and min value useing an optimality criterion.
}

