% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa--classified_class.R
\name{get_rank}
\alias{get_rank}
\title{Get rank of taxon in \code{\link{classified}}}
\usage{
get_rank(obj, taxon)
}
\arguments{
\item{obj}{(\code{\link{classified}})}

\item{taxon}{(\code{character} of length 1) The taxon_id used}
}
\value{
\code{numeric} of length 1
}
\description{
Get rank of taxon in \code{\link{classified}}
}
\keyword{internal}

