% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_taxa--taxmap--parsers.R
\name{get_sort_var}
\alias{get_sort_var}
\title{Get a vector from a vector/list/table to be used in mapping}
\usage{
get_sort_var(data, var)
}
\arguments{
\item{data}{A vector/list/table}

\item{var}{What to get.
* For tables, the names of columns can be used.
* `"\{\{index\}\}"` : This means to use the index of rows/items
* `"\{\{name\}\}"`  : This means to use row/item names.
* `"\{\{value\}\}"` : This means to use the values in vectors or lists. Lists}
}
\description{
Get a vector from a vector/list/table to be used in mapping
}
\keyword{internal}
