% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{get_stem_taxa}
\alias{get_stem_taxa}
\title{Identify shared taxonomy root}
\usage{
get_stem_taxa(taxa, parents, include_root = FALSE)
}
\arguments{
\item{taxa}{(\code{character}) Unique taxon IDs for every possible taxon.}

\item{parents}{(\code{character}) Unique taxon IDs for the supertaxa of every possible taxon. A
lack of a parent should be coded as \code{NA}.}

\item{include_root}{(\code{logical}) If \code{TRUE}, the index of the new root (taxon with the 
rank that still contains all other taxa).}
}
\value{
A interger vector of indexes corresponding to taxa shared by all subtaxa.
}
\description{
Given an adjacency/edge list, return the indexes of taxa shared by all subtaxa.
}
\keyword{internal}

