% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{get_subtaxa}
\alias{get_subtaxa}
\title{Get all subtaxa of a taxon}
\usage{
get_subtaxa(targets, taxa, parents, recursive = TRUE, simplify = FALSE)
}
\arguments{
\item{targets}{(\code{character}) Taxon IDs for which subtaxa will be returned.}

\item{taxa}{(\code{character}) Unique taxon IDs for every possible taxon.}

\item{parents}{(\code{character}) Unique taxon IDs for the subtaxa of every possible taxon. The
lack of a parent should be coded as \code{NA}.}

\item{recursive}{(\code{logical}) If \code{FALSE}, only return the subtaxa one level above the
target taxa. If \code{TRUE}, return all the subtaxa of every subtaxa, etc.}

\item{simplify}{(\code{logical}) If \code{TRUE}, then combine all the results into a single
vector of unique taxon IDs}

\item{include_target}{(\code{logical}) If \code{TRUE}, the target taxa are included in the output}
}
\value{
If \code{simplify = FALSE}, then a list of vectors of taxon IDs are returned
  corresponding to the \code{target} argument. If \code{simplify = TRUE}, then the unique taxon
  IDs for all \code{target} taxa are returned in a single vector.
}
\description{
Given one or more taxa IDs and the edge list defining the taxonomy, return the taxon IDs of all
subtaxa
}

