% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa--internal.R
\name{get_supertaxa}
\alias{get_supertaxa}
\title{Get all supertaxa of a taxon}
\usage{
get_supertaxa(targets, taxa, parents, recursive = TRUE, simplify = FALSE,
  include_target = FALSE)
}
\arguments{
\item{targets}{(\code{character}) Taxon IDs for which supertaxa will be returned.}

\item{taxa}{(\code{character}) Unique taxon IDs for every possible taxon.}

\item{parents}{(\code{character}) Unique taxon IDs for the supertaxa of every possible taxon. The
lack of a parent should be coded as \code{NA}.}

\item{recursive}{(\code{logical}) If \code{FALSE}, only return the supertaxa one level above the
target taxa. If \code{TRUE}, return all the supertaxa of every supertaxa, etc.}

\item{simplify}{(\code{logical}) If \code{TRUE}, then combine all the results into a single
vector of unique taxon IDs}

\item{include_target}{(\code{logical}) If \code{TRUE}, the target taxa are included in the output}
}
\value{
If \code{simplify = FALSE}, then a list of vectors of taxon IDs are returned
  corresponding to the \code{target} argument. If \code{simplify = TRUE}, then the unique taxon
  IDs for all \code{target} taxa are returned in a single vector.
}
\description{
Given one or more taxa IDs and the edge list defining the taxonomy, return the taxon IDs of all
supertaxa (i.e. all taxa the target taxa are a part of).
}
\keyword{internal}

