% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{get_taxmap_table}
\alias{get_taxmap_table}
\title{Get a table from a taxmap object}
\usage{
get_taxmap_table(obj, dataset, expected_cols = NULL)
}
\arguments{
\item{obj}{A taxmap object}

\item{dataset}{The name of the table}

\item{expected_cols}{The names/indexes of columns expected to exist. If a column is not found, issue a warning.}
}
\value{
A data.frame
}
\description{
Get a table from a taxmap object and complain if it does not exist.
}
\keyword{internal}
