% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_taxon_count}
\alias{get_taxon_count}
\title{Count occurrences throughout a taxonomy}
\usage{
get_taxon_count(taxa, parents, items)
}
\arguments{
\item{taxa}{(\code{character}) Unique taxon ids for every possible taxon.}

\item{parents}{(\code{character}) Unique taxon ids for the supertaxa of every possible taxon.}

\item{items}{(\code{character}) Taxon ids for a set of items to count.}
}
\value{
A named vector of counts corresponding to the taxa in \code{taxa}
}
\description{
Get counts of items at each unique taxon in a taxonomy, including internal taxa.
}

