% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{get_taxon_items}
\alias{get_taxon_items}
\title{Get all subtaxa of a taxon}
\usage{
get_taxon_items(targets, taxa, parents, items, recursive = TRUE,
  simplify = FALSE)
}
\arguments{
\item{targets}{(\code{character}) Taxon IDs for which items will be returned.}

\item{taxa}{(\code{character}) Unique taxon IDs for every possible taxon.}

\item{parents}{(\code{character}) Unique taxon IDs for the subtaxa of every possible taxon. The
lack of a parent should be coded as \code{NA}.}

\item{items}{(\code{character}) Taxon IDs for a set of items.}

\item{recursive}{(\code{logical}) If \code{FALSE}, only return the item assigned to the specified
target taxa, not its subtaxa. If \code{TRUE}, return all the items of every subtaxa, etc.}

\item{simplify}{(\code{logical}) If \code{TRUE}, then combine all the results into a single
  vector of unique item indexes.}
}
\value{
If \code{simplify = FALSE}, then a list of vectors of item indexes are returned
  corresponding to the \code{target} argument. If \code{simplify = TRUE}, then the item indexes
  for all \code{target} taxa are returned in a single vector.
}
\description{
Given one or more taxa IDs and the edge list defining the taxonomy, return the taxon IDs of all
subtaxa
}

