% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sequence_download.R
\name{get_taxon_sample}
\alias{get_taxon_sample}
\title{Download representative sequences for a taxon}
\usage{
get_taxon_sample(name = NULL, id = NULL, target_level, max_counts = NULL,
  interpolate_max = TRUE, min_counts = NULL, interpolate_min = TRUE,
  verbose = TRUE, max_length = 10000, min_length = 1,
  max_children = NULL, min_children = NULL, ...)
}
\arguments{
\item{target_level}{A character vector of length 1. The finest taxonomic level at which
to sample. The finest level at which replication occurs. Must be a finer level than
\code{taxon}.}

\item{max_counts}{A named numeric vector. The maximum number of sequences to download for each
taxonomic level. The names correspond to taxonomic levels. See
\code{\link{get_taxonomy_levels}} or \code{\link[taxize]{rank_ref}} for available taxonomic
levels.}

\item{taxon}{A character vector of length 1. The taxon to download a sample of sequences for.}
}
\description{
Downloads a sample of sequences meant to evenly capture the diversity of a given taxon.
}
\examples{
get_taxon_sample(name = "oomycetes", target_level = "genus")
fungi <- get_taxon_sample(name = "fungi", target_level = "family", max_counts = c(family = 5, order = 30), entrez_query = "18S[All Fields] AND 28S[All Fields]", min_length = 600, max_length = 10000)
}

