% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_taxonomy_levels}
\alias{get_taxonomy_levels}
\title{Get taxonomy levels}
\usage{
get_taxonomy_levels()
}
\value{
An ordered factor of taxonomy levels, such as "Subkingdom" and "Order", in order of the
  hierarchy.
}
\description{
Get taxonomy levels
}

