% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new--heat_tree--internal.R
\name{group_list_by_taxa}
\alias{group_list_by_taxa}
\title{Group list items by taxon id}
\usage{
group_list_by_taxa(obj, a_list, def_type_func)
}
\arguments{
\item{obj}{a taxmap}

\item{a_list}{a list to convert}

\item{def_type_func}{An as.* type function to corece values to a certain type.}
}
\value{
A list corresponding to taxa in obj
}
\description{
Group list items by taxon id and fill empty ids with NULL
}
\keyword{internal}
