% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new--heat_tree--internal.R
\name{heat_tree_colors}
\alias{heat_tree_colors}
\title{Determine colors used in heat_tree}
\usage{
heat_tree_colors(obj)
}
\arguments{
\item{obj}{a taxmap}
}
\value{
a taxmap
}
\description{
Determine the colors used in a heat_tree
}
\keyword{internal}
