% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new--heat_tree.R
\name{heat_tree_data}
\alias{heat_tree_data}
\title{Make data for plotting a heat tree using \code{\link{heat_tree_plot}}}
\usage{
heat_tree_data(obj, node_label = NULL, edge_label = NULL,
  tree_label = NULL, node_size = NULL, edge_size = node_size,
  node_label_size = node_size, edge_label_size = edge_size,
  tree_label_size = node_size, node_color = "grey",
  edge_color = node_color, node_label_color = "black",
  edge_label_color = "black", tree_label_color = "black",
  node_size_trans = "area", edge_size_trans = node_size_trans,
  node_label_size_trans = "linear", edge_label_size_trans = "linear",
  tree_label_size_trans = "linear", node_color_trans = "linear",
  edge_color_trans = node_color_trans,
  node_label_color_trans = "linear", edge_label_color_trans = "linear",
  tree_label_color_trans = "linear", node_size_range = NULL,
  edge_size_range = node_size_range,
  node_label_size_range = node_size_range,
  edge_label_size_range = NULL, tree_label_size_range = NULL,
  node_color_range = quantative_palette(),
  edge_color_range = node_color_range, node_label_color_range = NULL,
  edge_label_color_range = NULL, tree_label_color_range = NULL,
  node_size_interval = NULL, node_color_interval = NULL,
  edge_size_interval = NULL, edge_color_interval = NULL,
  margin_size = 0, aspect_ratio = 1, layout = "reingold-tilford",
  initial_layout = NULL, make_node_legend = TRUE,
  make_edge_legend = TRUE, title = NULL, title_size = 0.1,
  node_color_axis_label = NULL, node_size_axis_label = NULL,
  edge_color_axis_label = NULL, edge_size_axis_label = NULL,
  background_color = "#FFFFFF00", output_file = NULL,
  repel_labels = TRUE, repel_force = 1, repel_iter = 1000, ...)
}
\arguments{
\item{obj}{An object of type \code{\link[taxa]{Taxmap}}. This typically
contains all of the data that will be plotted. Any data that is to be
plotted that is not contained in this object will need to named by the
taxon ids used in this object or be the same length as the number of taxa
in this object.}

\item{node_label}{Labels associated with nodes. See the "labels" section
below for more information. Default: no node labels.}

\item{edge_label}{Labels associated with edges. See the "labels" section
below for more information. Default: no edge labels.}

\item{tree_label}{Labels associated with subtrees. When there are multiple
roots to the taxonomy, multiple trees are printed, one for each root, and
they are automatically arranged on the same graph using the same color/size
scales. This option allows each subtree to be labeled. Values assocaited
with the roots and given to this option will be used for the subtree
labels. See the "labels" section below for more information. Default: no
subtree labels.}

\item{node_size}{A numeric value used to determine relative node size. See
the "size" section below for more information. Default: constant size.}

\item{edge_size}{A numeric value used to determine relative edge size.
Although edges connect two nodes, in this function, their relative width is
associated with a single taxon, represented by the node the edge leads to.
For this reason, root taxa do not have edges. Keep in mind that edges will
look odd if they are bigger than their assocaited node. See the "size"
section below for more information. Default: half of node size.}

\item{node_label_size}{A numeric value used to determine relative node label
size. See the "size" section below for more information. Default: relative
to node size.}

\item{edge_label_size}{A numeric value used to determine relative edge label
size. See the "size" section below for more information. Default: relative
to edge size.}

\item{tree_label_size}{A numeric value used to determine relative tree label
size. See the "size" section below for more information. Default: relative
to subtree size.}

\item{node_color}{A numeric value used to determine relative node color or a
character value representing a color, either hex code or color name. See
the "color" section below for more information. Default: grey.}

\item{edge_color}{A numeric value used to determine relative edge color or a
character value representing a color, either hex code or color name. See
the "color" section below for more information. Default: same as node
color.}

\item{node_label_color}{A numeric value used to determine relative node label
color or a character value representing a color, either hex code or color
name. See the "color" section below for more information.. Default: black.}

\item{edge_label_color}{A numeric value used to determine relative edge label
color or a character value representing a color, either hex code or color
name. See the "color" section below for more information. Default: black.}

\item{tree_label_color}{A numeric value used to determine relative subtree
label color or a character value representing a color, either hex code or
color name. See the "color" section below for more information. Default:
black.}

\item{node_size_trans}{How the numbers used to determine node size is
transformed before being mapped to node size. See details on
transformations. Default: \code{"area"}.}

\item{edge_size_trans}{How the numbers used to determine edge size is
transformed before being mapped to edge size. See details on
transformations. Default: same as \code{node_size_trans}.}

\item{node_label_size_trans}{How the numbers used to determine node label
size is transformed before being mapped to node label size. See details on
transformations. Default: same as \code{node_size_trans}.}

\item{edge_label_size_trans}{How the numbers used to determine edge label
size is transformed before being mapped to edge label size. See details on
transformations. See details on transformations. Default: same as
\code{edge_size_trans}.}

\item{tree_label_size_trans}{How the numbers used to determine subtree label
size is transformed before being mapped to subtree label size. See details
on transformations. See details on transformations. Default: \code{"area"}.}

\item{node_color_trans}{How the numbers used to determine node color is
transformed before being mapped to node color. See details on
transformations. Default: \code{"linear"}.}

\item{edge_color_trans}{How the numbers used to determine edge color is
transformed before being mapped to edge color. See details on
transformations. Default: same as node color transformation.}

\item{node_label_color_trans}{How the numbers used to determine node label
color is transformed before being mapped to node label color. See details
on transformations. Default: \code{"linear"}.}

\item{edge_label_color_trans}{How the numbers used to determine edge label
color is transformed before being mapped to edge label color. See details
on transformations. Default: \code{"linear"}.}

\item{tree_label_color_trans}{How the numbers used to determine subtree label
color is transformed before being mapped to subtree label color. See
details on transformations. Default: \code{"linear"}.}

\item{node_size_range}{The minimum and maximum diameter of nodes, as a length
2 numeric vector, in units of proportion of graph size. For example
\code{c(0.01, 0.02)} would make nodes between 1% and 2% of graph size. See
details on size ranges. Default: Optimize to balance overlaps and range
size.}

\item{edge_size_range}{The minimum and maximum width of edges, as a length 2
numeric vector, in units of proportion of graph size. For example
\code{c(0.01, 0.02)} would make edge widths between 1% and 2% of graph
size. See details on size ranges. Default: One half of node diameter.}

\item{node_label_size_range}{The minimum and maximum size of node label text,
as a length 2 numeric vector, in units of proportion of graph size. For
example \code{c(0.01, 0.02)} would make text between 1% and 2% of graph
size. See details on size ranges. Default: Relative to node size.}

\item{edge_label_size_range}{The minimum and maximum size of edge label text,
as a length 2 numeric vector, in units of proportion of graph size. For
example \code{c(0.01, 0.02)} would make text between 1% and 2% of graph
size. See details on size ranges. Default: Relative to edge size.}

\item{tree_label_size_range}{The minimum and maximum size of tree label text,
as a length 2 numeric vector, in units of proportion of graph size. For
example \code{c(0.01, 0.02)} would make text between 1% and 2% of graph
size. See details on size ranges. Default: Relative to subtree size.}

\item{node_color_range}{A series of colors used to make the color ramp used
for nodes, as a character vector of colors. See details on color ranges
below for more information. Default: Color-blind friendly palette.}

\item{edge_color_range}{A series of colors used to make the color ramp used
for edges, as a character vector of colors. See details on color ranges
below for more information. Default: Color-blind friendly palette.}

\item{node_label_color_range}{A series of colors used to make the color ramp
used for node labels, as a character vector of colors. See details on color
ranges below for more information. Default: Color-blind friendly palette.}

\item{edge_label_color_range}{A series of colors used to make the color ramp
used for edge labels, as a character vector of colors. See details on color
ranges below for more information. Default: Color-blind friendly palette.}

\item{tree_label_color_range}{A series of colors used to make the color ramp
used for subtree labels, as a character vector of colors. See details on
color ranges below for more information. Default: Color-blind friendly
palette.}

\item{node_size_interval}{A numeric vector of length 2 that sets the minimum
and maximum values represented by node size. See details on intervals below
for more information. Default: The range of values in \code{node_size}.}

\item{node_color_interval}{A numeric vector of length 2 that sets the minimum
and maximum values represented by node color. See details on intervals
below for more information. Default: The range of values in
\code{node_color}.}

\item{edge_size_interval}{A numeric vector of length 2 that sets the minimum
and maximum values represented by edge size. See details on intervals below
for more information. Default: The range of values in \code{edge_size}.}

\item{edge_color_interval}{A numeric vector of length 2 that sets the minimum
and maximum values represented by edge color. See details on intervals
below for more information. Default: The range of values in
\code{edge_color}.}

\item{layout}{The layout algorithm used to position nodes. See details on
layouts. Default: \code{"reingold-tilford"}.}

\item{initial_layout}{he layout algorithm used to set the initial position of
nodes, passed as input to the \code{layout} algorithm. See details on
layouts. Default: Not used.}

\item{make_node_legend}{if TRUE, make legend for node size/color mappings.
Default: TRUE.}

\item{make_edge_legend}{if TRUE, make legend for edge size/color mappings.
Default: TRUE.}

\item{title}{Text to print above the graph.}

\item{title_size}{The size of the title relative to the rest of the graph.
For example a value of \code{.05} will make the title text's height 5% of
the graph's size.}

\item{node_color_axis_label}{The label on the scale axis corresponding to
\code{node_color}. Default: The expression given to \code{node_color}.}

\item{node_size_axis_label}{The label on the scale axis corresponding to
\code{node_size}. Default: The expression given to \code{node_size}.}

\item{edge_color_axis_label}{The label on the scale axis corresponding to
\code{edge_color}. Default: The expression given to \code{edge_color}.}

\item{edge_size_axis_label}{The label on the scale axis corresponding to
\code{edge_size}. Default: The expression given to \code{edge_size}.}

\item{repel_labels}{If \code{TRUE}, use code from the ggrepel package to
spread out labels. Default: TRUE.}

\item{repel_force}{The force of which overlapping labels will be repelled
from eachother. Default: 1.}

\item{repel_iter}{The number of iterations used when repelling labels.
Default: 1000.}

\item{...}{(other named arguments) Passed to the \code{\link{igraph}} layout
function used.}
}
\description{
This makes a \code{\link[taxa]{taxmap}} object containing all the information
needed to plot a heat tree. Using \code{\link{heat_tree_data}} and
\code{\link{heat_tree_plot}} is intended for advanced users when
customizations of the plot must be scripted.
}
\examples{
# See examples in the `heat_tree` function

}
