% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new--heat_tree--internal.R
\name{heat_tree_init_taxmap}
\alias{heat_tree_init_taxmap}
\title{Reformat heat_tree arguments to taxmap}
\usage{
heat_tree_init_taxmap(obj, arguments)
}
\arguments{
\item{obj}{The input taxmap object}

\item{arguments}{The validated arguments for heat_tree in a list.}
}
\value{
a taxmap
}
\description{
Reformat heat_tree arguments to taxmap
}
\keyword{internal}
