% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new--heat_tree--internal.R
\name{heat_tree_labels}
\alias{heat_tree_labels}
\title{Determine labels used in heat_tree}
\usage{
heat_tree_labels(obj)
}
\arguments{
\item{obj}{a taxmap}
}
\value{
a taxmap
}
\description{
Determine the labels used in a heat_tree
}
\keyword{internal}
