% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heat_tree_matrix.R
\name{heat_tree_matrix}
\alias{heat_tree_matrix}
\title{Plot a matrix of heat trees}
\usage{
heat_tree_matrix(obj, dataset, label_small_trees = FALSE, key_size = 0.6,
  seed = 1, ...)
}
\arguments{
\item{obj}{A \code{taxmap} object}

\item{dataset}{The name of a table in \code{obj$data} that is the output of 
\code{\link{compare_treatments}} or in the same format.}

\item{label_small_trees}{If \code{TRUE} add labels to small trees as well as 
the key tree. Otherwise, only the key tree will be labled.}

\item{key_size}{The size of the key tree relative to the whole graph. For
example, 0.5 means half the width/height of the graph.}

\item{seed}{That random seed used to make the graphs.}
}
\description{
Plot a matrix of heat trees for showing parwise comparision. A larger, 
labelled tree serves as a key for the matrix of smaller unlabelled trees.
}
