% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new--heat_tree.R
\name{heat_tree_plot}
\alias{heat_tree_plot}
\title{Plots the object made by \code{\link{heat_tree_data}}}
\usage{
heat_tree_plot(obj, margin_size = 0, aspect_ratio = 1,
  background_color = "#FFFFFF00", output_file = NULL)
}
\arguments{
\item{margin_size}{(\code{numeric} of length 2) The horizontal and vertical margins, measured in
proportion of graphic area. c(left, right, bottom, top). Default: \code{0, 0, 0, 0}.}

\item{aspect_ratio}{The aspect_ratio of the plot. Default: determined by layout.}

\item{background_color}{The background color of the plot. Default: Transparent}

\item{output_file}{The path to one or more files to save the plot in using
\code{\link[ggplot2]{ggsave}}. The type of the file will be determined by the extension given.
Default: Do not save plot.}
}
\description{
Uses the data in \code{\link{heat_tree_data}} to make a ggplot2 plot of a heat tree.
Using \code{\link{heat_tree_data}} and \code{\link{heat_tree_plot}} is
intended for advanced users when customizations of the plot must be scripted.
}
