% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new--heat_tree--internal.R
\name{heat_tree_shapes}
\alias{heat_tree_shapes}
\title{Determine shpaes used in heat_tree}
\usage{
heat_tree_shapes(obj)
}
\arguments{
\item{obj}{a taxmap}
}
\value{
a taxmap
}
\description{
Determine the shapes used in a heat_tree
}
\keyword{internal}
