% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new--heat_tree--internal.R
\name{heat_tree_sizes}
\alias{heat_tree_sizes}
\title{Determine sizes used in heat_tree}
\usage{
heat_tree_sizes(obj)
}
\arguments{
\item{obj}{a taxmap}
}
\value{
a taxmap
}
\description{
Determine the sizes used in a heat_tree
}
\keyword{internal}
