% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new--heat_tree--transformations.R
\name{heat_tree_transform_data}
\alias{heat_tree_transform_data}
\title{Apply transformations}
\usage{
heat_tree_transform_data(obj)
}
\arguments{
\item{obj}{The input taxmap object}
}
\value{
a taxmap
}
\description{
Apply transformations to user-supplied data contained in a taxmap object.
Used in the heat_tree function.
}
\keyword{internal}
