% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heat_tree--internal.R
\name{heat_tree_transform_funcs}
\alias{heat_tree_transform_funcs}
\title{Transformation functions}
\usage{
heat_tree_transform_funcs(inverse = FALSE)
}
\arguments{
\item{inverse}{(\code{logical} of length 1) If \code{TRUE}, return the inverse of the functions.}
}
\description{
Functions used by plotting functions to transform data and their inverses.
}
\keyword{internal}
