% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_taxa--taxmap--printers.R
\name{highlight_taxon_ids}
\alias{highlight_taxon_ids}
\title{Highlight taxon ID column}
\usage{
highlight_taxon_ids(table_text, header_index, row_indexes)
}
\arguments{
\item{table_text}{The print out of the table in a character vector, one element per line.}

\item{header_index}{The row index that contains the table column names}

\item{row_indexes}{The indexes of the rows to be formatted.}
}
\description{
Changes the font of a taxon ID column in a table print out.
}
