% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_documentation.R
\name{hmp_otus}
\alias{hmp_otus}
\title{A HMP subset}
\format{
A 1,000 x 52 tibble.
}
\source{
Subset from data available at https://www.hmpdacc.org/hmp/HMQCP/
}
\description{
A subset of the Human Microbiome Project abundance matrix produced by QIIME.
It contains OTU ids, taxonomic lineages, and the read counts for 50 samples.
See \code{\link{hmp_samples}} for the matching dataset of sample information.
}
\details{
The 50 samples were randomly selected such that there were 10 in each of 5
treatments: "Saliva", "Throat", "Stool", "Right_Antecubital_fossa",
"Anterior_nares". For each treatment, there were 5 samples from men and 5
from women.
}
\seealso{
Other hmp_data: 
\code{\link{hmp_samples}}
}
\concept{hmp_data}
\keyword{data}
