% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_documentation.R
\name{hmp_samples}
\alias{hmp_samples}
\title{Sample information for HMP subset}
\format{
A 50 x 3 tibble.
}
\source{
Subset from data available at https://www.hmpdacc.org/hmp/HMQCP/
}
\description{
The sample information for a subset of the Human Microbiome Project data. It
contains the sample ID, sex, and body site for each sample in the abundance
matrix stored in \code{\link{hmp_otus}}. The "sample_id" column corresponds
to the column names of \code{\link{hmp_otus}}.
}
\details{
The 50 samples were randomly selected such that there were 10 in each of 5
treatments: "Saliva", "Throat", "Stool", "Right_Antecubital_fossa",
"Anterior_nares". For each treatment, there were 5 samples from men and 5
from women. "Right_Antecubital_fossa" was renamed to "Skin" and
"Anterior_nares" to "Nose".
}
\seealso{
Other hmp_data: 
\code{\link{hmp_otus}}
}
\concept{hmp_data}
\keyword{data}
