% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_taxa--taxmap--internal.R
\name{init_taxmap_data}
\alias{init_taxmap_data}
\title{Convert `data` input for Taxamp}
\usage{
init_taxmap_data(self, data, input_ids, assume_equal = TRUE)
}
\arguments{
\item{self}{The newly created [taxmap()] object}

\item{data}{The `data` variable passed to the `Taxmap` constructor}

\item{input_ids}{The taxon IDs for the inputs that made the taxonomy}

\item{assume_equal}{If `TRUE`, and a data set length is the same as the
`input_ids` length, then assume that `input_ids` applies to the data set as
well.}
}
\value{
A `data` variable with the right format
}
\description{
Make sure `data` is in the right format and complain if it is not. Then, add
a `taxon_id` column to data with the same length as the input
}
\keyword{internal}
