% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heat_tree--vertex_size.R
\name{inter_circle_gap}
\alias{inter_circle_gap}
\title{Finds the gap/overlap of circle coordinates}
\usage{
inter_circle_gap(x, y, r)
}
\arguments{
\item{x}{(\code{numeric} of length 1) x coordinate of center}

\item{y}{(\code{numeric} of length 1) y coordinate of center}

\item{r}{(\code{numeric} of length 1) The diameter of the circle.}
}
\description{
Given a set of x, y coordinates and corresponding radii return the gap between every possible 
combination.
}
\keyword{internal}
