% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_taxonomy--legend.R
\name{inverse}
\alias{inverse}
\title{Generate the inverse of a function}
\usage{
inverse(f, interval)
}
\description{
http://stackoverflow.com/questions/10081479/solving-for-the-inverse-of-a-function-in-r
}

