% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new--heat_tree--internal.R
\name{is_categorical}
\alias{is_categorical}
\title{Check if a list/vector is categorical}
\usage{
is_categorical(x)
}
\description{
Check if a list/vector that is the value for a color aesthetic is composed of factors and perhaps characters, but not numbers.
}
