% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{is_color_char}
\alias{is_color_char}
\title{Test if a vector/list is colors}
\usage{
is_color_char(input)
}
\arguments{
\item{input}{A vector or list}
}
\description{
Test each element of a list is a hex color code or a color from colors()
}
\keyword{internal}
