% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa--classified_class.R
\name{item_count}
\alias{item_count}
\title{Count items in \code{\link{classified}}}
\usage{
item_count(obj, subset = taxon_ids(obj))
}
\arguments{
\item{obj}{(\code{\link{classified}})}

\item{subset}{(\code{character}) The \code{taxon_id}s to get counts for.}
}
\value{
\code{numeric} of length 1
}
\description{
Count items in \code{\link{classified}}
}
\keyword{internal}

