% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxmap--column_generators.R
\name{n_items}
\alias{n_items}
\title{Count items in \code{\link{taxmap}}}
\usage{
n_items(obj, subset = obj$taxon_data$taxon_ids)
}
\arguments{
\item{obj}{(\code{\link{taxmap}})}

\item{subset}{(\code{character}) The \code{taxon_ids}s to get counts for.}
}
\value{
\code{numeric}
}
\description{
Count items in \code{\link{taxmap}}
}

