% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa--classified_class.R
\name{item_data}
\alias{item_data}
\title{Return item data from \code{\link{classified}}}
\usage{
item_data(obj, col_subset = item_data_colnames(obj),
  row_subset = 1:nrow(obj$item_data), drop = FALSE)
}
\arguments{
\item{obj}{(\code{\link{classified}})}

\item{col_subset}{(\code{character})
The names of columns, either user defined or generated using \code{taxon_funcs}.
Default: All columns.}

\item{row_subset}{(\code{character})
The taxon_ids of a subset of \code{obj}.
Default: All rows.}

\item{drop}{(\code{logical} of length 1) If \code{TRUE}, if \code{subset} is a single column
name, then a \code{vector} is returned instead of a \code{data.frame}}
}
\value{
\code{data.frame} with rows corresponding to items in input
}
\description{
Return a table of data associated with items of an object of type
\code{\link{classified}}.
}

