% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classified--internal.R
\name{item_data_colnames}
\alias{item_data_colnames}
\title{Get column names of item_data}
\usage{
item_data_colnames(obj)
}
\arguments{
\item{obj}{a \code{classified} object}
}
\value{
\code{character}
}
\description{
Get column names of item_data without calculating columns
}

