% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_documentation.R
\name{its1_ex_data}
\alias{its1_ex_data}
\title{Example of ITS1 fungal data}
\format{An object of type \code{\link{taxmap}}}
\description{
A dataset containing information from 170 Chytridiomycota ITS sequences from the ITS1 reference database.
}
\examples{
\dontrun{

file_path <- system.file("extdata", "its1_chytridiomycota_hmm.fasta", package = "metacoder")
sequences <- ape::read.FASTA(file_path)
its1_ex_data <- extract_tax_data(names(sequences),
                                 regex = "^(.*)_ITS1_HMM\\\\|(.*)\\\\|tax_id:(.*)\\\\|(.*)$",
                                 key = c("seq_id", my_name = "info",
                                         my_id = "info", description = "info"),
                                 database = "ncbi")
}

}
\seealso{
Other example_fastas: \code{\link{bryophytes_ex_data}},
  \code{\link{contaminants}},
  \code{\link{genbank_ex_data}}, \code{\link{pr2_ex_data}},
  \code{\link{silva_ex_data}},
  \code{\link{unite_ex_data_1}},
  \code{\link{unite_ex_data_2}}
}
