% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heat_tree--internal.R
\name{label_bounds}
\alias{label_bounds}
\title{Bounding box coords for labels}
\usage{
label_bounds(label, x, y, height, rotation, just)
}
\arguments{
\item{x}{Horizontal position of center of text grob}

\item{y}{Vertical position of center of text grob}

\item{height}{Height of text grob}

\item{rotation}{Rotation in radians}

\item{just}{Justification. e.g. "left-top"}
}
\description{
Given a position, size, rotation, and justification of a label, calculate the bounding box coordinates
}
\keyword{internal}
