% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_plotting.R
\name{layout_functions}
\alias{layout_functions}
\title{Layout functions}
\usage{
layout_functions(name = NULL, graph = NULL, intitial_coords = NULL, ...)
}
\arguments{
\item{name}{(\code{character} of length 1 OR NULL) name of algorithm. Leave \code{NULL} to 
see all options.}

\item{graph}{(\code{igraph}) The graph to generate the layout for.}

\item{intitial_coords}{(\code{matrix}) Initial vertex layout to bawse new layout off of.}

\item{...}{(other arguments) Passed to igraph layout function used.}
}
\description{
Functions used to determine graph layout.
Calling the function with no parameters returns available function names.
}

