% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_taxa--internal.R
\name{length_of_thing}
\alias{length_of_thing}
\title{Check length of thing}
\usage{
length_of_thing(obj)
}
\arguments{
\item{obj}{}
}
\value{
\code{numeric} of length 1.
}
\description{
Check the length of an object, be it list, vector, or table.
}
\keyword{internal}
