% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heat_tree--internal.R
\name{look_for_na}
\alias{look_for_na}
\title{Look for NAs in parameters}
\usage{
look_for_na(taxon_ids, args)
}
\arguments{
\item{args}{(\code{character}) The names of arguments to verify.}
}
\description{
Look for NAs in parameters
}
\keyword{internal}
