% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{make_fasta_with_u_replaced}
\alias{make_fasta_with_u_replaced}
\title{Make a temporary file U's replaced with T}
\usage{
make_fasta_with_u_replaced(file_path)
}
\arguments{
\item{file_path}{}
}
\value{
A path to a temporary file.
}
\description{
Make a temporary fasta file U's replaced with T without reading in whole file.
}
\keyword{internal}
