% Generated by roxygen2 (4.0.2): do not edit by hand
\name{make_new_ids}
\alias{make_new_ids}
\title{Generate new unique ids}
\usage{
make_new_ids(count, existing)
}
\arguments{
\item{count}{(\code{numeric} of length 1) The number of new unique ids to generate}

\item{existing}{(\code{character}) Existing unique ids. These will not appear in the output.}
}
\description{
Makes a vector of unique ids that differ from a previously defined set of unique ids.
}

