% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resizing_text_grob.R
\name{make_text_grobs}
\alias{make_text_grobs}
\title{Create a list of text grobs}
\usage{
make_text_grobs(label, x, y, size, color, rotation, justification, x_range,
  y_range)
}
\arguments{
\item{label}{Text to display}

\item{x}{x}

\item{y}{y}

\item{size}{The height of text}

\item{color}{The color of text}

\item{rotation}{The rotation of text. In radians.}

\item{justification}{A \code{character} indicating the justification of text. 
Is split by \code{"-"}}

\item{x_range}{The minimum and maximum x value in the plot}

\item{y_range}{The minimum and maximum y value in the plot}
}
\value{
\code{list} of \code{\link{resizingTextGrob}}
}
\description{
Creates a list of resizing text grobs
}
\keyword{internal}

