% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{map_unique}
\alias{map_unique}
\title{Run a function on unique values of a iterable}
\usage{
map_unique(input, func, ...)
}
\arguments{
\item{input}{What to pass to \code{func}}

\item{func}{(\code{function})}

\item{...}{passend to \code{func}}
}
\description{
Run a function on unique values of a iterable
}
\keyword{internal}

