% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/distance_calculation.R
\name{model_kimura2}
\alias{model_kimura2}
\title{Calculate kimura-2 from unaligned sequences}
\usage{
model_kimura2(x, y, ...)
}
\arguments{
\item{x}{A character vector of DNA sequence.}

\item{y}{A character vector of DNA sequence.}

\item{...}{Additional arguments are passed to Biostrings::pairwiseAlignment}
}
\value{
A Kimura2 distance metric as a numeric vector of length 1.
}
\description{
The two input sequences are aligned using Biostrings::pairwiseAlignment and the kimura
2-parameter distance metric is calculated.
}
\details{
This is not set up to deal with IUPAC ambiguity codes. It will treat ambiguous SNPs as
transversions. Use with caution; this is still in the development stage.
}

