% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_plotting.R
\name{molten_dist}
\alias{molten_dist}
\title{Get all distances beween points}
\usage{
molten_dist(x, y)
}
\arguments{
\item{x}{(\code{numeric} of length 1) x coordinate}

\item{y}{(\code{numeric} of length 1) y coordinate}
}
\value{
A \code{data.frame}
}
\description{
Returns the distances bewteen every possible combination of two points.
}
\examples{
molten_dist(x = 1:5, y = 1:5)


}
\keyword{internal}

