% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_taxa--internal.R
\name{multi_sep_split}
\alias{multi_sep_split}
\title{Like `strsplit`, but with multiple separators}
\usage{
multi_sep_split(input, split, ...)
}
\arguments{
\item{input}{A character vector}

\item{split}{One or more separators to use to split `input`}

\item{...}{Passed to [base::strsplit()]}
}
\description{
Splits items in a vector by multiple separators.
}
\keyword{internal}
