% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classified--mutate.R
\name{mutate_taxa}
\alias{mutate_taxa}
\title{Add columns to \code{\link{classified}} objects}
\usage{
mutate_taxa(.data, ...)
}
\arguments{
\item{.data}{\code{\link{classified}}}

\item{...}{One or more column names to add to the new object.
Newly created columns can be referenced in the same function call.}
}
\value{
An object of type \code{\link{classified}}
}
\description{
Add columns to the \code{taxon_data} in \code{\link{classified}} objects.
Any column name that appears in \code{taxon_data(.data)} can be used as if it was a vector on its own.
}

