% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{my_print}
\alias{my_print}
\title{Print something}
\usage{
my_print(..., verbose = TRUE)
}
\arguments{
\item{...}{Something to print}

\item{verbose}{If \code{FALSE}, do not print anything.}
}
\description{
The standard print function for this package. This is a wrapper to make
package-wide changes easier.
}
\keyword{internal}
