% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_taxa--taxmap--docs.R
\name{n_obs_1}
\alias{n_obs_1}
\title{Count observation assigned in [taxmap()]}
\arguments{
\item{obj}{([taxmap()])}

\item{data}{Dataset name, index, or a logical vector that indicates which dataset in
`obj$data` to add columns to.}

\item{target}{DEPRECIATED. use "data" instead.}
}
\value{
`numeric`
}
\description{
Count observations for each taxon in a data set in a [taxmap()] object. This
includes observations for the specific taxon but NOT the observations of its
subtaxa. "Observations" in this sense are the items (for list/vectors) or
rows (for tables) in a dataset. By default, observations in the first data
set in the [taxmap()] object is used.  For example, if the data set is a
table, then a value of 3 for a taxon means that their are 3 rows in that
table assigned to that taxon.
\preformatted{
obj$n_obs_1(data)
n_obs_1(obj, data)}
}
\examples{
# Get number of observations for each taxon in first dataset
n_obs_1(ex_taxmap)

# Get number of observations in a specified data set
n_obs_1(ex_taxmap, "info")
n_obs_1(ex_taxmap, "abund")

# Filter taxa using number of observations in the first table
filter_taxa(ex_taxmap, n_obs_1 > 0)

}
\seealso{
Other taxmap data functions: 
\code{\link{n_obs}()}
}
\concept{taxmap data functions}
