% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ncbi_children_uid}
\alias{ncbi_children_uid}
\title{Search NCBI for children of a taxon}
\usage{
ncbi_children_uid(name = NULL, id = NULL, start = 0, max_return = 1000,
  ancestor = NULL)
}
\arguments{
\item{name}{(\code{character}) The string to search for. Only exact matches found the name given
will be returned. Not compatible with \code{id}.}

\item{id}{(\code{character}) The uid to search for. Not compatible with \code{name}.}

\item{start}{The first record to return. If omitted, the results are returned from the first
record (start=0).}

\item{max_return}{(\code{numeric; length=1}) The maximum number of children to return.}

\item{ancestor}{(\code{character}) The ancestor of the taxon being searched for. This is useful
if there could be more than one taxon with the same name. Has no effect if \code{id} is used.}
}
\value{
A list of character vectors of children uids. The names of the list elements are the
  input \code{name} or \code{id} values.
}
\description{
Search the NCBI Taxonomy databse for uids of children of taxa. Taxa can be referenced by name
or uid. Referencing by name is faster.
}
\details{
In a few cases, different taxa have the same name (e.g. Satyrium; see examples). If one of these
are searched for then the children of both taxa will be returned. This can be avoided by
using a uid instead of the name or specifying an ancestor. If an ancestor is provided, only
children of both the taxon and its ancestor are returned. This will only fail if there are two
taxa with the same name and the same specified ancestor.
}
\examples{
ncbi_children_uid(name="Satyrium") #Satyrium is the name of two different genera
ncbi_children_uid(name="Satyrium", ancestor="Eumaeini") # A genus of butterflies
ncbi_children_uid(name="Satyrium", ancestor="Orchidaceae") # A genus of orchids
ncbi_children_uid(id="266948") #"266948" is the uid for the butterfly genus
ncbi_children_uid(id="62858") #"62858" is the uid for the orchid genus
}

