% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_download.R
\name{ncbi_sequence}
\alias{ncbi_sequence}
\title{Downloads sequences from ids}
\usage{
ncbi_sequence(ids, batch_size = 100)
}
\arguments{
\item{ids}{(\code{character}) One or more accession numbers to get sequences for}

\item{batch_size}{(\code{numeric} of length 1) The number of sequences to request in each query.
To large of values might case failures and too small will increase time to completion.}
}
\value{
(\code{list} of \code{character})
}
\description{
Downloads the sequences associated with GenBank accession ids.
}
\keyword{internal}
