% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new--heat_tree--shapes.R
\name{node_shape_pentagon}
\alias{node_shape_pentagon}
\title{Node shape function: pentagon}
\usage{
node_shape_pentagon(node_color = "grey")
}
\arguments{
\item{node_color}{The color of the pentagon fill}
}
\description{
A function used to plot nodes in a heat tree. This function plots a simple pentagon in a solid color.
}
