% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new--heat_tree--shapes.R
\name{node_shape_square}
\alias{node_shape_square}
\title{Node shape function: square}
\usage{
node_shape_square(node_color = "grey")
}
\arguments{
\item{node_color}{The color of the square fill}
}
\description{
A function used to plot nodes in a heat tree. This function plots a simple square in a solid color.
}
