% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new--heat_tree--shapes.R
\name{node_shape_triangle}
\alias{node_shape_triangle}
\title{Node shape function: triangle}
\usage{
node_shape_triangle(node_color = "grey")
}
\arguments{
\item{node_color}{The color of the triagle fill}
}
\description{
A function used to plot nodes in a heat tree. This function plots a simple triangle in a solid color.
}
