% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/distance_calculation.R
\name{pairwise_distance}
\alias{pairwise_distance}
\title{Calculates a pairwise distance matrix from unaligned sequences}
\usage{
pairwise_distance(sequences, seq_name = NULL)
}
\arguments{
\item{sequences}{A list or vector of sequence data}
}
\value{
A distance matrix of class \code{dist}
}
\description{
This is still in the developmental stage and currently only uses the kimura 2-parameter metric.
}

