% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa--classified_class.R
\name{parent_ids}
\alias{parent_ids}
\title{Get parent taxon IDs}
\usage{
parent_ids(obj, subset = taxon_ids(obj))
}
\arguments{
\item{obj}{(\code{classified})
The \code{classified} object containing taxon information to be queried.}

\item{subset}{(\code{character})
Taxon IDs for which parents will be returned.
Default: All taxon in \code{obj} will be used.}
}
\value{
\code{character}
}
\description{
Return the taxon IDs for a parents of taxon IDs in an \code{\link{classified}} object.
}
\keyword{internal}

