% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_taxa--taxmap--option_parsers.R
\name{parse_dataset}
\alias{parse_dataset}
\title{Parse options specifying datasets}
\usage{
parse_dataset(obj, data, must_be_valid = TRUE, needed = TRUE, rm_na = TRUE)
}
\arguments{
\item{obj}{The taxmap object.}

\item{data}{The name/index of datasets in a taxmap object to use. Can also be a logical vector
of length equal to the number of datasets.}

\item{must_be_valid}{If TRUE, all datasets specified must be valid or an error occurs.}

\item{needed}{If TRUE, at least one dataset must be specified or an error occurs.}

\item{rm_na}{If TRUE, then invalid datasets do result in NAs in the output.}
}
\value{
The indexes for the datasets selected
}
\description{
Parse options specifying datasets in taxmap objects
}
\keyword{internal}
