% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_taxa--taxonomy--internal.R
\name{parse_heirarchies_to_taxonomy}
\alias{parse_heirarchies_to_taxonomy}
\title{Infer edge list from hierarchies}
\usage{
parse_heirarchies_to_taxonomy(heirarchies)
}
\value{
A list of [hierarchy()] objects.
}
\description{
Infer edge list and unique taxa from hierarchies.
}
\keyword{internal}
