% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{parse_hmp_qiime}
\alias{parse_hmp_qiime}
\title{Parse HMP QIIME results}
\usage{
parse_hmp_qiime(otu_file, mapping_file, min_abundance = 1, max_otus = -1)
}
\arguments{
\item{otu_file}{(\code{character} of length 1) A file path or URL to the OTU table.}

\item{mapping_file}{(\code{character} of length 1) A file path or URL to the sample mapping file.}

\item{min_abundance}{(\code{numeric} of length 1) Do not return rows less abundance. This can make the output object much smaller.}

\item{max_otus}{(\code{numeric} of length 1) Only parse some number of OTUs. Useful for making small datasets for testing.}
}
\value{
A \code{\link{taxmap}} object
}
\description{
NOTE: Not extensively tested
Parses the results of the Human Microbiome Project QIIME analysis of the 16S metagenomic data.
This is mostly a wrapper for \code{\link{extract_taxonomy}}.
}
\seealso{
Other parsers: \code{\link{parse_mothur_tax_summary}},
  \code{\link{parse_mothur_taxonomy}},
  \code{\link{parse_newick}}, \code{\link{parse_phyloseq}},
  \code{\link{parse_qiime_biom}}, \code{\link{parse_rdp}},
  \code{\link{parse_silva_fasta}},
  \code{\link{parse_unite_general}}
}
