% Generated by roxygen2 (4.0.2): do not edit by hand
\name{parse_lineage}
\alias{parse_lineage}
\title{Parse taxonomic classification strings}
\usage{
parse_lineage(lineage, taxon_sep, rank_sep, rev_taxon, rev_rank,
  taxon_col_name = "taxon", rank_col_name = "rank")
}
\arguments{
\item{lineage}{(\code{character}) Taxonomic classifications in the form of delimited strings.}

\item{taxon_sep}{(\code{character; length = 1}) Character that deliminates individual taxa
and potentially their ranks}

\item{rank_sep}{(\code{character; length = 1}) Character that deliminates a taxon and its rank.}

\item{rev_taxon}{If \code{TRUE}, the rank order of taxa read in a lineage is reversed to be
specific to broad.}

\item{rev_rank}{If \code{TRUE}, the rank information come after the taxon information.}

\item{taxon_col_name}{(\code{character}) The name of the taxon column in the \code{data.frame}s
representing each classification.}

\item{rank_col_name}{(\code{character}) The name of the rank column in the \code{data.frame}s
representing each classification.}
}
\description{
Parses taxonomic classification strings into a list of \code{data.frame}
Can extract taxonomic ranks if present.
}

