% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mothur.R
\name{parse_mothur_summary}
\alias{parse_mothur_summary}
\title{Parse mothur classification summary file}
\usage{
parse_mothur_summary(file_path, unclassified = FALSE)
}
\arguments{
\item{file_path}{(\code{character} of length 1)
The file path to the input file.}
}
\value{
\code{\link{classified}}
}
\description{
Parse mothur classification summary file
}

