% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{parse_mothur_tax_summary}
\alias{parse_mothur_tax_summary}
\title{Parse mothur classification summary file}
\usage{
parse_mothur_tax_summary(file_path = NULL, text = NULL,
  unclassified = FALSE)
}
\arguments{
\item{file_path}{(\code{character} of length 1)
The file path to the input file.}

\item{unclassified}{(\code{logical} of length 1)
If \code{FALSE}, remove any unclassified rows.}
}
\value{
\code{\link{taxmap}}
}
\description{
Parse mothur classification summary file
}
