% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{parse_mothur_taxonomy}
\alias{parse_mothur_taxonomy}
\title{Parse mothur taxonomy file}
\usage{
parse_mothur_taxonomy(file = NULL, text = NULL)
}
\arguments{
\item{file}{(\code{character} of length 1) The file path to the input file.}

\item{text}{(\code{character}) An alternate input to "file". The contents of
the file as a character. Either "file" or "text" must be used, but not both.}
}
\value{
\code{\link{taxmap}}
}
\description{
Parse the `.taxonomy`` file that is returned by the `Classify.seqs` command
in mothur.
}
\details{
The input file has a format like:

AY457915	Bacteria(100);Firmicutes(99);Clostridiales(99);Joh...

or..

AY457915	Bacteria;Firmicutes;Clostridiales;Johnsonella_et_r...
}
