% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{parse_newick}
\alias{parse_newick}
\title{Parse a Newick file}
\usage{
parse_newick(file)
}
\arguments{
\item{file}{(\code{character} of length 1) The file path to the input file.}
}
\value{
\code{\link{taxmap}}
}
\description{
Parse a Newick file
}
\details{
The input file has a format like:

\preformatted{
}
}
\seealso{
Other parsers: \code{\link{parse_hmp_qiime}},
  \code{\link{parse_mothur_tax_summary}},
  \code{\link{parse_mothur_taxonomy}},
  \code{\link{parse_phyloseq}}
}
