% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{parse_phylo}
\alias{parse_phylo}
\title{Parse a phylo object}
\usage{
parse_phylo(obj)
}
\arguments{
\item{obj}{A phylo object from the ape package.}
}
\value{
\code{\link{taxmap}}
}
\description{
Parses a phylo object from the ape package.
}
\seealso{
Other parsers: \code{\link{parse_dada2}},
  \code{\link{parse_edge_list}},
  \code{\link{parse_greengenes}},
  \code{\link{parse_mothur_tax_summary}},
  \code{\link{parse_mothur_taxonomy}},
  \code{\link{parse_newick}}, \code{\link{parse_phyloseq}},
  \code{\link{parse_qiime_biom}}, \code{\link{parse_rdp}},
  \code{\link{parse_silva_fasta}},
  \code{\link{parse_ubiome}},
  \code{\link{parse_unite_general}}
}
\concept{parsers}
