% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mothur.R
\name{parse_summary_seqs}
\alias{parse_summary_seqs}
\title{Parse summary.seqs output}
\usage{
parse_summary_seqs(text)
}
\arguments{
\item{text}{The text output of \code{summary.seqs}}
}
\value{
A \code{data.frame} of statistics
}
\description{
Extract statistics from the command line output of mothur command \code{summary.seqs} and
return the resutls in a \code{data.frame}
}

