% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{parse_unite_general}
\alias{parse_unite_general}
\title{Parse UNITE general release FASTA}
\usage{
parse_unite_general(file, include_seqs = TRUE)
}
\arguments{
\item{file}{(\code{character} of length 1) The file path to the input file.}

\item{include_seqs}{(\code{logical} of length 1) If \code{TRUE}, include
sequences in the output object.}
}
\value{
\code{\link{taxmap}}
}
\description{
Parse the UNITE general release FASTA file
}
\details{
The input file has a format like:

\preformatted{
>Glomeromycota_sp|KJ484724|SH523877.07FU|reps|k__Fungi;p__Glomeromycota;c__unid...
ATAATTTGCCGAACCTAGCGTTAGCGCGAGGTTCTGCGATCAACACTTATATTTAAAACCCAACTCTTAAATTTTGTAT...
}
}
\seealso{
Other parsers: \code{\link{parse_greengenes}},
  \code{\link{parse_mothur_tax_summary}},
  \code{\link{parse_mothur_taxonomy}},
  \code{\link{parse_newick}}, \code{\link{parse_phyloseq}},
  \code{\link{parse_qiime_biom}}, \code{\link{parse_rdp}},
  \code{\link{parse_silva_fasta}}
}
