% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_plotting.R
\name{plot.classified}
\alias{plot.classified}
\title{Plot a taxonomic tree}
\usage{
\method{plot}{classified}(classified_data, ...)
}
\arguments{
\item{classified_data}{An object of class \link{classified}.}

\item{...}{Passed to \link{plot_taxonomy}}
}
\description{
Plot a taxonomic tree
}

