% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_image_tree}
\alias{plot_image_tree}
\title{Makes igraph from images}
\usage{
plot_image_tree(graph, image_file_paths, labels = NA, scaling = 1,
  exclude = c(), root_index = 1, label_color = "black")
}
\description{
Makes igraph from images
}

