% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot_value_tree}
\alias{plot_value_tree}
\title{Makes igraph from values}
\usage{
plot_value_tree(graph, values, labels = NA, scaling = 1, exclude = c(),
  root_index = 1, label_color = "black", display = FALSE, fade = FALSE,
  legend_text = "", value_range = c(0, 1), highlight_outliers = TRUE,
  background = "#00000000", save = NULL)
}
\arguments{
\item{graph}{An igraph graph object or an adjacency list (i.e. edge list) represented by
something that can be coerced to a two column \code{data.frame} where the first column is
every unique taxon id and the second is its parent/super taxon id.}
}
\description{
Makes igraph from values
}

