% Generated by roxygen2 (4.0.2): do not edit by hand
\name{postDrawDetails.resizingTextGrob}
\alias{postDrawDetails.resizingTextGrob}
\title{Clean up after the drawing.}
\usage{
\method{postDrawDetails}{resizingTextGrob}(x)
}
\description{
Clean up after the drawing.
}
\details{
Taken from: http://ryouready.wordpress.com/2012/08/01/creating-a-text-grob-that-automatically-adjusts-to-viewport-size/
}

