% Generated by roxygen2 (4.0.2): do not edit by hand
\name{preDrawDetails.resizingTextGrob}
\alias{preDrawDetails.resizingTextGrob}
\title{Adjusts text size to viewport}
\usage{
\method{preDrawDetails}{resizingTextGrob}(x)
}
\description{
Automatically called before any drawing occures.
}
\details{
Taken from: http://ryouready.wordpress.com/2012/08/01/creating-a-text-grob-that-automatically-adjusts-to-viewport-size/
}

