% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_taxa--taxmap--printers.R
\name{prefixed_print}
\alias{prefixed_print}
\title{Print a object with a prefix}
\usage{
prefixed_print(x, prefix, ...)
}
\arguments{
\item{x}{What to print.}
}
\description{
Print a object with a prefix. Uses the standard print method of the object.
}
\keyword{internal}
