% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primersearch.R
\name{primersearch_is_installed}
\alias{primersearch_is_installed}
\title{Test if primersearch is installed}
\usage{
primersearch_is_installed(must_be_installed = TRUE)
}
\arguments{
\item{must_be_installed}{(\code{logical} of length 1)
If \code{TRUE}, throw an error if primersearch is not installed.}
}
\value{
\code{logical} of length 1
}
\description{
Test if primersearch is installed
}
\keyword{internal}
