% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_taxa--taxmap--printers.R
\name{print__vector}
\alias{print__vector}
\title{Generic vector printer}
\usage{
print__vector(
  obj,
  data,
  name,
  prefix,
  max_width,
  max_rows,
  type = class(data)[1]
)
}
\arguments{
\item{obj}{The taxmap object containing the thing to print}

\item{data}{Something to print}

\item{name}{The name of the thing to print}

\item{prefix}{What to put before the thing printed. Typically a space.}

\item{max_width}{Maximum width in number of characters to print}

\item{max_rows}{Maximum number of rows to print}

\item{type}{The name of the type of vector to print (e.g. numeric).}
}
\description{
Print a vector for the print method of taxmap objects.
}
\details{
Which print method is called is determined by its name, so changing the name
of this function will change when it is called.
}
\seealso{
Other taxmap print methods: 
\code{\link{print__character}()},
\code{\link{print__data.frame}()},
\code{\link{print__default_}()},
\code{\link{print__factor}()},
\code{\link{print__integer}()},
\code{\link{print__list}()},
\code{\link{print__logical}()},
\code{\link{print__matrix}()},
\code{\link{print__numeric}()},
\code{\link{print__ordered}()},
\code{\link{print__tbl_df}()}
}
\concept{taxmap print methods}
\keyword{internal}
