% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_taxa--taxmap--printers.R
\name{print_item}
\alias{print_item}
\title{Print a item}
\usage{
print_item(
  obj,
  data,
  name = NULL,
  max_rows = 3,
  max_items = 3,
  max_width = getOption("width") - 10,
  prefix = ""
)
}
\arguments{
\item{obj}{The taxmap object containing the thing to print}

\item{data}{The item to be printed}

\item{max_rows}{(`numeric` of length 1) The maximum number of rows in
tables to print.}

\item{max_items}{(`numeric` of length 1) The maximum number of list
items to print.}

\item{max_width}{(`numeric` of length 1) The maximum number of
characters to print.}

\item{prefix}{(`numeric` of length 1) What to print in front of each
line.}
}
\description{
Used to print each item in the `taxmap` print method.
}
\keyword{internal}
