% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_taxa--taxonomy--docs.R
\name{print_tree}
\alias{print_tree}
\title{Print a text tree}
\arguments{
\item{obj}{A \code{taxonomy} or \code{taxmap} object}

\item{value}{What data to return. Default is taxon names. Any result of
[all_names()] can be used, but it usually only makes sense to use data with
one value per taxon, like taxon names.}
}
\description{
Print a text-based tree of a [taxonomy()] or [taxmap()] object.
}
\examples{
print_tree(ex_taxmap)

}
