% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_taxa--taxmap--internal.R
\name{progress_lapply}
\alias{progress_lapply}
\title{lappy with progress bars}
\usage{
progress_lapply(X, FUN, progress = interactive(), ...)
}
\arguments{
\item{X}{The thing to iterate over}

\item{FUN}{The function to apply to each element}

\item{progress}{(logical of length 1) Whether or not to print a progress bar. Default is to only print a progress bar during interactive use.}

\item{...}{Passed to function}
}
\value{
list
}
\description{
Immitates lapply with optional progress bars
}
\keyword{internal}
