% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_plotting.R
\name{qualitative_palette}
\alias{qualitative_palette}
\title{The defualt qualitative color palette}
\usage{
qualitative_palette()
}
\description{
The default color palette for qualitative data
}
\keyword{internal}

