% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_plotting.R
\name{quantative_palette}
\alias{quantative_palette}
\title{The defualt quantative color palette}
\usage{
quantative_palette()
}
\description{
The default color palette for quantative data
}
\keyword{internal}

