% Generated by roxygen2 (4.0.2): do not edit by hand
\name{query_taxon}
\alias{query_taxon}
\title{Queries sequences for a given taxon}
\usage{
query_taxon(query_id, taxon, key = character(0), type)
}
\arguments{
\item{query_id}{A character vector of length 1 specifying the variable name the query will be
bound to.}

\item{taxon}{A character vector of taxa. Specifing multiple taxa is equivalent to
running the command multiple times with each taxa and concatenating the results.}

\item{key}{A character vector of keywords. All keywords must be present for a given
sequence to be returned.}

\item{type}{The type of the sequence to return (e.g. RRNA). Use `getType()` to see options.}
}
\value{
A instance of class seqinr::qaw, as returned by seqinr::query.
}
\description{
Produces a sequinr query object that can be used to download sequences.
}
\details{
Based off a exercise in the seqinr vignette.
}
\examples{
\dontrun{
choosebank("genbank")
query_taxon("test",
            c("phytophthora", "pythium"),
            c("@18S@", "@28S@"),
            "RRNA")
str(test)
closebank()}
}
\seealso{
\code{\link[taxize]{ncbi_getbyname}} \code{\link[seqinr]{query}}
}

