% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_taxa--dataset_documentation.R
\docType{data}
\name{ranks_ref}
\alias{ranks_ref}
\title{Lookup-table for IDs of taxonomic ranks}
\description{
Composed of two columns:
\itemize{
 \item rankid - the ordered identifier value. lower values mean higher rank
 \item ranks - all the rank names that belong to the same level, with
 different variants that mean essentially the same thing
}
}
\keyword{data}
