% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{rarefy_obs}
\alias{rarefy_obs}
\title{Calculate proportions from observation counts}
\usage{
rarefy_obs(obj, dataset, cols = NULL, sample_size = NULL,
  other_cols = FALSE)
}
\arguments{
\item{obj}{A taxmap object}

\item{dataset}{The name of a table in \code{obj} that contains counts.}

\item{cols}{The names/indexes of columns in \code{data} that have counts. By 
Default, all numeric columns in \code{data} are used.}

\item{sample_size}{The sample size counts will be rarefied to. This can be 
either a single integer or a vector of integers of equal length to the 
number of columns.}

\item{other_cols}{If \code{TRUE}, keep non-count cols in the input data.
The "taxon_id" column will always be preserved.}
}
\value{
A tibble
}
\description{
For a given table in a taxmap object, rarefy counts to a constant total. This
is a wrapper around \code{\link[vegan]{rrarefy}} that automatically detects
which columns are numeric and handles the reformatting needed to use tibbles.
}
\seealso{
Other calculations: \code{\link{calc_n_samples}},
  \code{\link{calc_obs_props}},
  \code{\link{calc_taxon_abund}},
  \code{\link{compare_treatments}}
}
