% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fasta.R
\name{read_fasta}
\alias{read_fasta}
\title{Read a FASTA file}
\usage{
read_fasta(file_path)
}
\arguments{
\item{file_path}{(\code{character} of length 1) The path to a file to read.}
}
\value{
names \code{character}
}
\description{
Reads a FASTA file. This is the FASTA parser for metacoder. It simply tries
to read a FASTA into a named character vector with minimal fuss. It does not
do any checks for valid characters etc.
}
