% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fasta.R
\name{read_lines_apply}
\alias{read_lines_apply}
\title{Apply a function to chunks of a file}
\usage{
read_lines_apply(
  file_path,
  func,
  buffer_size = 1000,
  simplify = FALSE,
  skip = 0
)
}
\arguments{
\item{file_path}{(\code{character} of length 1) The path to a file to read.}

\item{func}{(\code{function}) The function to run on each chunk of the file.}

\item{buffer_size}{(\code{numeric} of length 1) The number of lines in each chunk}

\item{simplify}{(\code{logical} of length 1) If \code{TRUE}, then the result is simplified to a vector.}

\item{skip}{(\code{numeric} of length 1) Where to start reading the file.}
}
\value{
\code{list} of results of \code{func}
}
\description{
Reads a file in chunks, applies a function to each of them, and returns to results of the function calls.
}
\keyword{internal}
