% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classified--internal.R
\name{remove_redundant_names}
\alias{remove_redundant_names}
\title{Remove the redundant taxon names}
\usage{
remove_redundant_names(obj, name_col)
}
\arguments{
\item{obj}{a \code{classified} object}

\item{name_col}{(\code{character} of length 1)
The name of a column in \code{obj$taxon_data}}
}
\value{
\code{character}
}
\description{
Remove the names of parent taxa in the begining of their children's names in a \code{classified} object.
This is useful for remove genus names in species binomials.
}

