% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_taxonomy--internal.R
\name{rename_duplicated}
\alias{rename_duplicated}
\title{Number duplicated names}
\usage{
rename_duplicated(my_names)
}
\arguments{
\item{my_names}{(\code{character})}
}
\value{
\code{character}
}
\description{
Add a number to any character values that appear more that once.
}
\keyword{internal}

