% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/internal.R
\name{rescale_limits}
\alias{rescale_limits}
\title{Scale numbers to a range of values}
\usage{
rescale_limits(input, new_min, new_max)
}
\arguments{
\item{input}{(\code{numeric}) Numbers to rescale.}

\item{new_min}{(\code{numeric} of length 1) The minimum of the rescaled numbers.}

\item{new_max}{(\code{numeric} of length 1) The maximum of the rescaled numbers.}
}
\description{
Scales a set of numbers to a differnt range based on their range
}
\examples{
rescale_limits(1:10, -2.2, 7.5)
hist(rescale_limits(rnorm(1000), 5, 10))
}

