% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa--classified_class.R
\name{reserved_col_names}
\alias{reserved_col_names}
\title{invalid column names for user data}
\usage{
reserved_col_names()
}
\description{
invalid column names for user data
}
\keyword{internal}

