% Generated by roxygen2 (4.0.2): do not edit by hand
\name{resizingTextGrob}
\alias{resizingTextGrob}
\title{Adds text grob that scales with viewport size}
\usage{
resizingTextGrob(...)
}
\description{
Creates a new grob class called resizingTextGrob that is supposed to resize automatically
It is a thin layer over \code{grid::textGrob}.
}
\details{
Taken from: http://ryouready.wordpress.com/2012/08/01/creating-a-text-grob-that-automatically-adjusts-to-viewport-size/
}

