% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{rev_comp}
\alias{rev_comp}
\title{Revere complement sequences}
\usage{
rev_comp(seqs)
}
\arguments{
\item{seqs}{A character vector with one element per sequence.}
}
\description{
Make the reverse complement of one or more sequences stored as a character
vector. This is a wrapper for \code{\link[seqinr]{comp}} for character
vectors instead of lists of character vectors with one value per letter.
IUPAC ambiguity codes are handled and the upper/lower case is preserved.
}
\examples{

rev_comp(c("aagtgGGTGaa", "AAGTGGT"))

}
\seealso{
Other sequence transformations: 
\code{\link{complement}()},
\code{\link{reverse}()}
}
\concept{sequence transformations}
