% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{reverse}
\alias{reverse}
\title{Reverse sequences}
\usage{
reverse(seqs)
}
\arguments{
\item{seqs}{A character vector with one element per sequence.}
}
\description{
Find the reverse of one or more sequences stored as a character
vector. This is a wrapper for \code{\link{rev}} for character
vectors instead of lists of character vectors with one value per letter.
}
\examples{

reverse(c("aagtgGGTGaa", "AAGTGGT"))

}
\seealso{
Other sequence transformations: 
\code{\link{complement}()},
\code{\link{rev_comp}()}
}
\concept{sequence transformations}
