% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa--classified_class.R
\name{roots}
\alias{roots}
\title{Get root taxa}
\usage{
roots(obj, subset = taxon_ids(obj))
}
\arguments{
\item{obj}{(\code{classified}) The \code{classified} object containing taxon information to be queried.}

\item{subset}{(\code{character})
Taxon IDs for which supertaxa will be returned.
Default: All taxon in \code{obj} will be used.}
}
\value{
\code{character}
}
\description{
Return the root taxa for a \code{\link{classified}} object.
Can also be used to get the roots of a subset of taxa.
}

