% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxmap--sample.R
\name{sample_frac_items}
\alias{sample_frac_items}
\title{Sample a proportion of items from \code{\link{taxmap}}}
\usage{
sample_frac_items(.data, size = 1, replace = FALSE, taxon_weight = NULL,
  item_weight = NULL, use_supertaxa = TRUE, collapse_func = mean, ...)
}
\arguments{
\item{.data}{(\code{\link{taxmap}}) The object to sample from.}

\item{size}{(\code{numeric} of length 1) The proportion of items to sample.}

\item{replace}{(\code{logical} of length 1) If \code{TRUE}, sample with replacement.}

\item{taxon_weight}{(\code{numeric}) Non-negative sampling weights of each taxon. If 
\code{use_supertaxa} is \code{TRUE}, the weights for each taxon in an item's classification are
supplied to \code{collapse_func} to get the item weight. The expression given is evaluated in 
the context of \code{\link{taxon_data}}. In other words, any column name that appears in 
\code{\link{taxon_data}(.data)} can be used as if it was a vector on its own. If 
\code{item_weight} is also specified, the two weights are multiplied (after \code{taxon_weight}
for each item is calculated).}

\item{item_weight}{(\code{numeric}) Sampling weights of each item. The expression given is 
evaluated in the context of \code{\link{item_data}}. In other words, any column name that 
appears in \code{\link{item_data}(.data)} can be used as if it was a vector on its own. If 
\code{taxon_weight} is also specified, the two weights are multiplied (after 
\code{taxon_weight} for each item is calculated).}

\item{use_supertaxa}{(\code{logical} of length 1) Affects how the \code{taxon_weight} is used. If
\code{TRUE}, the weights for each taxon in an item's classification are multiplied to get the 
item weight. Otherwise, just the taxonomic level the item is assign to it considered.}

\item{collapse_func}{(\code{function} of length 1) If \code{taxon_weight} option is used and 
\code{supertaxa} is \code{TRUE}, the weights for each taxon in an item's classification are 
supplied to \code{collapse_func} to get the item weight. This function should take  numeric 
vector and return a single number.}

\item{...}{Additional options are passed to \code{\link{filter_items}}.}
}
\value{
An object of type \code{\link{taxmap}}
}
\description{
Randomly sample some propoortion of items from a \code{\link{taxmap}} object. Weights can be 
specified for items or the taxa they are taxmap by.
See \link[dplyr]{sample_frac} for the inspiration for this function.
}
\seealso{
Other dplyr-like functions: \code{\link{arrange_items}},
  \code{\link{arrange_taxa}}, \code{\link{filter_items}},
  \code{\link{filter_taxa}}, \code{\link{mutate_items}},
  \code{\link{mutate_taxa}},
  \code{\link{sample_frac_taxa}},
  \code{\link{sample_n_items}},
  \code{\link{sample_n_taxa}}, \code{\link{select_items}},
  \code{\link{select_taxa}}, \code{\link{transmute_items}},
  \code{\link{transmute_taxa}}
}

