% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classified--sample.R
\name{sample_n_items}
\alias{sample_n_items}
\title{Sample n items from \code{\link{classified}}}
\usage{
sample_n_items(.data, size, replace = FALSE, taxon_weight = NULL,
  item_weight = NULL, supertaxa = TRUE)
}
\arguments{
\item{.data}{(\code{\link{classified}})}

\item{size}{(\code{numeric} of length 1) The number of items to sample.}

\item{replace}{(\code{logical} of length 1) If \code{TRUE}, sample with replacement.}

\item{supertaxa}{(\code{logical} of length 1) Affects how the \code{taxon_weight} is used. If
\code{TRUE}, the weights for each taxon in an item's classification are multiplied to get the
item weight. Otherwise, just the taxonomic level the item is assign to it considered.}

\item{itemless}{(\code{logical} of length 1) If \code{TRUE}, preserve taxa even if all of their
items are filtered out. If \code{FALSE}, remove taxa for which all items were filtered out.
Note that only taxa that are itemless due to this filtering will be removed; there might be
other taxa without items to begin with that will not be removed.}
}
\value{
An object of type \code{\link{classified}}
}
\description{
Randomly sample some number of items from a \code{\link{classified}} object. Weighs can be
specified for items or the taxa they are classified by.
}
\details{
See \link[dplyr]{sample_n} for the inspiration for this function.
}
\seealso{
Other dplyr-like functions: \code{\link{arrange_items}},
  \code{\link{arrange_taxa}}, \code{\link{filter_items}},
  \code{\link{filter_taxa}}, \code{\link{mutate_items}},
  \code{\link{mutate_taxa}}, \code{\link{select_items}},
  \code{\link{select_taxa}}, \code{\link{transmute_items}},
  \code{\link{transmute_taxa}}
}

