% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxmap--sample.R
\name{sample_n_obs}
\alias{sample_n_obs}
\title{Sample n observations from \code{\link{taxmap}}}
\usage{
sample_n_obs(.data, size, replace = FALSE, taxon_weight = NULL,
  obs_weight = NULL, use_supertaxa = TRUE, collapse_func = mean, ...)
}
\arguments{
\item{.data}{(\code{\link{taxmap}}) The object to sample from.}

\item{size}{(\code{numeric} of length 1) The number of observations to sample.}

\item{replace}{(\code{logical} of length 1) If \code{TRUE}, sample with replacement.}

\item{taxon_weight}{(\code{numeric}) Non-negative sampling weights of each taxon. If 
\code{use_supertaxa} is \code{TRUE}, the weights for each taxon in an observation's classification are
supplied to \code{collapse_func} to get the observation weight. The expression given is evaluated in 
the context of \code{\link{taxon_data}}. In other words, any column name that appears in 
\code{\link{taxon_data}(.data)} can be used as if it was a vector on its own. If 
\code{obs_weight} is also specified, the two weights are multiplied (after \code{taxon_weight}
for each observation is calculated).}

\item{obs_weight}{(\code{numeric}) Sampling weights of each observation. The expression given is 
evaluated in the context of \code{\link{obs_data}}. In other words, any column name that 
appears in \code{\link{obs_data}(.data)} can be used as if it was a vector on its own. If 
\code{taxon_weight} is also specified, the two weights are multiplied (after 
\code{taxon_weight} for each observation is calculated).}

\item{use_supertaxa}{(\code{logical} of length 1) Affects how the \code{taxon_weight} is used. If
\code{TRUE}, the weights for each taxon in an observation's classification are multiplied to get the 
observation weight. Otherwise, just the taxonomic level the observation is assign to it considered.}

\item{collapse_func}{(\code{function} of length 1) If \code{taxon_weight} option is used and 
\code{supertaxa} is \code{TRUE}, the weights for each taxon in an observation's classification are 
supplied to \code{collapse_func} to get the observation weight. This function should take  numeric 
vector and return a single number.}

\item{...}{Additional options are passed to \code{\link{filter_obs}}.}
}
\value{
An object of type \code{\link{taxmap}}
}
\description{
Randomly sample some number of observations from a \code{\link{taxmap}} object. Weights can be 
specified for observations or the taxa they are taxmap by.
See \link[dplyr]{sample_n} for the inspiration for this function.
}
\examples{
# Subsample without replacement, keeping all taxa
sample_n_obs(unite_ex_data_3, 100)
# Subsample without replacement and remove unsampled taxa
sample_n_obs(unite_ex_data_3, 100, unobserved = FALSE)
# Subsample with taxon weight 
sample_n_obs(unite_ex_data_3, 100, unobserved = FALSE, taxon_weight = 1 / n_obs)
# Sample with replacement
sample_n_obs(unite_ex_data_3, 10000, replace = TRUE)
  
}
\seealso{
Other dplyr-like functions: \code{\link{arrange_obs}},
  \code{\link{arrange_taxa}}, \code{\link{filter_obs}},
  \code{\link{filter_taxa}}, \code{\link{mutate_obs}},
  \code{\link{mutate_taxa}}, \code{\link{sample_frac_obs}},
  \code{\link{sample_frac_taxa}},
  \code{\link{sample_n_taxa}}, \code{\link{select_obs}},
  \code{\link{select_taxa}}, \code{\link{transmute_obs}},
  \code{\link{transmute_taxa}}
}
