% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classified--select.R
\name{select_items}
\alias{select_items}
\title{Subset columns in a \code{\link{classified}} object}
\usage{
select_items(.data, ...)
}
\arguments{
\item{.data}{\code{\link{classified}}}

\item{...}{One or more column names to return in the new object.
This can be one of three things:
\describe{
   \item{\code{character}}{The name of a column in \code{item_data}}
   \item{\code{unquoted column name}}{The name of a column in \code{item_data} typed as if it was a varaible on its own.}
   \item{\code{logical}}{A \code{TRUE}/\code{FALSE} vector of length equal to the number of columns in \code{item_data}}
   \item{\code{numeric}}{Indexes of columns in \code{item_data}}
 }}
}
\value{
An object of type \code{\link{classified}}
}
\description{
Subsets \code{item_data} columns in a \code{\link{classified}} object.
Takes and returns a \code{\link{classified}} object.
Any column name that appears in \code{item_data(.data)} can be used as if it was a vector on its own.
}

