% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heat_tree--internal.R
\name{select_labels}
\alias{select_labels}
\title{Pick labels to show}
\usage{
select_labels(my_data, label_max, sort_by_column, label_column)
}
\arguments{
\item{my_data}{\code{data.frame}}

\item{label_max}{\code{numeric} of length 1}

\item{sort_by_column}{\code{character} of length 1; the name of a column in \code{my_data}}

\item{label_column}{\code{character} of length 1; the name of a column in \code{my_data}
containing labels}
}
\value{
\code{character} IDs of rows with labels to show
}
\description{
Pick labels to show based off a column name to sort by and a maximum number
}
\keyword{internal}
