% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_documentation.R
\docType{data}
\name{silva_ex_data}
\alias{silva_ex_data}
\title{Example dataset from SILVA}
\format{An object of type \code{\link{taxmap}}}
\usage{
silva_ex_data
}
\description{
https://www.arb-silva.de/
}
\examples{
\dontrun{

file_path <- system.file("extdata", "silva_nr99.fasta", package = "metacoder")
sequences <- ape::read.FASTA(file_path)
silva_ex_data <- extract_taxonomy(sequences,
                                  regex = "^(.*?) (.*)$",
                                  key = c(id = "obs_info", "class"),
                                  class_sep = ";")
}

}
\keyword{datasets}
