% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_taxonomy--internal.R
\name{split_class_list}
\alias{split_class_list}
\title{Split a list of classifications by a row/column value}
\usage{
split_class_list(classifications, row_index, col_index)
}
\arguments{
\item{classifications}{(\code{list} of \code{data.frame}) 
The classifications of a set of observations.
Not necessarily unique.
Rows should correspond to taxa and columns to information associated with those taxa.}

\item{row_index}{The row in each \code{data.frame}}

\item{col_index}{The column in each \code{data.frame}}
}
\value{
\code{list} of \code{list} of \code{data.frame}
}
\description{
Split a list of classifications into a list of list by unique values of a
specific row/column.
}
\keyword{internal}
