% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_taxon_class.R
\name{[[.classified}
\alias{[[.classified}
\title{Create a restrictive subset of \code{\link{classified}}}
\usage{
\method{[[}{classified}(...)
}
\value{
\code{\link{classified}}
}
\description{
Create a subset of items classified by a taxonomy.
Child and parent taxa of specified taxa are not preserved.
}

