% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{subsample_by_taxonomy}
\alias{subsample_by_taxonomy}
\title{subsample_by_taxonomy}
\usage{
subsample_by_taxonomy(distance_matrix, taxon, taxon_level, level_order,
  triangular = TRUE, level_to_analyze = "subtaxon", max_subset = NA)
}
\description{
subsample_by_taxonomy
}

