% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/test_taxon_class.R
\name{subset_classified}
\alias{subset_classified}
\title{Create a inclusive subset of \code{\link{classified}}}
\usage{
subset_classified(obj, taxon, item, subtaxa = TRUE, supertaxa = FALSE)
}
\arguments{
\item{taxon}{A key to filter the taxon data.frame rows on}

\item{item}{A key to filter the item data.frame rows on}
}
\value{
\code{\link{classified}}
}
\description{
Create a subset of items classified by a taxonomy.
Only unspecified taxa with no items or children with items are discarded.
}

