% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa--classified_class.R
\name{subtaxa}
\alias{subtaxa}
\title{Get all subtaxa of a taxon}
\usage{
subtaxa(obj, subset = taxon_ids(obj), recursive = TRUE, simplify = FALSE,
  include_input = FALSE)
}
\arguments{
\item{obj}{(\code{classified})
The \code{classified} object containing taxon information to be queried.}

\item{subset}{(\code{character})
Taxon IDs for which subtaxa will be returned.}

\item{recursive}{(\code{logical})
If \code{FALSE}, only return the subtaxa one level above the
  target taxa. If \code{TRUE}, return all the subtaxa of every subtaxa, etc.}

\item{simplify}{(\code{logical})
If \code{TRUE}, then combine all the results into a single
  vector of unique taxon IDs}

\item{include_input}{(\code{logical})
If \code{TRUE}, the input taxa are included in the output}
}
\value{
If \code{simplify = FALSE}, then a list of vectors of taxon IDs are returned
  corresponding to the \code{target} argument. If \code{simplify = TRUE}, then the unique taxon
  IDs for all \code{target} taxa are returned in a single vector.
}
\description{
Return the taxon IDs of all subtaxa in an object of type \code{classified}
}

