% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa--classified_class.R
\name{sum.classified}
\alias{sum.classified}
\title{Combine classified data}
\usage{
\method{sum}{classified}(...)
}
\arguments{
\item{...}{(\code{\link{classified}}) One or more objects of type \code{\link{classified}}}
}
\value{
An object of type \code{\link{classified}}
}
\description{
Combine the contents of multiple objects of type \code{\link{classified}}
}
\keyword{internal}

