% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_taxon_class.R
\name{taxon_apply}
\alias{taxon_apply}
\title{Apply a function to every taxon}
\usage{
taxon_apply(obj, func)
}
\arguments{
\item{obj}{(\code{\link{classified}})}

\item{func}{(\code{function}) A function that accepts an object of type \code{\link{classified}}.}
}
\value{
\code{list} of length equal to the number of taxa in \code{obj}
}
\description{
Apply  a function to every taxon in an object of type \code{\link{classified}}.
}

