% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_taxon_class.R
\name{taxon_data}
\alias{taxon_data}
\title{Return taxon data from \code{\link{classified}}}
\usage{
taxon_data(obj, calculated_cols = TRUE)
}
\arguments{
\item{obj}{(\code{\link{classified}})}

\item{calculated_cols}{(\code{logical} of length 1) If \code{TRUE}, return calculated columns using
\code{\link{classified}$taxon_funcs}.}
}
\value{
\code{data.frame} with rows corresponding to taxa in input
}
\description{
Return a table of data associated with taxa of and object of type
\code{\link{classified}}.
}

