% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa--classified_class.R
\name{taxon_data_colnames}
\alias{taxon_data_colnames}
\title{Return taxon data column names}
\usage{
taxon_data_colnames(obj)
}
\arguments{
\item{obj}{(\code{\link{classified}})}
}
\description{
Return taxon data column names of and object of type \code{\link{classified}}.
This includes "taxon_id", "parent_id", user-defined columns, and columns generated by \code{taxon_funcs}.
}

