% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classified--internal.R
\name{taxon_data_cols_used}
\alias{taxon_data_cols_used}
\title{Get names of taxon_data in an unevaluated expression}
\usage{
taxon_data_cols_used(obj, ...)
}
\arguments{
\item{obj}{a \code{classified} object}

\item{...}{unevaluated expression}
}
\value{
\code{character}
}
\description{
Get names of taxon_data in an unevaluated expression
}
\keyword{internal}

