% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa--classified_class.R
\name{taxon_ids}
\alias{taxon_ids}
\title{Get taxon IDs}
\usage{
taxon_ids(obj)
}
\arguments{
\item{obj}{(\code{classified}) The \code{classified} object containing taxon information to be queried.}
}
\value{
\code{character}
}
\description{
Return the taxon IDs for a \code{\link{classified}} object.
}
\keyword{internal}

