% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barcode_gap_estimation.R
\name{taxon_output_path_preparation}
\alias{taxon_output_path_preparation}
\title{taxon_output_path_preparation}
\usage{
taxon_output_path_preparation(output_directory, sub_directory = NULL,
  name = NULL, id = NULL, level_name = NULL, ext = "", ...)
}
\description{
taxon_output_path_preparation
}

