% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa--classified_class.R
\name{taxon_ranks}
\alias{taxon_ranks}
\title{Get rank of taxa}
\usage{
taxon_ranks(obj, subset = taxon_ids(obj))
}
\arguments{
\item{obj}{(\code{\link{classified}})}

\item{subset}{(\code{character}) The \code{taxon_ids}s to get ranks for.}
}
\value{
\code{numeric}
}
\description{
Get rank of taxa in an object of type \code{\link{classified}}
}

