% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{taxonomic_sample}
\alias{taxonomic_sample}
\title{Recursivly sample a set of taxonomic assignments}
\usage{
taxonomic_sample(classified_data, max_counts = c(), min_counts = c(),
  max_children = c(), min_children = c(), item_filters = list(),
  subtaxa_filters = list(), stop_conditions = list(), ...)
}
\arguments{
\item{classified_data}{(An object of type \link{classified})}

\item{max_counts}{(\code{numeric}) A named vector that defines that maximum number of
items in for each level specified. The names of the vector specifies that level each number
applies to. If more than the maximum number of items exist for a given taxon, it is randomly
subsampled to this number.}

\item{min_counts}{(\code{numeric}) A named vector that defines that minimum number of
items in for each level specified. The names of the vector specifies that level each number
applies to.}

\item{max_children}{(\code{numeric}) A named vector that defines that maximum number of
subtaxa per taxon for each level specified. The names of the vector specifies that level each
number applies to. If more than the maximum number of subtaxa exist for a given taxon, they
are randomly subsampled to this number of subtaxa.}

\item{min_children}{(\code{numeric}) A named vector that defines that minimum number of
subtaxa in for each level specified. The names of the vector specifies that level each number
applies to.}

\item{item_filters}{(\code{list} of \code{function(items, id)}) A list of functions that take a data
structure containing the information of multiple items and a taxon id.
Returns a object of the same type with some of the items potentially removed.}

\item{subtaxa_filters}{(\code{list} of \code{function(items, id)}) A list of functions that take a data
structure containing the information of multiple subtaxa IDs and the current taxon id.
Returns a object of the same type with some of the subtaxa potentially removed. If a function returns
\code{NULL}, then no items for the current taxon are returned.}

\item{stop_conditions}{(\code{list} of \code{function(id)}) A list of functions that take the
current taxon id. If any of the functions return \code{TRUE}, the items for the current taxon are 
returned rather than looking for items of subtaxa, stopping the recursion.}

\item{...}{Additional parameters are passed to all of the function options.}
}
\value{
Returns an object of type \code{classified}
}
\description{
Recursivly sample a set of items with taxonomic assignments and an associated taxonomy.
}

