% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{taxonomy_to_adj_list}
\alias{taxonomy_to_adj_list}
\title{Make an adjacency list from classifications}
\usage{
taxonomy_to_adj_list(classifications)
}
\arguments{
\item{classifications}{(\code{list} of \code{data.frame}) Taxnomic classifications used to build
adjacency list. Each classification must be a \code{data.frame} with a column named "taxon".}
}
\description{
Makes an adjacency list from a list of taxonomic classifications meant to represent the tree 
structure shared by the classifications given.
}

