% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_taxonomy--resizing_text.R
\name{text_grob_length}
\alias{text_grob_length}
\title{Estimate text grob length}
\usage{
text_grob_length(text)
}
\arguments{
\item{text}{\code{character} The text to be printed}
}
\value{
The estimated length of the printed text as a multiple of its height
}
\description{
Estimate the printed length of `resizingTextGrob` text
}
\keyword{internal}

