% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barcode_gap_estimation.R
\name{threshold_optimization}
\alias{threshold_optimization}
\title{Optimize clustering threshold}
\usage{
threshold_optimization(distance, threshold_resolution = 0.001,
  output_file_path = NULL, plot_file_path = NULL, name = NULL, ...)
}
\description{
Optimize clustering threshold
}

