% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_taxa--internal.R
\name{to_percent}
\alias{to_percent}
\title{Format a proportion as a printed percent}
\usage{
to_percent(prop, digits = 3, ...)
}
\arguments{
\item{prop}{The proportion}

\item{digits}{a positive integer indicating how many significant digits
    are to be used for
    numeric and complex \code{x}.  The default, \code{NULL}, uses
    \code{\link[base]{getOption}("digits")}.  This is a suggestion: enough decimal
    places will be used so that the smallest (in magnitude) number has
    this many significant digits, and also to satisfy \code{nsmall}.
    (For more, notably the interpretation for complex numbers see
    \code{\link[base]{signif}}.)}

\item{...}{passed to `format`}
}
\value{
character
}
\description{
Format a proportion as a printed percent
}
\keyword{internal}
