% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classified--mutate.R
\name{transmute_items}
\alias{transmute_items}
\title{Replace columns in \code{\link{classified}} objects}
\usage{
transmute_items(.data, ...)
}
\arguments{
\item{.data}{\code{\link{classified}}}

\item{...}{One or more column names to add to the new object.
Newly created columns can be referenced in the same function call.}
}
\value{
An object of type \code{\link{classified}}
}
\description{
Replace columns of \code{item_data} in \code{\link{classified}} objects.
Any column name that appears in \code{item_data(.data)} can be used as if it was a vector on its own.
See \code{\link[dplyr]{transmute}} for inspiration and more information.
}

