% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R, R/old_taxa--internal.R
\name{unique_mapping}
\alias{unique_mapping}
\title{get indexes of a unique set of the input}
\usage{
unique_mapping(input)

unique_mapping(input)
}
\description{
get indexes of a unique set of the input

Get indexes of a unique set of the input
}
\keyword{internal}
