% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{unique_taxa}
\alias{unique_taxa}
\title{Indentify unique taxa in classifications}
\usage{
unique_taxa(classifications, id_column = NULL)
}
\arguments{
\item{classifications}{(\code{list} of \code{data.frame})
Each classification must be a \code{data.frame} with a column named "taxon".}

\item{id_column}{(\code{character}) The column name in each classification \code{data.frame}
that will be used to name the elements of the returned list.}
}
\value{
(\code{list} of \code{data.frame}) Taxonomic classifications of every unique taxon in 
the list of classifications given.
}
\description{
Indentifies unique taxa in a list of taxonomic classifications.
}
\keyword{internal}

