% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_documentation.R
\docType{data}
\name{unite_ex_data}
\alias{unite_ex_data}
\title{Example of UNITE fungal ITS data}
\format{An object of type \code{\link{classified}}}
\source{
\url{https://unite.ut.ee/}
}
\usage{
unite_ex_data
}
\description{
A dataset containing information from 500 sequences from the UNITE reference database.
}
\examples{
\dontrun{
file_path <- system.file("extdata", "unite_general_release.fasta", package = "metacoder")
sequences <- ape::read.FASTA(file_path)
unite_ex_data <- extract_taxonomy(names(sequences),
                                  regex = "^(.*)\\\\|(.*)\\\\|(.*)\\\\|.*\\\\|(.*)$",
                                  key = c(name = "item_info", seq_id = "item_info",
                                          other_id = "item_info", "class_name"),
                                  database = "none")
}

}
\keyword{datasets}

