% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{validate_edge_list}
\alias{validate_edge_list}
\title{Validate a taxon edge list}
\usage{
validate_edge_list(taxa, parents)
}
\arguments{
\item{taxa}{(\code{character}) Unique taxon IDs for every possible taxon.}

\item{parents}{(\code{character}) Unique taxon IDs for the supertaxa of every possible taxon. The
lack of a parent should be coded as \code{NA}.}
}
\value{
(\code{logical}) Returns \code{TRUE} if the list is valid.
}
\description{
Validate a taxon edge list for use in other functions
}
\keyword{internal}

