% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_taxonomy--internal.R
\name{validate_regex_match}
\alias{validate_regex_match}
\title{Check that all match input}
\usage{
validate_regex_match(input, regex, max_print = 10)
}
\arguments{
\item{input}{(\code{character})}

\item{regex}{(\code{character} of length 1)}

\item{max_print}{(\code{numeric} of length 1)
The maximum number of lines to print in error/warning messages.}
}
\value{
\code{character} Parts of \code{input} matching \code{regex}
}
\description{
Ensure that all of a character vector matches a regex.
Inputs that do not match are excluded.
}
\keyword{internal}
