% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_taxa--taxmap--internal.R
\name{validate_taxmap_funcs}
\alias{validate_taxmap_funcs}
\title{Validate `funcs` input for Taxamp}
\usage{
validate_taxmap_funcs(funcs)
}
\arguments{
\item{funcs}{The `funcs` variable passed to the `Taxmap` constructor}
}
\value{
A `funcs` variable with the right format
}
\description{
Make sure `funcs` is in the right format and complain if it is not.
NOTE: This currently does nothing.
}
\keyword{internal}
