% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heat_tree--internal.R
\name{verify_color_range}
\alias{verify_color_range}
\title{Verify color range parameters}
\usage{
verify_color_range(args)
}
\arguments{
\item{args}{(\code{character}) The names of arguments to verify.}
}
\description{
Verify color range parameters
}
\keyword{internal}
