% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heat_tree--internal.R
\name{verify_label_count}
\alias{verify_label_count}
\title{Verify label count}
\usage{
verify_label_count(args)
}
\arguments{
\item{args}{(\code{character}) The names of arguments to verify.}
}
\description{
Verify label count
}
\keyword{internal}
