% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_taxonomy--internal.R
\name{verify_size_range}
\alias{verify_size_range}
\title{Verify size range parameters}
\usage{
verify_size_range(args)
}
\arguments{
\item{args}{(\code{character}) The names of arguments to verify.}
}
\description{
Verify size range parameters
}
\keyword{internal}

