% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/option_parsers.R
\name{verify_taxmap}
\alias{verify_taxmap}
\title{Check that an object is a taxmap}
\usage{
verify_taxmap(obj)
}
\arguments{
\item{obj}{A taxmap object}
}
\description{
Check that an object is a taxmap
This is intended to be used to parse options in other functions.
}
\seealso{
Other option parsers: 
\code{\link{get_taxmap_cols}()},
\code{\link{get_taxmap_data}()},
\code{\link{get_taxmap_other_cols}()},
\code{\link{get_taxmap_table}()}
}
\concept{option parsers}
\keyword{internal}
