% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heat_tree--internal.R
\name{verify_trans}
\alias{verify_trans}
\title{Verify transformation function parameters}
\usage{
verify_trans(args)
}
\arguments{
\item{args}{(\code{character}) The names of arguments to verify.}
}
\description{
Verify transformation function parameters
}
\keyword{internal}
