% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_taxonomy--internal.R
\name{vigilant_report}
\alias{vigilant_report}
\title{Report a error/warning if needed}
\usage{
vigilant_report(text, vigilance = c("error", "none", "message", "warning"))
}
\arguments{
\item{text}{The error to report}

\item{vigilance}{(\code{character} of length 1) Controls the reporting of possible problems.
The following values are possible: 
 \describe{
   \item{\code{"none"}}{No warnings or errors are generated if the function can complete.}
   \item{\code{"message"}}{A message is generated when atypical events occur.}
   \item{\code{"warning"}}{Warnings are generated when atypical events occur.}
   \item{\code{"error"}}{Errors are generated when atypical events occur, stopping the completion of the function.}
 }}
}
\value{
\code{NULL}
}
\description{
Report a error/warning if needed
}
\keyword{internal}

