% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writers.R
\name{write_greengenes}
\alias{write_greengenes}
\title{Write an immitation of the Greengenes databse}
\usage{
write_greengenes(obj, tax_file = NULL, seq_file = NULL,
  tax_names = taxon_names, ranks = gg_rank, ids = gg_id,
  sequences = gg_seq)
}
\arguments{
\item{obj}{A taxmap object}

\item{tax_file}{(\code{character} of length 1) The file path to save the
taxonomy file.}

\item{seq_file}{(\code{character} of length 1) The file path to save the
sequence fasta file. This is optional.}

\item{tax_names}{(\code{character} named by taxon ids) The names of taxa}

\item{ranks}{(\code{character} named by taxon ids) The ranks of taxa}

\item{ids}{(\code{character} named by taxon ids) Sequence ids}

\item{sequences}{(\code{character} named by taxon ids) Sequences}
}
\value{
\code{\link{taxmap}}
}
\description{
Attempts to save taxonomic and sequence information of a taxmap object in the
Greengenes output format. If the taxmap object was created using
[parse_greengenes()], then it should be able to replicate the format
exactly.
}
\details{
The taxonomy output file has a format like:

\preformatted{
228054  k__Bacteria; p__Cyanobacteria; c__Synechococcophycideae; o__Synech...
844608  k__Bacteria; p__Cyanobacteria; c__Synechococcophycideae; o__Synech...
...
}

The optional sequence file has a format like:

\preformatted{
>1111886 
AACGAACGCTGGCGGCATGCCTAACACATGCAAGTCGAACGAGACCTTCGGGTCTAGTGGCGCACGGGTGCGTA...
>1111885 
AGAGTTTGATCCTGGCTCAGAATGAACGCTGGCGGCGTGCCTAACACATGCAAGTCGTACGAGAAATCCCGAGC...
...
}
}
\seealso{
Other writers: \code{\link{write_mothur_taxonomy}},
  \code{\link{write_rdp}}
}
