% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writers.R
\name{write_mothur_taxonomy}
\alias{write_mothur_taxonomy}
\title{Write an immitation of the Mothur taxonomy file}
\usage{
write_mothur_taxonomy(obj, file, tax_names = taxon_names, ids = sequence_id,
  scores = score)
}
\arguments{
\item{obj}{A taxmap object}

\item{file}{(\code{character} of length 1) The file path to save the
sequence fasta file. This is optional.}

\item{tax_names}{(\code{character} named by taxon ids) The names of taxa}

\item{ids}{(\code{character} named by taxon ids) Sequence ids}

\item{ranks}{(\code{character} named by taxon ids) The ranks of taxa}

\item{info}{(\code{character} named by taxon ids) Info associated with
sequences. In the example output shown here, this field corresponds to
"Sparassis crispa; MBUH-PIRJO&ILKKA94-1587/ss5"}

\item{sequences}{(\code{character} named by taxon ids) Sequences}
}
\value{
\code{\link{taxmap}}
}
\description{
Attempts to save taxonomic information of a taxmap object in the
mothur `*.taxonomy` format. If the taxmap object was created using
[parse_mothur_taxonomy()], then it should be able to replicate the format
exactly.
}
\details{
The output file has a format like:

\preformatted{
AY457915	Bacteria(100);Firmicutes(99);Clostridiales(99);Johnsone...
AY457914	Bacteria(100);Firmicutes(100);Clostridiales(100);Johnso...
AY457913	Bacteria(100);Firmicutes(100);Clostridiales(100);Johnso...
AY457912	Bacteria(100);Firmicutes(99);Clostridiales(99);Johnsone...
AY457911	Bacteria(100);Firmicutes(99);Clostridiales(98);Ruminoco...
}

or...

\preformatted{
AY457915	Bacteria;Firmicutes;Clostridiales;Johnsonella_et_rel.;J...
AY457914	Bacteria;Firmicutes;Clostridiales;Johnsonella_et_rel.;J...
AY457913	Bacteria;Firmicutes;Clostridiales;Johnsonella_et_rel.;J...
AY457912	Bacteria;Firmicutes;Clostridiales;Johnsonella_et_rel.;J...
AY457911	Bacteria;Firmicutes;Clostridiales;Ruminococcus_et_rel.;...
}
}
\seealso{
Other writers: \code{\link{write_greengenes}},
  \code{\link{write_rdp}}, \code{\link{write_silva_fasta}},
  \code{\link{write_unite_general}}
}
