% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writers.R
\name{write_rdp}
\alias{write_rdp}
\title{Write an immitation of the RDP FASTA databse}
\usage{
write_rdp(obj, file, tax_names = taxon_names, ranks = rdp_rank,
  ids = rdp_id, info = seq_name, sequences = rdp_seq)
}
\arguments{
\item{obj}{A taxmap object}

\item{file}{(\code{character} of length 1) The file path to save the
sequence fasta file. This is optional.}

\item{tax_names}{(\code{character} named by taxon ids) The names of taxa}

\item{ranks}{(\code{character} named by taxon ids) The ranks of taxa}

\item{ids}{(\code{character} named by taxon ids) Sequence ids}

\item{info}{(\code{character} named by taxon ids) Info associated with
sequences. In the example output shown here, this field corresponds to
"Sparassis crispa; MBUH-PIRJO&ILKKA94-1587/ss5"}

\item{sequences}{(\code{character} named by taxon ids) Sequences}
}
\value{
\code{\link{taxmap}}
}
\description{
Attempts to save taxonomic and sequence information of a taxmap object in the
RDP FASTA format. If the taxmap object was created using
[parse_rdp()], then it should be able to replicate the format
exactly.
}
\details{
The output file has a format like:

\preformatted{
>S000448483 Sparassis crispa; MBUH-PIRJO&ILKKA94-1587/ss5	Lineage=Root;rootrank;Fun...
ggattcccctagtaactgcgagtgaagcgggaagagctcaaatttaaaatctggcggcgtcctcgtcgtccgagttgtaa
tctggagaagcgacatccgcgctggaccgtgtacaagtctcttggaaaagagcgtcgtagagggtgacaatcccgtcttt
...
}
}
\seealso{
Other writers: \code{\link{write_greengenes}}
}
