% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writers.R
\name{write_silva_fasta}
\alias{write_silva_fasta}
\title{Write an immitation of the SILVA FASTA databse}
\usage{
write_silva_fasta(obj, file, tax_names = taxon_names,
  other_names = other_name, ids = ncbi_id, start = start_pos,
  end = end_pos, sequences = silva_seq)
}
\arguments{
\item{obj}{A taxmap object}

\item{file}{(\code{character} of length 1) The file path to save the
sequence fasta file. This is optional.}

\item{tax_names}{(\code{character} named by taxon ids) The names of taxa}

\item{other_names}{(\code{character} named by taxon ids) Alternate names
of taxa. Will be added after the primary name.}

\item{ids}{(\code{character} named by taxon ids) Sequence ids}

\item{start}{(\code{character}) The start position of the
sequence.}

\item{end}{(\code{character}) The end position of the
sequence.}

\item{sequences}{(\code{character} named by taxon ids) Sequences}
}
\value{
\code{\link{taxmap}}
}
\description{
Attempts to save taxonomic and sequence information of a taxmap object in the
SILVA FASTA format. If the taxmap object was created using
[parse_silva_fasta()], then it should be able to replicate the format
exactly.
}
\details{
The output file has a format like:

\preformatted{
>GCVF01000431.1.2369 Bacteria;Proteobacteria;Gammaproteobacteria;Oceanospiril...
CGUGCACGGUGGAUGCCUUGGCAGCCAGAGGCGAUGAAGGACGUUGUAGCCUGCGAUAAGCUCCGGUUAGGUGGCAAACA
ACCGUUUGACCCGGAGAUCUCCGAAUGGGGCAACCCACCCGUUGUAAGGCGGGUAUCACCGACUGAAUCCAUAGGUCGGU
...
}
}
\seealso{
Other writers: \code{\link{write_greengenes}},
  \code{\link{write_mothur_taxonomy}},
  \code{\link{write_rdp}},
  \code{\link{write_unite_general}}
}
