% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writers.R
\name{write_unite_general}
\alias{write_unite_general}
\title{Write an immitation of the UNITE general FASTA databse}
\usage{
write_unite_general(obj, file, tax_names = taxon_names, ranks = unite_rank,
  ids = unite_id, gb_acc = acc_num, sequences = unite_seq,
  seq_name = organism)
}
\arguments{
\item{obj}{A taxmap object}

\item{file}{(\code{character} of length 1) The file path to save the
sequence fasta file. This is optional.}

\item{tax_names}{(\code{character} named by taxon ids) The names of taxa}

\item{ranks}{(\code{character} named by taxon ids) The ranks of taxa}

\item{ids}{(\code{character} named by taxon ids) UNITE sequence ids}

\item{sequences}{(\code{character} named by taxon ids) Sequences}

\item{seq_name}{(\code{character} named by taxon ids) Name of sequences.
Usually a taxon name.}

\item{acc_num}{(\code{character} named by taxon ids) Genbank accession
numbers}
}
\value{
\code{\link{taxmap}}
}
\description{
Attempts to save taxonomic and sequence information of a taxmap object in the
UNITE general FASTA format. If the taxmap object was created using
[parse_unite_general()], then it should be able to replicate the format
exactly.
}
\details{
The output file has a format like:

\preformatted{
>Glomeromycota_sp|KJ484724|SH523877.07FU|reps|k__Fungi;p__Glomeromycota;c__unid...
ATAATTTGCCGAACCTAGCGTTAGCGCGAGGTTCTGCGATCAACACTTATATTTAAAACCCAACTCTTAAATTTTGTAT...
...
}
}
\seealso{
Other writers: \code{\link{write_greengenes}},
  \code{\link{write_mothur_taxonomy}},
  \code{\link{write_rdp}}
}
