% Generated by roxygen2 (4.0.2): do not edit by hand
\name{add_taxon_ids}
\alias{add_taxon_ids}
\title{Add ids for classifications}
\usage{
add_taxon_ids(classifications, id_key = NULL, id_col_name = "id")
}
\arguments{
\item{classifications}{(\code{list} of \code{data.frame}) Taxnomic classifications.}

\item{id_key}{(\code{list} of named \code{data.frame}) A list defining what classification every
unique id is. If not provided, unique ids will be generated from \code{classifications}.}

\item{id_col_name}{(\code{character}) The name of the column of unique ids that will be added
to each classification.}
}
\description{
Add ids for classifications
}
\seealso{
unique_taxa
}

