% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{add_taxon_ids}
\alias{add_taxon_ids}
\title{Add IDs for classifications}
\usage{
add_taxon_ids(classifications, id_key = NULL, id_col_name = "taxon_id")
}
\arguments{
\item{classifications}{(\code{list} of \code{data.frame}) Taxnomic classifications.}

\item{id_key}{(\code{list} of named \code{data.frame}) A list defining what classification every
unique id is. If not provided, unique IDs will be generated from \code{classifications}.}

\item{id_col_name}{(\code{character}) The name of the column of unique IDs that will be added
to each classification.}
}
\description{
Add IDs for classifications
}
\seealso{
unique_taxa
}

