% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxmap--arrange.R
\name{arrange_taxa}
\alias{arrange_taxa}
\title{Sort columns of \code{\link{taxmap}} objects}
\usage{
arrange_taxa(.data, ...)
}
\arguments{
\item{.data}{\code{\link{taxmap}}}

\item{...}{One or more column names to sort on. Newly created columns can be referenced in the
same function call.}
}
\value{
An object of type \code{\link{taxmap}}
}
\description{
Sort columns of \code{taxon_data} in \code{\link{taxmap}} objects. Any column name that
appears in \code{taxon_data(.data)} can be used as if it was a vector on its own. See
\link[dplyr]{arrange} for more details.
}
\seealso{
Other dplyr-like functions: \code{\link{arrange_items}},
  \code{\link{filter_items}}, \code{\link{filter_taxa}},
  \code{\link{mutate_items}}, \code{\link{mutate_taxa}},
  \code{\link{sample_frac_items}},
  \code{\link{sample_frac_taxa}},
  \code{\link{sample_n_items}},
  \code{\link{sample_n_taxa}}, \code{\link{select_items}},
  \code{\link{select_taxa}}, \code{\link{transmute_items}},
  \code{\link{transmute_taxa}}
}

