% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classified--column_generators.R
\name{assigned}
\alias{assigned}
\title{Count item assigned in \code{\link{classified}}}
\usage{
assigned(obj, subset = obj$taxon_data$taxon_ids)
}
\arguments{
\item{obj}{(\code{\link{classified}})}

\item{subset}{(\code{character}) The \code{taxon_ids}s to get counts for.}
}
\value{
\code{numeric}
}
\description{
Count items assigned to a specific taxon in \code{\link{classified}}
}

