% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{calc_obs_props}
\alias{calc_obs_props}
\title{Calculate proportions from observation counts}
\usage{
calc_obs_props(obj, dataset, cols = NULL, other_cols = FALSE,
  new_names = NULL)
}
\arguments{
\item{obj}{A taxmap object}

\item{dataset}{The name of a table in \code{obj} that contains counts.}

\item{cols}{The names/indexes of columns in \code{dataset} to use. By
default, all numeric columns are used. Takes one of the following inputs:
\describe{
  \item{TRUE/FALSE:}{All/No columns will used.}
  \item{Vector of TRUE/FALSE of length equal to the number of columns:}{Use the columns
corresponding to \code{TRUE} values.}
  \item{Character vector:}{The names of columns to use}
  \item{Numeric vector:}{The indexes of columns to use}
}}

\item{other_cols}{Preserve in the output non-target columns present in the
input data. The "taxon_id" column will always be preserved. Takes one of
the following inputs:
\describe{
  \item{TRUE/FALSE:}{All non-target columns will be preserved or not.}
  \item{Vector of TRUE/FALSE of length equal to the number of columns:}{Preserve the columns
corresponding to \code{TRUE} values.}
  \item{Character vector:}{The names of columns to preserve}
  \item{Numeric vector:}{The indexes of columns to preserve}
}}

\item{new_names}{If supplied, rename the output proportion columns. Must be
the same length as \code{cold}.}
}
\value{
A tibble
}
\description{
For a given table in a taxmap object, convert one or more columns containing
counts to proportions. This is meant to be used with counts associated with
observations (e.g. OTUs), as opposed to counts that have already been summed
per taxon. If counts have already been summed per taxon, use
[calc_tax_props()] instead.
}
\examples{
\dontrun{
# Parse dataset for examples
x = parse_tax_data(hmp_otus, class_cols = "lineage", class_sep = ";",
                   class_key = c(tax_rank = "info", tax_name = "taxon_name"),
                   class_regex = "^(.+)__(.+)$")
                   
# Calculate proportions for all numeric columns
calc_obs_props(x, "tax_data")

# Calculate proportions for a subset of columns
calc_obs_props(x, "tax_data", cols = c("700035653", "700097433", "700100489"))
calc_obs_props(x, "tax_data", cols = 4:6)
calc_obs_props(x, "tax_data", cols = startsWith(colnames(x$data$tax_data), "7"))

# Including all other columns in ouput
calc_obs_props(x, "tax_data", other_cols = TRUE)

# Inlcuding specific columns in output
calc_obs_props(x, "tax_data", cols = c("700035653", "700097433", "700100489"),
               other_cols = 2:3)


}
}
\seealso{
Other calculations: \code{\link{calc_n_samples}},
  \code{\link{calc_taxon_abund}},
  \code{\link{compare_treatments}},
  \code{\link{rarefy_obs}}
}
