% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{calc_obs_props}
\alias{calc_obs_props}
\title{Calculate proportions from observation counts}
\usage{
calc_obs_props(obj, dataset, cols = NULL, other_cols = FALSE)
}
\arguments{
\item{obj}{A taxmap object}

\item{dataset}{The name of a table in \code{obj} that contains counts.}

\item{cols}{The names/indexes of columns in \code{data} that have counts. By Default,
all numeric columns in \code{data} are used.}

\item{other_cols}{If \code{TRUE}, keep non-count cols in the input data.
The "taxon_id" column will always be preserved.}
}
\value{
A tibble
}
\description{
For a given table in a taxmap object, convert one or more columns containing
counts to proportions. This is meant to be used with counts associated with
observations (e.g. OTUs), as opposed to counts that have already been summed
per taxon. If counts have already been summed per taxon, use
[calc_tax_props()] instead.
}
\examples{
\dontrun{
# Parse dataset for plotting
x = parse_tax_data(hmp_otus, class_cols = "lineage", class_sep = ";",
                   class_key = c(tax_rank = "info", tax_name = "taxon_name"),
                   class_regex = "^(.+)__(.+)$")
                   
# Calculate proportions for all numeric columns
calc_obs_props(x, "tax_data")

# Calculate proportions for a subset of columns
calc_obs_props(x, "tax_data", cols = c("700097433", "700100489", "700111174"))

# Including all other columns in ouput
calc_obs_props(x, "tax_data", other_cols = TRUE)

# Inlcuding specific columns in output



}
}
\seealso{
Other calculations: \code{\link{calc_n_samples}},
  \code{\link{calc_taxon_abund}},
  \code{\link{compare_treatments}},
  \code{\link{rarefy_obs}}
}
