% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/barcode_gap_estimation.R
\name{calculate_barcode_statistics}
\alias{calculate_barcode_statistics}
\title{Calculate barcode statistics}
\usage{
calculate_barcode_statistics(distance_matrix,
  taxonomy_levels = get_taxonomy_levels(), saved_output_path = getwd(),
  level_to_analyze = "Species", distance_type = "distance",
  max_sequences_to_compare = 500, return_raw_data = FALSE,
  save_statistics = FALSE, save_raw_data = FALSE, save_plots = FALSE,
  taxonomy_separator = ";", functions_to_apply = list("overall_statistics",
  "intertaxon_statistics", "intrataxon_statistics", "distance_distribution",
  "threshold_optimization"), ...)
}
\description{
Calculate barcode statistics
}

