% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new--heat_tree--validation.R
\name{check_color_range}
\alias{check_color_range}
\title{Verify color range parameters}
\usage{
check_color_range(ranges, args)
}
\arguments{
\item{ranges}{A list of range arguments to check}

\item{args}{A list of all arguments}
}
\value{
A named list of argument values, potentially modified
}
\description{
Verify color range parameters
}
\keyword{internal}
