% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa--classified_class.R
\name{classifications}
\alias{classifications}
\title{Get classification of taxa}
\usage{
classifications(obj, subset = taxon_ids(obj), sep = ";")
}
\arguments{
\item{obj}{(\code{\link{classified}})}

\item{subset}{(\code{character})
The \code{taxon_ids}s to get classifications for.}

\item{sep}{(\code{character} of length 1)
The character(s) to place between taxon IDs}
}
\value{
\code{character} of length equal to \code{subset}
}
\description{
Get classification strings of taxa in an object of type \code{\link{classified}}.
Each classification is constructed by concatenating the taxon ids of the given taxon and its supertaxa.
}

