% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa--classified_class.R
\name{classified}
\alias{classified}
\title{Create an instance of \code{classified}}
\usage{
classified(taxon_ids, parent_ids, item_taxon_ids, taxon_data = NULL,
  item_data = NULL, taxon_funcs = list(item_counts = item_counts,
  taxon_ranks = taxon_ranks, classifications = classifications),
  item_funcs = list())
}
\arguments{
\item{taxon_ids}{Unique taxon ids}

\item{parent_ids}{Taxon ids of parent taxa of \code{taxon_ids}}

\item{item_taxon_ids}{Taxon ids of items}

\item{taxon_data}{A \code{data.frame} with rows pretaining to \code{taxon_ids}}

\item{item_data}{A \code{data.frame} with rows pretaining to \code{item_taxon_ids}}

\item{taxon_funcs}{A function that accepts a subset of this object containing information for a single taxa.
A single value must be returned derived from that information.
These values will be acessible as if it were a column in \code{taxon_data}.}

\item{item_funcs}{A function that accepts a subset of this object containing information for a single item.
A single value must be returned derived from that information.
These values will be acessible as if it were a column in \code{item_data}.}
}
\value{
An object of type \code{classified}
}
\description{
Create an instance of \code{classified} containing items classified by a taxonomy
}

