% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_taxon_class.R
\name{classified}
\alias{classified}
\title{Create an instance of \code{classified}}
\usage{
classified(taxon_id, parent_id, item_taxon_id, taxon_data = NULL,
  item_data = NULL, taxon_funcs = list(), item_funcs = list())
}
\arguments{
\item{taxon_id}{Unique taxon ids}

\item{parent_id}{Parent taxa of \code{taxon_id}}

\item{item_taxon_id}{Taxon ids of items}

\item{taxon_data}{A \code{data.frame} with rows pretaining to \code{taxon_id}}

\item{item_data}{A \code{data.frame} with rows pretaining to \code{item_taxon_id}}

\item{taxon_funcs}{A function that accepts a subset of this object containing information for a single taxa.
A single value must be returned derived from that information. 
These values will be acessible as if it were a column in \code{taxon_data}.}

\item{item_funcs}{A function that accepts a subset of this object containing information for a single item.
A single value must be returned derived from that information. 
These values will be acessible as if it were a column in \code{item_data}.}
}
\description{
Create an instance of \code{classified} containing items classified by a taxonomy
}

