% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa--classified_class.R
\name{count_items}
\alias{count_items}
\title{Count items in \code{\link{classified}}}
\usage{
count_items(obj, taxon)
}
\arguments{
\item{obj}{(\code{\link{classified}})}

\item{taxon}{(\code{character} of length 1) The taxon_id used}
}
\value{
\code{numeric} of length 1
}
\description{
Count items in \code{\link{classified}}
}
\keyword{internal}

