% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{delete_vetices_and_children}
\alias{delete_vetices_and_children}
\title{delete_vetices_and_children}
\usage{
delete_vetices_and_children(graph, nodes)
}
\description{
delete_vetices_and_children
}
\keyword{internal}
