% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/barcode_gap_estimation.R
\name{distance_distribution}
\alias{distance_distribution}
\title{distance_distribution}
\usage{
distance_distribution(distance, identity = NULL, distance_bin_width = 0.001,
  output_file_path = NULL, plot_file_path = NULL, name = NULL, ...)
}
\description{
distance_distribution
}

