% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heat_tree--mapping.R
\name{diverging_palette}
\alias{diverging_palette}
\title{The default diverging color palette}
\usage{
diverging_palette()
}
\value{
\code{character} of hex color codes
}
\description{
Returns the default color palette for diverging data
}
\examples{
diverging_palette()

}
