% Generated by roxygen2 (4.0.2): do not edit by hand
\name{download_gb_query}
\alias{download_gb_query}
\title{Download the sequences from an seqinr query object and formats them with their annotations}
\usage{
download_gb_query(query_req)
}
\arguments{
\item{query_req}{A list of class  \code{\link[seqinr]{SeqAcnucWeb}} (e.g. what
\code{\link[seqinr]{query}} produces in the \code{x$req} element of the output.}
}
\value{
A data.frame in the format of the output of \code{\link[taxize]{ncbi_getbyid}}.
}
\description{
Formats the output of \code{\link[seqinr]{getSequence}} and \code{\link[seqinr]{getAnnot}}
to the output of \code{\link[taxize]{ncbi_getbyid}}.
}

