% Generated by roxygen2 (4.0.2): do not edit by hand
\name{drawDetails.resizingTextGrob}
\alias{drawDetails.resizingTextGrob}
\title{Draws a resizingTextGrob}
\usage{
\method{drawDetails}{resizingTextGrob}(x, recording = TRUE)
}
\description{
Called automatically when drawing a grob using grob.draw.
}
\details{
Taken from: http://ryouready.wordpress.com/2012/08/01/creating-a-text-grob-that-automatically-adjusts-to-viewport-size/
}

