% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/resizing_text_grob.R
\name{drawDetails.resizingTextGrob}
\alias{drawDetails.resizingTextGrob}
\title{Draws a resizingTextGrob}
\usage{
drawDetails.resizingTextGrob(x, recording = TRUE)
}
\description{
Called automatically when drawing a grob using grob.draw.
}
\details{
Taken from: http://ryouready.wordpress.com/2012/08/01/creating-a-text-grob-that-automatically-adjusts-to-viewport-size/
}

