% Generated by roxygen2 (4.0.2): do not edit by hand
\name{extract_organism}
\alias{extract_organism}
\title{Extract the binomial organism name from genbank annotations.}
\usage{
extract_organism(annotation)
}
\description{
Takes a genebank annotation as a character vector with one line per element and returns the
organism name
}

