% Generated by roxygen2 (4.0.2): do not edit by hand
\name{extract_taxonomy}
\alias{extract_taxonomy}
\title{Extract taxonomy information from sequence headers}
\usage{
extract_taxonomy(input, regex, key, lineage_tax_sep = ";",
  lineage_rank_sep = "__", lineage_tax_rev = FALSE,
  lineage_rank_rev = FALSE, taxon_in_lineage = TRUE, database = "ncbi")
}
\arguments{
\item{input}{(\code{character}) A vector from which to extract taxonomy information.}

\item{regex}{(\code{character; length == 1}) A regular expression with capturing groups
indicating the locations of relevant information. The identity of the information must
be specified using the \code{key} argument.}

\item{key}{(\code{character}) The identity of the caturing groups defined using \code{regex}.
The length of \code{key} must be equal to the number of capturing groups specified in \code{regex}.
Any names added to the terms will be used as column names in the output.
Each term must be one of those decribed below:
\describe{
  \item{\code{taxon_name}}{The name of a taxon. Not necessarily unique.}
  \item{\code{taxon_id}}{A unique numeric id for a taxon.}
  \item{\code{taxon_rank}}{A taxonomic rank name (e.g. "genus").}
  \item{\code{taxon_info}}{Arbitrary taxon info you want included in the output. Can be used more than once.}
  \item{\code{lineage}}{A list of taxa names that consitute the full taxonomic classification
from broad to specific (see \code{lineage_tax_rev}). Individual names are not necessarily unique.
Individual taxa are separated by the \code{lineage_tax_sep} argument and the taxon-rank group is separated
by the \code{lineage_rank_sep} argument.}
  \item{\code{lineage_id}}{A list of taxa unique ids that consitute the full taxonomic
classification from broad to specific. Same usage as \code{lineage}}.
  \item{\code{item_name}}{An item (e.g. sequence) name. Not necessarily unique.}
  \item{\code{item_id}}{An unique item (e.g. sequence) identifier. The taxonomy information will be
looked up if available. Requires an internet connection.}
  \item{\code{item_info}}{Arbitrary item info you want included in the output. Can be used more than once.}
}}

\item{lineage_tax_sep}{(\code{character; length == 1}) Used with the \code{lineage} term in the \code{key}
argument. The characters used to separate individual taxa within a lineage.}

\item{lineage_rank_sep}{(\code{character; length == 1}) Used with the \code{lineage} term in the \code{key}
argument when a lineage contiains both taxon and rank information. This is the characters used to separate
th rank and the taxon name within an individual taxa in a lineage.}

\item{lineage_tax_rev}{Used with the \code{lineage} term in the \code{key} argument.
If TRUE, the rank order of taxa read in a lineage is reversed to be specific to broad.}

\item{lineage_rank_rev}{Used with the \code{lineage} term in the \code{key} argument  when a lineage
contiains both taxon and rank information. If TRUE, the rank information come after the taxon information.}

\item{taxon_in_lineage}{If \code{TRUE}, the lineage string included the taxon itself as its most specific
classification.}

\item{database}{(\code{character; length == 1}): The name of the database that patterns given in
\code{parser} will apply to. Currently, only \code{ncbi} is being supported.}
}
\value{
Returns a list of two elements:
 \describe{
   \item{\code{taxonomy}}{A data.frame with 2 columns representing A type of
   \href{http://en.wikipedia.org/wiki/Adjacency_list}{adjacency list}.
   The columns are the ids of a taxon and its subtaxon
   respectivly. If a taxon has more than subtaxon, it will be
   present multiple times in the first "parent taxon" column. However, a taxon can only be
   present once in the second "sub taxon" column; otherwise, it would imply the taxon has more than
   one "parent". This data structure can be converted to more intuitive ones with ... TO BE CONTINUED.}
   \item{\code{taxa}}{A data.frame with one row per taxon, containing available information
   on each taxon. The number and nature of columns depend on the input data. Typically, a column
   of taxon names is present.}
   \item{\code{items}}{A data.frame with one row per input item (typically sequences)
   containing their taxon ids and other informtion, depending on input.}
   }
}
\description{
Extracts the taxonomy used by a set of sequences based on their header information. A data
structure representing the heirerarchical nature of the taxonomy as well as a vector
identifing the taxon of each sequence is returned. Taxa are translated into unique codes if they
are not already encoded this way.
}

