% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{extract_taxonomy}
\alias{extract_taxonomy}
\title{Extract taxonomy information from sequence headers}
\usage{
extract_taxonomy(input, regex, key, class_tax_sep = ";",
  class_rank_sep = "__", class_tax_rev = FALSE, class_rank_rev = FALSE,
  taxon_in_lineage = TRUE, database = "ncbi", arbitrary_ids = "warn")
}
\arguments{
\item{input}{(\code{character}) A vector from which to extract taxonomy information.}

\item{regex}{(\code{character; length == 1}) A regular expression with capturing groups
indicating the locations of relevant information. The identity of the information must
be specified using the \code{key} argument.}

\item{key}{(\code{character}) The identity of the capturing groups defined using \code{regex}.
The length of \code{key} must be equal to the number of capturing groups specified in \code{regex}.
Any names added to the terms will be used as column names in the output.
Each term must be one of those decribed below:
\describe{
  \item{\code{taxon_id}}{A unique numeric id for a taxon for a particular \code{database}}
  \item{\code{taxon_name}}{The name of a taxon. Not necessarily unique, but are specific (i.e. interperable)
  to a particular \code{database}.}
  \item{\code{taxon_info}}{Arbitrary taxon info you want included in the output. Can be used more than once.}
  \item{\code{class_id}}{A list of taxa unique ids that consitute the full taxonomic classification
from broad to specific (see \code{class_tax_rev}) for a particular \code{database}. Individual taxa
are separated by the \code{class_tax_sep} argument and the taxon-rank group is separated by the
\code{class_rank_sep} argument.}
  \item{\code{class_name}}{A list of taxa names that consitute the full taxonomic
classification from broad to specific. Same usage as \code{class_id}}.
Individual names are not necessarily unique, but are specific (i.e. interperable)
to a particular \code{database}.
  \item{\code{item_id}}{An unique item (e.g. sequence) identifier. The taxonomy information will be
looked up if available. Requires an internet connection.}
  \item{\code{item_name}}{An item (e.g. sequence) name. Not necessarily unique.}
  \item{\code{item_info}}{Arbitrary item info you want included in the output. Can be used more than once.}
}}

\item{class_tax_sep}{(\code{character; length == 1}) Used with the \code{class_name} term in the \code{key}
argument. The characters used to separate individual taxa within a lineage.}

\item{class_rank_sep}{(\code{character; length == 1}) Used with the \code{class_name} term in the \code{key}
argument when a lineage contiains both taxon and rank information. This is the characters used to separate
th rank and the taxon name within an individual taxa in a lineage.}

\item{class_tax_rev}{Used with the \code{class_name} term in the \code{key} argument.
If TRUE, the rank order of taxa read in a lineage is reversed to be specific to broad.}

\item{class_rank_rev}{Used with the \code{class_name} term in the \code{key} argument  when a lineage
contiains both taxon and rank information. If TRUE, the rank information come after the taxon information.}

\item{taxon_in_lineage}{If \code{TRUE}, the lineage string included the taxon itself as its most specific
classification.}

\item{database}{(\code{character; length == 1}): The name of the database that patterns given in
\code{parser} will apply to. Valid databases include "ncbi", "itis", "eol", "col", "tropicos",
and "nbn".}

\item{arbitrary_ids}{(\code{character} of length 1) Determines how the generation of arbitrary ids is
handled. Possible options are:
\describe{
  \item{\code{"allow"}}{Arbitrary ids are automatically generated if needed. These can occur intermixed
  with offical database ids in the case of failed database lookups.}
  \item{\code{"warn"}}{Like \code{"allow"} but issue a warning when arbitrary ids are used.}
  \item{\code{"error"}}{Cause an error if arbitrary ids are needed.}
  \item{\code{"na"}}{Put \code{NA}s where arbitrary are needed.}
  \item{\code{"none"}}{Do not use a database to look up information.}
}}
}
\value{
Returns a list of two elements:
 \describe{
   \item{\code{taxonomy}}{A list of \code{data.frame}s containing the classification of each
   unique taxon. The order of the elements corresponds to the rows in the "taxa"
   \code{data.frame} described below.}
   \item{\code{taxa}}{A data.frame with one row per taxon, containing available information
   on each taxon. The number and nature of columns depend on the input data. Typically, a column
   of taxon names is present.}
   \item{\code{items}}{A data.frame with one row per input item (typically sequences)
   containing their taxon ids and other informtion, depending on input.}
   }
}
\description{
Extracts the taxonomy used by a set of sequences based on their header information. A data
structure representing the heirerarchical nature of the taxonomy as well as a vector
identifing the taxon of each sequence is returned. Taxa are translated into unique codes if they
are not already encoded this way.
}

