% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{fapply}
\alias{fapply}
\title{under development}
\usage{
fapply(iterable, functions, .preprocessor = {     function(x) x },
  .preprocessor_args = list(), .allow_complex = TRUE, ...)
}
\description{
under development
}

