% Generated by roxygen2 (4.0.2): do not edit by hand
\name{format_taxize}
\alias{format_taxize}
\title{Format taxize output to string}
\usage{
format_taxize(taxa)
}
\arguments{
\item{taxa}{Output from taxize::classification}
}
\value{
A character vecotr of taxonomy strings pasted together with delimiters.
}
\description{
Format taxize output to string
}

