% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxmap--internal.R
\name{format_taxon_subset}
\alias{format_taxon_subset}
\title{Format taxon subset value}
\usage{
format_taxon_subset(obj, index)
}
\arguments{
\item{obj}{(\code{taxmap})
The \code{taxmap} object containing taxon information to be queried.}

\item{index}{If a \code{character}, then it should be values of \code{taxon_ids}.
If a \code{numeric}, then it should be row indexes of \code{taxon_data}.
If a \code{logical}, then it should correspond to rows of \code{taxon_data}.}
}
\value{
\code{numeric}
}
\description{
Format an input to a \code{subset} option on functions like \code{\link{supertaxa}}.
Converts logical and \code{taxon_ids} into indexes of \code{taxon_data}.
}
\keyword{internal}

