% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_documentation.R
\name{genbank_ex_data}
\alias{genbank_ex_data}
\title{Fungal ITS Genbank refseq}
\format{An object of type \code{\link{taxmap}}}
\source{
\url{http://www.ncbi.nlm.nih.gov/nuccore}
}
\description{
A dataset containing information for 299 sequences obtained from NCBI using the following query:
}
\details{
\code{
(18s[All Fields] AND 28s[All Fields]) AND "basidiomycetes"[porgn] AND (refseq[filter] AND ("700"[SLEN] : "800"[SLEN]))
}
}
\examples{
\dontrun{

file_path <- system.file("extdata", "ncbi_basidiomycetes.fasta", package = "metacoder")
sequences <- ape::read.FASTA(file_path)

# Note: This can take a few minutes
genbank_ex_data <- extract_tax_data(names(sequences),
                                    regex = "^.*\\\\|(.*)\\\\|.*\\\\|(.*)\\\\|(.*)$",
                                    key = c(gi_no = "info", "seq_id", desc = "info"),
                                    database = "ncbi")
                                    
}
}
\seealso{
Other example_fastas: \code{\link{bryophytes_ex_data}},
  \code{\link{contaminants}}, \code{\link{its1_ex_data}},
  \code{\link{pr2_ex_data}}, \code{\link{silva_ex_data}},
  \code{\link{unite_ex_data_1}},
  \code{\link{unite_ex_data_2}}
}
