% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{get_numerics}
\alias{get_numerics}
\title{Return numeric values in a character}
\usage{
get_numerics(input)
}
\arguments{
\item{input}{}
}
\description{
Returns just valid numeric values and ignores others.
}
\keyword{internal}
