% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heat_tree--vertex_size.R
\name{get_optimal_range}
\alias{get_optimal_range}
\title{Find optimal range}
\usage{
get_optimal_range(
  max_range,
  min_range,
  resolution,
  opt_crit,
  choose_best,
  minimize = TRUE
)
}
\arguments{
\item{max_range}{(\code{numeric} of length 2) The min and max boundaries to the search space for
the optimal maximum value.}

\item{min_range}{(\code{numeric} of length 2) The min and max boundaries to the search space for
the optimal minimum value.}

\item{resolution}{(\code{numeric} of length 2) The number of increments in each dimension.}

\item{opt_crit}{(\code{function}) A function that takes two arguments, the max and min, and
returns the optimality statistic.}

\item{choose_best}{(\code{function}) A function that takes a list of \code{opt_crit} outputs
and returns the index of the best one.}
}
\description{
Finds optimal max and min value using an optimality criterion.
}
\keyword{internal}
