% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/option_parsers.R
\name{get_taxmap_table}
\alias{get_taxmap_table}
\title{Get a table from a taxmap object}
\usage{
get_taxmap_table(obj, data)
}
\arguments{
\item{obj}{A taxmap object}

\item{data}{Which data set to use. Can be any of the following:
\describe{
  \item{Name}{The name of the data set to use.}
  \item{Index}{The index of the data set to use.}
  \item{TRUE/FALSE vector}{A TRUE/FALSE vector the same length as the
  number of datasets, with exactly one TRUE corresponding to the
  selected data set.}
}}
}
\value{
A table
}
\description{
Get a table from a taxmap object and complain if it does not exist.
This is intended to be used to parse options in other functions.
}
\seealso{
Other option parsers: 
\code{\link{get_taxmap_cols}()},
\code{\link{get_taxmap_data}()},
\code{\link{get_taxmap_other_cols}()},
\code{\link{verify_taxmap}()}
}
\concept{option parsers}
\keyword{internal}
