% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_taxonomy_levels}
\alias{get_taxonomy_levels}
\title{Get taxonomy levels using taxize}
\usage{
get_taxonomy_levels()
}
\value{
A character vector of taxonomy levels, such as "Subkingdom" and "Order", in order of the
  hiearchy.
}
\description{
Get taxonomy levels using taxize
}

