% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_download.R
\name{get_taxonomy_levels}
\alias{get_taxonomy_levels}
\title{Get taxonomy levels}
\usage{
get_taxonomy_levels()
}
\description{
Return An ordered factor of taxonomy levels, such as "Subkingdom" and "Order", in order of the
  hierarchy.
}
\keyword{internal}
