% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new--heat_tree--internal.R,
%   R/new--heat_tree--validation.R
\name{heat_tree_validate_arguments}
\alias{heat_tree_validate_arguments}
\title{Validate arguments given to heat_tree}
\usage{
heat_tree_validate_arguments(obj, args)

heat_tree_validate_arguments(obj, args)
}
\arguments{
\item{obj}{taxmap object}

\item{args}{A named list of argument values}
}
\value{
A named list of argument values, potentially modified
}
\description{
Validate arguments given to \code{\link{heat_tree}} or
\code{\link{heat_tree_data}} and issue warnings or errors if things dont make
sense. NOTE: This function uses variables defined in \code{\link{heat_tree_data}}.

Validate arguments given to \code{\link{heat_tree}} or
\code{\link{heat_tree_data}} and issue warnings or errors if things dont make
sense.
}
\keyword{internal}
