% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_taxa--hierarchies.R
\name{hierarchies}
\alias{hierarchies}
\title{Make a set of many [hierarchy()] class objects}
\usage{
hierarchies(..., .list = NULL)
}
\arguments{
\item{...}{Any number of object of class [hierarchy()]}

\item{.list}{Any number of object of class [hierarchy()] in a list}
}
\value{
An `R6Class` object of class [hierarchy()]
}
\description{
NOTE: This will soon be depreciated.
Make a set of many [hierarchy()] class objects.
This is just a thin wrapper over a standard list.
}
\seealso{
Other classes: 
\code{\link{hierarchy}()},
\code{\link{taxa}()},
\code{\link{taxmap}()},
\code{\link{taxon}()},
\code{\link{taxon_database}()},
\code{\link{taxon_id}()},
\code{\link{taxon_name}()},
\code{\link{taxon_rank}()},
\code{\link{taxonomy}()}
}
\concept{classes}
