% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxmap--column_generators.R
\name{hierarchies}
\alias{hierarchies}
\title{Get classification of taxa}
\usage{
hierarchies(obj, subset = 1:nrow(obj$taxon_data), sep = ";")
}
\arguments{
\item{obj}{(\code{\link{taxmap}})}

\item{subset}{(\code{character})
The \code{taxon_ids}s or indexes in \code{taxon_data} to get classifications for.}

\item{sep}{(\code{character} of length 1)
The character(s) to place between taxon IDs}
}
\value{
\code{character} of length equal to \code{subset}
}
\description{
Get classification strings of taxa in an object of type \code{\link{taxmap}}.
Each classification is constructed by concatenating the taxon ids of the given taxon and its supertaxa.
}
\seealso{
Other taxon_funcs: \code{\link{n_obs_1}},
  \code{\link{n_obs}}, \code{\link{n_subtaxa_1}},
  \code{\link{n_subtaxa}}, \code{\link{n_supertaxa}}
}

