% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heat_tree--legend.R
\name{inverse}
\alias{inverse}
\title{Generate the inverse of a function}
\usage{
inverse(f, interval)
}
\arguments{
\item{f}{(\code{function} with one argument) A function to derive and inverse from}

\item{interval}{(\code{character} of length 2) The range of the value the inverse function can return.}
}
\value{
(\code{function}) Return the inverse of the function given
}
\description{
http://stackoverflow.com/questions/10081479/solving-for-the-inverse-of-a-function-in-r
}
\keyword{internal}
