% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_taxon_class.R
\name{item_apply}
\alias{item_apply}
\title{(NOT IMPLEMENTED) Apply a function to every item}
\usage{
item_apply(obj, func)
}
\arguments{
\item{obj}{(\code{\link{classified}})}

\item{func}{(\code{function}) A function that accepts an object of type \code{\link{classified}}.}
}
\value{
\code{list} of length equal to the number of items in \code{obj}
}
\description{
Apply  a function to every item in an object of type \code{\link{classified}}.
}

