% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa--classified_class.R
\name{item_counts}
\alias{item_counts}
\title{Count items in \code{\link{classified}}}
\usage{
item_counts(obj, subset = taxon_ids(obj))
}
\arguments{
\item{obj}{(\code{\link{classified}})}

\item{subset}{(\code{character}) The \code{taxon_ids}s to get counts for.}
}
\value{
\code{numeric}
}
\description{
Count items in \code{\link{classified}}
}

