% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa--classified_class.R
\name{item_data}
\alias{item_data}
\alias{item_data.classified}
\title{Return item data from \code{\link{classified}}}
\usage{
item_data(...)

\method{item_data}{classified}(obj, subset = item_data_colnames(obj),
  drop = FALSE, ...)
}
\arguments{
\item{...}{Passed to \code{taxon_data.classified}}

\item{obj}{(\code{\link{classified}})}

\item{subset}{(\code{character}) The names of columns, either user defined or generated using \code{item_funcs}.
Default: All columns.}

\item{drop}{(\code{logical} of length 1) If \code{TRUE}, if \code{subset} is a single column
name, then a \code{vector} is returned instead of a \code{data.frame}}
}
\value{
\code{data.frame} with rows corresponding to items in input
}
\description{
Return a table of data associated with items of an object of type
\code{\link{classified}}.
}

