% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxmap--internal.R
\name{obs_data_colnames}
\alias{obs_data_colnames}
\title{Get column names of obs_data}
\usage{
obs_data_colnames(obj)
}
\arguments{
\item{obj}{a \code{taxmap} object}
}
\value{
\code{character}
}
\description{
Get column names of obs_data without calculating columns
}

