% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa--classified_class.R
\name{item_data_colnames}
\alias{item_data_colnames}
\title{Return item data column names}
\usage{
item_data_colnames(obj)
}
\arguments{
\item{obj}{(\code{\link{classified}})}
}
\description{
Return item data column names of and object of type \code{\link{classified}}.
This includes "taxon_id", user-defined columns, and columns generated by \code{item_funcs}.
}

