% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa--classified_class.R
\name{items}
\alias{items}
\title{Get items associated with taxa}
\usage{
items(obj, subset = taxon_ids(obj), recursive = TRUE, simplify = FALSE)
}
\arguments{
\item{obj}{(\code{classified})
The \code{classified} object containing taxon information to be queried.}

\item{subset}{(\code{character})
Taxon IDs for which items will be returned.}

\item{recursive}{(\code{logical})
If \code{FALSE}, only return the item assigned to the specified input taxa, not subtaxa.
If \code{TRUE}, return all the items of every subtaxa, etc.}

\item{simplify}{(\code{logical}) If \code{TRUE}, then combine all the results into a single
vector of unique item indexes.}
}
\value{
If \code{simplify = FALSE}, then a list of vectors of item indexes are returned
  corresponding to the \code{target} argument. If \code{simplify = TRUE}, then the item indexes
  for all \code{target} taxa are returned in a single vector.
}
\description{
Given one or more taxa IDs and a \code{\link{classified}} object, return the items
(e.g. sequence information) associated with each taxon.
}
\keyword{internal}

