% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_taxonomy--legend.R
\name{make_plot_legend}
\alias{make_plot_legend}
\title{Make color/size legend}
\usage{
make_plot_legend(x, y, length, width_range, width_stat_range, group_prefix,
  tick_size = 0.007, width_stat_trans = function(x) {     x },
  width_title = "Size", width_sig_fig = 3, color_range, color_stat_range,
  color_stat_trans = function(x) {     x }, color_title = "Color",
  color_sig_fig = 3, divisions = 100, label_count = 8, title = NULL,
  label_size = 0.05, title_size = 0.05, axis_label_size = 0.05,
  color_axis_label = NULL, size_axis_label = NULL, hide_size = FALSE,
  hide_color = FALSE)
}
\arguments{
\item{x}{bottom left}

\item{y}{bottom left}

\item{length}{(\code{numeric} of length 1) the length of the scale bar}

\item{width_range}{(\code{numeric} of length 1 or 2) the width of the scale bar or the range}

\item{width_stat_range}{(\code{numeric} of length 1 or 2) The stat range to display in the size labels}

\item{group_prefix}{(\code{character} of length 1) The prefix of the group field in the shape data returned}

\item{tick_size}{(\code{numeric} of length 1) the thickness of tick marks}

\item{width_stat_trans}{(\code{function}) The transformation used to convert the statistic to size}

\item{width_title}{(\code{character} of length 1) The title of the size labels.}

\item{width_sig_fig}{(\code{numeric} of length 1) The number of significant figures to use in size labels.}

\item{color_range}{(\code{character}) One ore more hex codes constituting a color scale.}

\item{color_stat_range}{(\code{numeric} of length 1 or 2) The stat range to display in the color labels}

\item{color_stat_trans}{(\code{function}) The transformation used to convert the statistic to size}

\item{color_title}{(\code{character} of length 1) The title of the color labels.}

\item{color_sig_fig}{(\code{numeric} of length 1) The number of significant figures to use in color labels.}

\item{divisions}{(\code{numeric} of length 1) The number of colors to display.}

\item{label_count}{(\code{numeric} of length 1) The number of labels.}

\item{title}{(\code{character} of length 1) The title of the legend}

\item{axis_label_size}{(\code{numeric} of length 1)}

\item{color_axis_label}{(\code{character} of length 1) The label for the color axis}

\item{size_axis_label}{(\code{character} of length 1) The label for the size axis}

\item{hide_size}{(\code{logical} of length 1) If \code{TRUE} hide size axis}

\item{hide_color}{(\code{logical} of length 1) If \code{TRUE} hide color axis}
}
\description{
Make color/size legend
}
\keyword{internal}

