% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{map_unique}
\alias{map_unique}
\title{run a function on unique values of a iterable}
\usage{
map_unique(input, func, ...)
}
\description{
run a function on unique values of a iterable
}
\keyword{internal}

