% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R, R/old_taxa--internal.R
\name{map_unique}
\alias{map_unique}
\title{Run a function on unique values of a iterable}
\usage{
map_unique(input, func, ...)

map_unique(input, func, ...)
}
\arguments{
\item{input}{What to pass to \code{func}}

\item{func}{(\code{function})}

\item{...}{passed to \code{func}}
}
\description{
Runs a function on unique values of a list/vector and then reformats the
output so there is a one-to-one relationship with the input.
}
\keyword{internal}
