% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heat_tree--vertex_size.R
\name{molten_dist}
\alias{molten_dist}
\title{Get all distances between points}
\usage{
molten_dist(x, y)
}
\arguments{
\item{x}{(\code{numeric} of length 1) x coordinate}

\item{y}{(\code{numeric} of length 1) y coordinate}
}
\value{
A \code{data.frame}
}
\description{
Returns the distances between every possible combination of two points.
}
\keyword{internal}
