% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxmap--column_generators.R
\name{n_obs_1}
\alias{n_obs_1}
\title{Count observation assigned in \code{\link{taxmap}}}
\usage{
n_obs_1(obj, subset = 1:nrow(obj$taxon_data))
}
\arguments{
\item{obj}{(\code{\link{taxmap}})}

\item{subset}{(\code{character}) The \code{taxon_ids} or indexes in \code{taxon_data} to get counts for.}
}
\value{
\code{numeric}
}
\description{
Count observations assigned to a specific taxon in \code{\link{taxmap}}.
This does not include observations assigned to subtaxa.
}
\seealso{
Other taxon_funcs: \code{\link{hierarchies}},
  \code{\link{n_obs}}, \code{\link{n_subtaxa_1}},
  \code{\link{n_subtaxa}}, \code{\link{n_supertaxa}}
}

