% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new--heat_tree--shapes.R
\name{node_shape_circle}
\alias{node_shape_circle}
\title{Node shape function: circle}
\usage{
node_shape_circle(node_color = "grey", n_verticies = 30)
}
\description{
A function used to plot nodes in a heat tree. This function plots a simple circle in a solid color.
}
