% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pairwise_distance}
\alias{pairwise_distance}
\title{NOTE: dist.dna with a list of only 2 breaks R once in a while. DONT USE
Calculates a pairwise distance matrix from unaligned sequences}
\usage{
pairwise_distance(sequences)
}
\arguments{
\item{sequences}{A list or vector of sequence data}
}
\value{
A distance matrix of class \code{dist}
}
\description{
This is still in the developmental stage and currently only uses the kimura 2-parameter metric.
}

