% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mothur.R
\name{parse_mothur_summary}
\alias{parse_mothur_summary}
\title{Parse mothur classification summary file}
\usage{
parse_mothur_summary(file_path, untaxmap = FALSE)
}
\arguments{
\item{file_path}{(\code{character} of length 1)
The file path to the input file.}

\item{untaxmap}{(\code{logical} of length 1)
If \code{FALSE}, remove any untaxmap rows.}
}
\value{
\code{\link{taxmap}}
}
\description{
Parse mothur classification summary file
}

