% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{parse_phyloseq}
\alias{parse_phyloseq}
\title{Convert a phyloseq to taxmap}
\usage{
parse_phyloseq(obj)
}
\arguments{
\item{obj}{A phyloseq object}
}
\value{
A taxmap object
}
\description{
Converts a phyloseq object to a taxmap object.
}
\seealso{
Other parsers: \code{\link{parse_hmp_qiime}},
  \code{\link{parse_mothur_tax_summary}},
  \code{\link{parse_mothur_taxonomy}},
  \code{\link{parse_newick}}
}
