% Generated by roxygen2 (4.0.2): do not edit by hand
\name{parse_primersearch}
\alias{parse_primersearch}
\title{Parse EMBOSS primersearch output}
\usage{
parse_primersearch(file_path)
}
\arguments{
\item{file_path}{The path to a primersearch output file.}
}
\value{
A data frame with each row corresponding to amplicon data
}
\description{
Parses the output file from EMBOSS primersearch into a data.frame with rows corresponding to
predicted amplicons and their associated information.
}
\seealso{
\code{\link{run_primersearch}}
}

