% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{parse_silva_fasta}
\alias{parse_silva_fasta}
\title{Parse SILVA FASTA release}
\usage{
parse_silva_fasta(file, include_seqs = TRUE)
}
\arguments{
\item{file}{(\code{character} of length 1) The file path to the input file.}

\item{include_seqs}{(\code{logical} of length 1) If \code{TRUE}, include
sequences in the output object.}
}
\value{
\code{\link{taxmap}}
}
\description{
Parses an SILVA FASTA file that can be found at
https://www.arb-silva.de/no_cache/download/archive/release_128/Exports/.
}
\details{
The input file has a format like:

\preformatted{
>GCVF01000431.1.2369 Bacteria;Proteobacteria;Gammaproteobacteria;Oceanospiril...
CGUGCACGGUGGAUGCCUUGGCAGCCAGAGGCGAUGAAGGACGUUGUAGCCUGCGAUAAGCUCCGGUUAGGUGGCAAACA
ACCGUUUGACCCGGAGAUCUCCGAAUGGGGCAACCCACCCGUUGUAAGGCGGGUAUCACCGACUGAAUCCAUAGGUCGGU
...
}
}
\seealso{
Other parsers: \code{\link{parse_hmp_qiime}},
  \code{\link{parse_mothur_tax_summary}},
  \code{\link{parse_mothur_taxonomy}},
  \code{\link{parse_newick}}, \code{\link{parse_phyloseq}},
  \code{\link{parse_qiime_biom}}, \code{\link{parse_rdp}},
  \code{\link{parse_unite_general}}
}
