% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mothur.R
\name{parse_summary_seqs}
\alias{parse_summary_seqs}
\title{Parse summary.seqs output}
\usage{
parse_summary_seqs(text = NULL, file = NULL)
}
\arguments{
\item{text}{The text output of \code{summary.seqs}}

\item{file}{The path to saved output of \code{summary.seqs}}
}
\value{
A \code{data.frame} of statistics
}
\description{
Extract statistics from the command line output of mothur command \code{summary.seqs} and
return the results in a \code{data.frame}
}
\keyword{internal}
