% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_alignment}
\alias{plot_alignment}
\title{Display sequence alignment}
\usage{
plot_alignment(alignment)
}
\arguments{
\item{alignment}{(\code{DNAbin}) A matrix representing a sequence alignment.}
}
\description{
Make a plot pf a sequence alignment for an overview of alignment structure. 
NOT FINISHED.
}
\references{
ColorBrewer2 was used for the color palette
}

