% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_distance_distribution}
\alias{plot_distance_distribution}
\title{plot_distance_distribution}
\usage{
plot_distance_distribution(input, title = NULL, save_png = NULL,
  display = FALSE, background = "transparent", smoothness = 3,
  bin_width = NULL)
}
\description{
plot_distance_distribution
}

