% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_taxonomy}
\alias{plot_taxonomy}
\title{Plot items on taxonomy}
\usage{
plot_taxonomy(taxon_id, parent_id, vertex_size = NULL, vertex_color = NULL,
  vertex_alpha = NULL, line_size = NULL, line_color = NULL,
  line_alpha = NULL, label = NULL)
}
\arguments{
\item{taxon_id}{(\code{character} of length 1) The name of the column of \code{data} that
contains the unique id of the taxon for that row.}

\item{parent_id}{(\code{character} of length 1) The name of the column of \code{data} that
contains the unique id of supertaxon \code{taxon_id} is a part of.}

\item{vertex_size}{(\code{character} of length 1) The name of the column of \code{data} that
contains the value to base vertex size on.}

\item{vertex_color}{(\code{character} of length 1) The name of the column of \code{data} that
contains the value to base vertex size on.}

\item{vertex_alpha}{(\code{character} of length 1) The name of the column of \code{data} that
contains the value to base vertex size on.}

\item{line_size}{(\code{character} of length 1) The name of the column of \code{data} that
contains the value to base vertex size on.}

\item{line_color}{(\code{character} of length 1) The name of the column of \code{data} that
contains the value to base vertex size on.}

\item{line_alpha}{(\code{character} of length 1) The name of the column of \code{data} that
contains the value to base vertex size on.}

\item{data}{(\code{data.frame}) Must have columns with names specified by options \code{taxon_id}
and \code{parent_id}.}
}
\description{
Plots the distribution of values associated with item with an associated taxonomic classification.
Uses \code{igraph} to make layout and \code{ggplot2} to make plots.
}

