% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_taxonomy}
\alias{plot_taxonomy}
\title{Plot items on taxonomy}
\usage{
plot_taxonomy(taxon_id, parent_id, size = NULL, vertex_color = NULL,
  vertex_alpha = NULL, vertex_label = NULL, line_color = NULL,
  line_alpha = NULL, line_label = NULL, overlap_bias = 5,
  min_label_size = 0.015)
}
\arguments{
\item{taxon_id}{The unique ids of the taxon for each row.}

\item{parent_id}{The unique id of supertaxon \code{taxon_id} is a part of.}

\item{size}{The value to base vertex and line size on.}

\item{vertex_color}{The value to base vertex color on.}

\item{vertex_alpha}{The value to base vertex transparency on.}

\item{vertex_label}{The values of labels over vertcies.}

\item{line_color}{The value to base line color on.}

\item{line_alpha}{The value to base line transparency on.}

\item{line_label}{The values of labels over lines.}

\item{overlap_bias}{(\code{numeric} > 0) The factor by which overlaps are punished relative to
spaces when optimizing vertex size range.}

\item{min_label_size}{(\code{numeric} of length 1) The minimum label size that will be shown. A
proportion of viewport size. Labels that would be smaller are not added to save time.}

\item{data}{(\code{data.frame}) Must have columns with names specified by options \code{taxon_id}
and \code{parent_id}.}
}
\description{
Plots the distribution of values associated with item with an associated taxonomic classification.
Uses \code{igraph} to make layout and \code{ggplot2} to make plots.
}

