% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_taxonomy.R
\name{plot_taxonomy}
\alias{plot.classified}
\alias{plot_taxonomy}
\title{Plot a taxonomic tree}
\usage{
plot_taxonomy(taxon_id, parent_id, vertex_label = NA, edge_label = NA,
  tree_label = NA, vertex_size = 1, edge_size = vertex_size,
  tree_size = 1, vertex_label_size = vertex_size,
  edge_label_size = edge_size, tree_label_size = as.numeric(NA),
  vertex_color = "#999999", edge_color = vertex_color, tree_color = NA,
  vertex_label_color = "#000000", edge_label_color = "#000000",
  tree_label_color = "#000000", vertex_size_trans = "area",
  edge_size_trans = vertex_size_trans,
  vertex_label_size_trans = vertex_size_trans,
  edge_label_size_trans = edge_size_trans, tree_label_size_trans = "area",
  vertex_color_trans = "area", edge_color_trans = vertex_color_trans,
  tree_color_trans = "area", vertex_label_color_trans = "area",
  edge_label_color_trans = "area", tree_label_color_trans = "area",
  vertex_size_range = c(NA, NA), edge_size_range = c(NA, NA),
  vertex_label_size_range = c(NA, NA), edge_label_size_range = c(NA, NA),
  tree_label_size_range = c(NA, NA),
  vertex_color_range = quantative_palette(),
  edge_color_range = vertex_color_range,
  tree_color_range = quantative_palette(),
  vertex_label_color_range = quantative_palette(),
  edge_label_color_range = quantative_palette(),
  tree_label_color_range = quantative_palette(), vertex_label_max = 50,
  edge_label_max = 50, tree_label_max = 50, overlap_avoidance = 1,
  margin_size = c(0, 0), layout = "reingold-tilford",
  initial_layout = "fruchterman-reingold", make_legend = TRUE,
  title = NULL, title_size = 0.1, vertex_color_axis_label = NULL,
  vertex_size_axis_label = NULL, edge_color_axis_label = NULL,
  edge_size_axis_label = NULL, ...)

\method{plot}{classified}(classified_data, ...)
}
\arguments{
\item{taxon_id}{The unique ids of taxa.}

\item{parent_id}{The unique id of supertaxon \code{taxon_id} is a part of.}

\item{vertex_label}{See details on labels.
Default: no labels.}

\item{edge_label}{See details on labels.
Default: no labels.}

\item{tree_label}{See details on labels.
The label to display above each graph.
The value of the root of each graph will be used.
Default: None.}

\item{vertex_size}{See details on size.
Default: constant size.}

\item{edge_size}{See details on size.
Default: relative to vertex size.}

\item{vertex_label_size}{See details on size.
Default: relative to veterx size.}

\item{edge_label_size}{See details on size.
Default: relative to edge size.}

\item{tree_label_size}{See details on size.
Default: relative to graph size.}

\item{vertex_color}{See details on colors.
Default: grey.}

\item{edge_color}{See details on colors.
Default: same as vertex color.}

\item{tree_color}{See details on colors.
The value of the root of each graph will be used.
Overwrites the vertex and edge color if specified.
Default: Not used.}

\item{vertex_label_color}{See details on colors.
Default: black.}

\item{edge_label_color}{See details on colors.
Default: black.}

\item{tree_label_color}{See details on colors.
Default: black.}

\item{vertex_size_trans}{See details on transformations.
Default: \code{"area"}.}

\item{edge_size_trans}{See details on transformations. 
Default: same as \code{vertex_size_trans}.}

\item{vertex_label_size_trans}{See details on transformations. 
Default: same as \code{vertex_size_trans}.}

\item{edge_label_size_trans}{See details on transformations. 
Default: same as \code{edge_size_trans}.}

\item{tree_label_size_trans}{See details on transformations.
Default: \code{"area"}.}

\item{vertex_color_trans}{See details on transformations. 
Default: \code{"area"}.}

\item{edge_color_trans}{See details on transformations.
Default: same as vertex color transformation.}

\item{tree_color_trans}{See details on transformations.
Default: \code{"area"}.}

\item{vertex_label_color_trans}{See details on transformations.
Default: \code{"area"}.}

\item{edge_label_color_trans}{See details on transformations.
Default: \code{"area"}.}

\item{tree_label_color_trans}{See details on transformations. 
Default: \code{"area"}.}

\item{vertex_size_range}{See details on ranges.
Defualt: Optimize to balance overlaps and range size.}

\item{edge_size_range}{See details on ranges.
Default: relative to vertex size range.}

\item{vertex_label_size_range}{See details on ranges.
Default: relative to vertex size.}

\item{edge_label_size_range}{See details on ranges.
Default: relative to edge size.}

\item{tree_label_size_range}{See details on ranges.
Default: relative to tree size.}

\item{vertex_color_range}{See details on ranges.
Default: Color-blind friendly palette.}

\item{edge_color_range}{See details on ranges.
Default: same as vertex color.}

\item{tree_color_range}{See details on ranges.
Default: Color-blind friendly palette.}

\item{vertex_label_color_range}{See details on ranges.
Default: Color-blind friendly palette.}

\item{edge_label_color_range}{See details on ranges.
Default: Color-blind friendly palette.}

\item{tree_label_color_range}{See details on ranges.
Default: Color-blind friendly palette.}

\item{vertex_label_max}{The maximum number of vertex labels.
Default: 20.}

\item{edge_label_max}{The maximum number of edge labels.
Default: 20.}

\item{tree_label_max}{The maximum number of tree labels.
Default: 20.}

\item{overlap_avoidance}{(\code{numeric})
The relative importance of avoiding overlaps vs maximizing size range.
Higher numbers will cause vertex size optimazation to avoid overlaps more.
Default: \code{1}.}

\item{margin_size}{(\code{numeric} of length 2)
The horizontal and vertical margins.
Default: \code{0, 0}.}

\item{layout}{The layout algorithm used to position vertexes.
See details on layouts.
Default: \code{"reingold-tilford"}.}

\item{initial_layout}{he layout algorithm used to set the initial position
of vertexes, passed as input to the \code{layout} algorithm.
See details on layouts.
Default: Not used.}

\item{make_legend}{if TRUE...}

\item{title}{Name to print above the graph.}

\item{title_size}{The size of the title realtive to the rest of the graph.}

\item{vertex_color_axis_label}{The label on the scale axis corresponding to \code{vertex_color}.
Default: The expression given to \code{vertex_color}.}

\item{vertex_size_axis_label}{The label on the scale axis corresponding to \code{vertex_size}.
Default: The expression given to \code{vertex_size}.}

\item{edge_color_axis_label}{The label on the scale axis corresponding to \code{edge_color}.
Default: The expression given to \code{edge_color}.}

\item{edge_size_axis_label}{The label on the scale axis corresponding to \code{edge_size}.
Default: The expression given to \code{edge_size}.}

\item{...}{(other named arguments)
Passed to the \code{\link{igraph}} layout function used.}

\item{tree_size_trans}{See details on transformations.
Default: \code{"area"}.}
}
\description{
Plots the distribution of values associated with a taxonomic classification.
Taxonomic classifications can have multiple roots, resulting in multiple trees on the same plot.
Sizes and colors of vertexes, edges, labels, and individual trees can be displayed relative to
numbers (e.g. taxon statistics, such as abundance).
The displayed range of colors and sizes can be explicitly defined or automatically genereated.
Various transforamtions can be applied to numbers sizes/colors are mapped to.
Several types of tree layout algorithms from \code{\link{igraph}} can be used.
}
\section{size}{



The size of vertexes, edges, labels, and trees can be mapped to arbitrary numbers.
This is useful for displaying statistics for taxa, such as abundance.
Only the relative size of numbers is used, not the values themeselves.
They can be transformed to make the mapping non-linear using the transformation options.
The range of actual sizes displayed on the graph can be set using the range options.

Accepts a \code{numeric} vector, the same length \code{taxon_id} or a
factor of its length.
}

\section{colors}{


The colors of vertexes, edges, labels, and trees can be mapped to arbitrary numbers.
This is useful for highlighting groups of taxa.
Only the relative size of numbers is used, not the values themeselves.
They can be transformed to make the mapping non-linear using the transformation options.
The range of actual colors displayed on the graph can be set using the range options.

Accepts a vector, the same length \code{taxon_id} or a factor of its length.
If a numeric vector is given, it is mapped to a color scale.
Hex values or color names can be used (e.g. \code{#000000} or \code{"black"}).
}

\section{labels}{


The labels of vertexes, edges, and trees can be added.
Vertex labels are centered over their vertex.
Edge labels are displayed over edges, in the same orientation.
Tree labels are displayed over their tree.

Accepts a vector, the same length \code{taxon_id} or a factor of its length.
}

\section{transformations}{


Before any numbers specified are mapped to color/size, they can be transformed to make
the mapping non-linear. 
Any of the transformations listed below can be used by specifying their name.
A customized function can also be supplied to do the transformation.

\describe{
  \item{"radius"}{Proprotional to radius/diameter of vertex}
  \item{"area"}{circular area; better perceptual accuracy than \code{"radius"}}
  \item{"log10 radius"}{Log base 10 of radius}
  \item{"log2 radius"}{Log base 2 of radius}
  \item{"ln radius"}{Log base e of radius}
  \item{"log10 area"}{Log base 10 of circular area}
  \item{"log2 area"}{Log base 2 of circular area}
  \item{"ln area"}{Log base e of circular area}
}
}

\section{ranges}{


The displayed range of colors and sizes can be explicitly defined or automatically genereated.
Size ranges are specified by supplying a \code{numeric} vector with two values: the minimum and maximum.
The units used should be between 0 and 1, representing the proportion of a dimension of the graph.
Since the dimensions of the graph are determined by layout, and not always square, the value
that \code{1} corresponds to is the square root of the graph area (i.e. the side of a square with 
the same area as the plotted space).
Color ranges can be any number of color values as either HEX codes (e.g. \code{#000000}) or
color names (e.g. \code{"black"}).
}

\section{layout}{


Layouts determine the position of vertexes on the graph.
The are implemented using the \code{\link{igraph}} package.
Any additional arguments passed to \code{plot_taxonomy} are passed to the  \code{\link{igraph}}
function used.
The following \code{character} values are understood:

\describe{
  \item{"automatic"}{Use \code{\link[igraph]{nicely}}. Let \code{\link{igraph}} choose the layout.}
  \item{"reingold-tilford"}{Use \code{\link[igraph]{as_tree}}. A circular tree-like layout.}
  \item{"davidson-harel"}{Use \code{\link[igraph]{with_dh}}. A type of simulated annealing.}
  \item{"gem"}{Use \code{\link[igraph]{with_gem}}. A force-directed layout.}
  \item{"graphopt"}{Use \code{\link[igraph]{with_graphopt}}. A force-directed layout.}
  \item{"mds"}{Use \code{\link[igraph]{with_mds}}. Multidimensional scaling.}
  \item{"fruchterman-reingold"}{Use \code{\link[igraph]{with_fr}}. A force-directed layout.}
  \item{"kamada-kawai"}{Use \code{\link[igraph]{with_kk}}. A layout based on a phyisical model of springs.}
  \item{"large-graph"}{Use \code{\link[igraph]{with_lgl}}. Meant for larger graphs.}
  \item{"drl"}{Use \code{\link[igraph]{with_drl}}. A force-directed layout.}
}
}

