% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_taxonomy}
\alias{plot_taxonomy}
\title{Plot items on taxonomy}
\usage{
plot_taxonomy(taxon_id, parent_id, size = NULL, vertex_color = NULL,
  vertex_label = NULL, line_color = NULL, line_label = NULL,
  overlap_bias = 10, min_label_size = 0.015, line_label_offset = 1,
  margin_size = 0, aspect_raito = 1, data_only = FALSE,
  layout_func = layout.reingold.tilford, layout_args = list(circular =
  TRUE))
}
\arguments{
\item{taxon_id}{The unique ids of the taxon for each row.}

\item{parent_id}{The unique id of supertaxon \code{taxon_id} is a part of.}

\item{size}{The value to base vertex and line size on. If a numeric vector is given, it is rescaled
to an optimum size for display based on minimizing overlaps and spaces beteween adjacent verticies.
See option \code{overlap_bias} for how this optimization is done.}

\item{vertex_color}{The value to base vertex color on. If a numeric vector is given, it is used to
construct a color scale. Hex values or color names can be used (e.g. \code{#000000} or
\code{"black"}).}

\item{vertex_label}{The values of labels over vertcies.}

\item{line_color}{The value to base line color on. If a numeric vector is given, it is used to
construct a color scale. Hex values or color names can be used (e.g. \code{#000000} or
\code{"black"}).}

\item{line_label}{The values of labels over lines.}

\item{overlap_bias}{(\code{numeric} > 0) The factor by which overlaps are punished relative to
spaces when optimizing displayed vertex size range. Only nearby pairs of verticies are considered.
For example, a value of \code{10} would mean that overlaps are considered 10 time more important than
gaps to avoid amoung nearby verticies when choosing an optimum vertex size range.}

\item{min_label_size}{(\code{numeric} of length 1) The minimum label size that will be shown. A
proportion of viewport size. Labels that would be smaller are not added to save time. The type
of text grob used to display the labels that scale with output dimensions is time consuming to
draw, so printing all labels can take a long time. Giving a value of \code{0} will cause all labels
to be printed.}

\item{line_label_offset}{(\code{numeric}) The length the line label is offset from its associated
vertex centerin multiples of vertex radius. For example, a value of \code{1} will place the label on the
edge of the vertex circle and \code{0} would place it in the center.}

\item{margin_size}{(\code{numeric} < 0.5) The amount of space around the plot in terms of the proportion to
the range of space occupied by verticies. For example, a value of \code{.5} would make margins as wide
as the plotted data.}

\item{data}{(\code{data.frame}) Must have columns with names specified by options \code{taxon_id}
and \code{parent_id}.}

\item{aspect_ratio}{(\code{numeric}) The height / width of the plot.}
}
\description{
Plots the distribution of values associated with item with an associated taxonomic classification.
Uses \code{igraph} to make layout and \code{ggplot2} to make plots.
}

