% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_value_distribution_by_level}
\alias{plot_value_distribution_by_level}
\title{plot_value_distribution_by_level}
\usage{
plot_value_distribution_by_level(taxon_data, value_column,
  level_column = "level", ...)
}
\description{
plot_value_distribution_by_level
}

