% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot_value_tree}
\alias{plot_value_tree}
\title{Makes igraph from values}
\usage{
plot_value_tree(graph, values, labels = NA, scaling = 1, exclude = c(),
  root_index = 1, label_color = "black", display = FALSE, fade = FALSE,
  legend_text = "", value_range = c(0, 1), highlight_outliers = TRUE,
  background = "#00000000", save = NULL)
}
\description{
Makes igraph from values
}

