% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heat_tree--shape_generators.R
\name{polygon_coords}
\alias{polygon_coords}
\title{Makes coordinates for a regular polygon}
\usage{
polygon_coords(n = 5, x = 0, y = 0, radius = 1, angle = 0)
}
\arguments{
\item{n}{(\code{numeric} of length 1) The number of nodes in the polygon.}

\item{x}{(\code{numeric} of length 1) x coordinate of center}

\item{y}{(\code{numeric} of length 1) y coordinate of center}

\item{radius}{(\code{numeric} of length 1) The diameter of the circle.}

\item{angle}{(\code{numeric} of length 1) Angle to rotate points around the center of the circle.}
}
\description{
Generates an n x 2 matrix containing x and y coordinates between 1 and 0 for the points of a 
regular polygon.
}
\details{
Inspired by (i.e. stolen from) https://gist.github.com/baptiste/2224724, which was
itself inspired from a post by William Dunlap on r-help (10/09/09)
}
\keyword{internal}
