% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resizing_text_grob.R
\name{postDrawDetails.resizingTextGrob}
\alias{postDrawDetails.resizingTextGrob}
\title{Clean up after the drawing.}
\usage{
postDrawDetails.resizingTextGrob(x)
}
\description{
Clean up after the drawing.
}
\details{
Taken from: http://ryouready.wordpress.com/2012/08/01/creating-a-text-grob-that-automatically-adjusts-to-viewport-size/
}

