% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_documentation.R
\name{pr2_ex_data}
\alias{pr2_ex_data}
\title{Example of PR2 SSU data}
\format{An object of type \code{\link{taxmap}}}
\description{
A dataset containing information from 249 Stramenopile sequences from the PR2 reference database.
}
\examples{
\dontrun{

file_path <- system.file("extdata", "pr2_stramenopiles_gb203.fasta", package = "metacoder")
sequences <- ape::read.FASTA(file_path)
library(taxa) # The parsers in taxa are used
pr2_ex_data <- extract_tax_data(names(sequences),
                                regex = "^(.*\\\\..*?)\\\\|(.*)$",
                                key = c("info", "class"),
                                class_sep = "|")
}

}
\seealso{
Other example_datasets: \code{\link{bryophytes_ex_data}},
  \code{\link{contaminants}},
  \code{\link{genbank_ex_data}},
  \code{\link{its1_ex_data}}, \code{\link{silva_ex_data}},
  \code{\link{unite_ex_data_1}},
  \code{\link{unite_ex_data_2}}
}
