% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/in_silico_pcr.R
\name{primersearch}
\alias{primersearch}
\alias{primersearch.classified}
\alias{primersearch.default}
\title{Use EMBOSS primersearch for in silico PCR}
\usage{
primersearch(...)

\method{primersearch}{default}(input, forward, reverse, seq_name = NULL,
  pair_name = NULL, mismatch = 5, ...)

\method{primersearch}{classified}(input, embed = TRUE, ...)
}
\arguments{
\item{...}{Additional arguments are passed to \code{\link{run_primersearch}}.}

\item{input}{A list of character vectors of DNA sequence, a \code{\link[ape]{DNAbin}} object,
or an object of type \code{\link{classified}}}

\item{forward}{(\code{character} of length 1) The forward primer sequence}

\item{reverse}{(\code{character} of length 1) The reverse primer sequence}

\item{seq_name}{(\code{character}) Names of sequences}

\item{pair_name}{(\code{character} of length 1) The name of the primer pair.}

\item{mismatch}{An integer vector of length 1. The percentage of mismatches allowed.}

\item{embed}{If \code{TRUE}, embed output in input object and return}
}
\value{
An object of type \code{\link{classified}}
}
\description{
A pair of primers are aligned against a set of sequences.
The location of the best hits, quality of match, and predicted amplicons are returned.
Requires the EMBOSS tool kit (\url{http://emboss.sourceforge.net/}) to be installed.
}
\section{Installing EMBOSS}{


The command-line tool "primersearch" from the EMBOSS tool kit is needed to use this function.
How you install EMBOSS will depend on your operating system:

\strong{Linux:}

Open up a terminal and type:

\code{sudo apt-get install emboss}

\strong{Mac OSX:}

The easiest way to install EMBOSS on OSX is to use \href{http://brew.sh/}{homebrew}.
After installing homebrew, open up a terminal and type:

\code{brew install homebrew/science/emboss}

NOTE: This has not been tested by us yet.

\strong{Windows:}

There is an installer for Windows here:

ftp://emboss.open-bio.org/pub/EMBOSS/windows/mEMBOSS-6.5.0.0-setup.exe

NOTE: This has not been tested by us yet.
}
\examples{
\dontrun{
result <- primersearch(rdp_ex_data, 
                       forward = "CAGYMGCCRCGGKAAHACC",
                       reverse = "GGACTACNSGGGTMTCTAAT",
                       pair_name = "U519F_Arch806R",
                       mismatch = 10)
                       
plot(result, 
     vertex_size = item_count,
     vertex_label = name,
     vertex_color = prop_amplified,
     vertex_color_range = c("red", "yellow", "green"),
     vertex_color_trans = "radius",
     vertex_color_interval = c(0, 1),
     layout = "fruchterman-reingold")
}

}

