% Generated by roxygen2 (4.0.2): do not edit by hand
\name{primersearch}
\alias{primersearch}
\title{Use EMBOSS primersearch for in silico PCR}
\usage{
primersearch(sequence, forward, reverse, seq_name = NULL, pair_name = NULL,
  mismatch = 5, ...)
}
\arguments{
\item{sequence}{A list of character vectors of DNA sequence or a file path to FASTA file
containing sequences.}

\item{forward}{A character vector or list of primer sequences. If named and pair_name is not set,
the names are used to construct the primer pair names.}

\item{reverse}{A character vector or list of primer sequences. If named and pair_name is not set,
the names are used to construct the primer pair names.}

\item{seq_name}{Names of sequences}

\item{pair_name}{A character vector of names for primer pairs.}

\item{mismatch}{An integer vector of length 1. The percentage of mismatches allowed.}

\item{...}{Additional arguments are passed to \code{\link{run_primersearch}}.}
}
\value{
Output from \code{\link{parse_primersearch}} (A dataframe)
}
\description{
Use EMBOSS primersearch for in silico PCR
}

