% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classified--class.R
\name{print.classified}
\alias{print.classified}
\title{Print a \code{\link{classified}} object}
\usage{
\method{print}{classified}(x, ...)
}
\arguments{
\item{x}{object to print}

\item{...}{Passed to other methods}
}
\description{
Print a \code{\link{classified}} object
}

