% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxmap--class.R
\name{print.taxmap}
\alias{print.taxmap}
\title{Print a \code{\link{taxmap}} object}
\usage{
\method{print}{taxmap}(x, max_rows = 7, ...)
}
\arguments{
\item{x}{object to print}

\item{max_rows}{(\code{integer} of length 1)
The maximum number of rows to print in tables.}

\item{...}{Not used}
}
\description{
Print a \code{\link{taxmap}} object
}

