% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heat_tree--mapping.R
\name{qualitative_palette}
\alias{qualitative_palette}
\title{The default qualitative color palette}
\usage{
qualitative_palette()
}
\value{
\code{character} of hex color codes
}
\description{
Returns the default color palette for qualitative data
}
\examples{
qualitative_palette()

}
