% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heat_tree--mapping.R
\name{quantative_palette}
\alias{quantative_palette}
\title{The default quantative color palette}
\usage{
quantative_palette()
}
\value{
\code{character} of hex color codes
}
\description{
Returns the default color palette for quantative data.
}
\examples{
quantative_palette()

}
