% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sequence_download.R
\name{query_req_to_dataframe}
\alias{query_req_to_dataframe}
\title{Converts seqinr::qaw to data.frame}
\usage{
query_req_to_dataframe(query_req)
}
\arguments{
\item{query_req}{A list of class \code{\link[seqinr]{SeqAcnucWeb}} or a single
\code{\link[seqinr]{qaw}} object.}
}
\value{
A data frame with columns corresponding to \code{\link[seqinr]{SeqAcnucWeb}} attributes.
}
\description{
Converts a list of class \code{\link[seqinr]{SeqAcnucWeb}} or \code{\link[seqinr]{qaw}} to a
data.frame.
}

