% Generated by roxygen2 (4.0.2): do not edit by hand
\name{query_req_to_dataframe}
\alias{query_req_to_dataframe}
\title{Converts a list of class seqinr::SeqAcnucWeb to a data.frame}
\usage{
query_req_to_dataframe(query_req)
}
\arguments{
\item{query_req}{A list of class seqinr::SeqAcnucWeb}
}
\value{
A data frame with rows named after the sequence names and columns 'length' and 'frame'.
}
\description{
Converts a list of class seqinr::SeqAcnucWeb to a data.frame
}

