% Generated by roxygen2 (4.0.2): do not edit by hand
\name{query_taxon}
\alias{query_taxon}
\title{Queries ACNUC for taxa}
\usage{
query_taxon(taxon, key = NULL, type = NULL, query_id = NULL,
  execute = TRUE, parent = FALSE, ...)
}
\arguments{
\item{taxon}{A character vector of taxa to search for.}

\item{key}{A character vector of keywords. All keywords must be present for a given
sequence to be returned.}

\item{type}{The type of the sequence to return (e.g. RRNA). Use \code{\link[seqinr]{getType}}
to see options.}

\item{query_id}{A character vector of length 1. The name of the query passed to
\code{\link[seqinr]{query}}. The result will be saved in a global variable with this name. If
NULL, a temporary variable is made and deleted upon completion of the function.}

\item{execute}{If FALSE, the query is not executed. Instead the query string will be returned.}

\item{parent}{If TRUE, results are replaced with their parent sequences (i.e. subsequences are
replaced with full sequences). See ACNUC documentation for more info.}
}
\value{
If \code{execute == TRUE} A instance of class \code{\link[seqinr]{qaw}}, as returned by
\code{\link[seqinr]{query}}. Otherwise, a query string will be returned as a character vector.
}
\description{
Produces a \pkg{seqinr} query object that can be used to download sequences with other
\pkg{seqinr} functions. It can also be parsed for result sequence IDs and lengths. Note that
this function requires an open connection to a database provided by
\code{\link[seqinr]{choosebank}} (see examples).
}
\details{
This function uses \code{\link[seqinr]{query}} to search for sequences.
\code{\link[seqinr]{query}} assignes results to a varaible in the global environment as well as
returning the variable. By default, this function avoids this behavior by using a temporary
variable and deleting it on function exit, even if an error is encountered. If you want to let
\code{\link[seqinr]{query}} assign to the global environment, use the \code{query_id} option.
}
\examples{
\dontrun{
choosebank("genbank")
x <- query_taxon(c("Phytophthora", "Pythium"), key=c("@18S@"), type="RRNA")
closebank()}
}
\seealso{
\code{\link[taxize]{ncbi_getbyname}},
  \code{\link[seqinr]{getSequence}},
  \code{\link[seqinr]{query}},
  \code{\link[seqinr]{getType}}
}

