% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reading_files.R
\name{read_fasta}
\alias{read_fasta}
\title{Read a FASTA file}
\usage{
read_fasta(file_path, subset = NULL)
}
\arguments{
\item{file_path}{(\code{character} of length 1) The path to a file to read.}

\item{subset}{(\code{numeric}) Indexes of entries to return.
If not \code{NULL}, the file will first be indexed without loading the whole file into RAM.}
}
\value{
names \code{character}
}
\description{
Read a FASTA file
}

