% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxmap--internal.R
\name{remove_redundant_names}
\alias{remove_redundant_names}
\title{Remove the redundant taxon names}
\usage{
remove_redundant_names(obj, name_col, all_supertaxa = TRUE)
}
\arguments{
\item{obj}{a \code{taxmap} object}

\item{name_col}{(\code{character} of length 1)
The name of a column in \code{obj$taxon_data}}

\item{all_supertaxa}{(\code{logical} of length 1) If \code{TRUE}, check all supertaxa for redundant names instead of just the one immediate supertaxa.}
}
\value{
\code{character}
}
\description{
Remove the names of parent taxa in the begining of their children's names in a \code{taxmap} object.
This is useful for remove genus names in species binomials.
}

