% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{restrict_taxonomy}
\alias{restrict_taxonomy}
\title{Make minimum taxonomy for a subset of taxa}
\usage{
restrict_taxonomy(taxa, parents, subset)
}
\arguments{
\item{taxa}{(\code{character}) Unique taxon IDs for every possible taxon.}

\item{parents}{(\code{character}) Unique taxon IDs for the supertaxa of every possible taxon.
Root taxa should have \code{NA} in this column.}

\item{subset}{(\code{character}) Taxon IDs to restrict the new taxonomy to.}
}
\value{
A vector of unique taxon IDs.
}
\description{
Subsets a table representing every unique taxon based on a subset of taxa. 
This preserves supertaxa, not specified in the subset, but needed to define the internal tree 
structure of the taxonomy.
}

