% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heat_tree--legend.R
\name{scale_bar_coords}
\alias{scale_bar_coords}
\title{Make scale bar division}
\usage{
scale_bar_coords(x1, x2, y1, y2, color, group)
}
\arguments{
\item{x1}{(\code{numeric} of length 1) x of top right}

\item{x2}{(\code{numeric} of length 1) x of bottom right}

\item{y1}{(\code{numeric} of length 1) y of top right}

\item{y2}{(\code{numeric} of length 1) y of bottom right}

\item{color}{}

\item{group}{}
}
\value{
\code{data.frame}
}
\description{
Make scale bar division
}
\keyword{internal}
