% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_documentation.R
\name{silva_ex_data}
\alias{silva_ex_data}
\title{Example dataset from SILVA}
\format{An object of type \code{\link{taxmap}}}
\description{
https://www.arb-silva.de/
}
\examples{
\dontrun{

file_path <- system.file("extdata", "silva_nr99.fasta", package = "metacoder")
sequences <- ape::read.FASTA(file_path)
silva_ex_data <- extract_tax_data(names(sequences),
                                  regex = "^(.*?) (.*)$",
                                  key = c(id = "info", tax = "class"),
                                  class_sep = ";")
}

}
\seealso{
Other example_fastas: \code{\link{bryophytes_ex_data}},
  \code{\link{contaminants}},
  \code{\link{genbank_ex_data}},
  \code{\link{its1_ex_data}}, \code{\link{pr2_ex_data}},
  \code{\link{unite_ex_data_1}},
  \code{\link{unite_ex_data_2}}
}
