% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heat_tree--internal.R
\name{split_by_level}
\alias{split_by_level}
\title{Splits a taxonomy at a specific level or rank}
\usage{
split_by_level(taxa, parents, level, rank = NULL)
}
\arguments{
\item{taxa}{(\code{character}) Unique taxon IDs for every possible taxon.}

\item{parents}{(\code{character}) Unique taxon IDs for the supertaxa of every possible taxon.}

\item{level}{(\code{character} or \code{numeric} of length 1)}

\item{rank}{(\code{character}) The rank designation (e.g. "genus") corresponding to each observation in}
}
\value{
a \code{list} of taxon id \code{character} vectors. 
\code{taxa}.
}
\description{
Breaks one taxonomy into multiple, each with a root of a specified distance from the root.
}
\keyword{internal}
