% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa--classified_class.R
\name{split_by_taxon}
\alias{split_by_taxon}
\title{Split \code{\link{classified}} into individual taxa}
\usage{
split_by_taxon(obj)
}
\arguments{
\item{obj}{(\code{\link{classified}}) The object to split.}
}
\value{
\code{list} of \code{\link{classified}}
}
\description{
Splits an object of type \code{\link{classified}} into a list  of
\code{\link{classified}} objects, one for each taxon in the input.
}

