% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/test_taxon_class.R
\name{[[.classified}
\alias{[[.classified}
\title{Create a restrictive subset of \code{\link{classified}}}
\usage{
\method{[[}{classified}(taxon, item)
}
\arguments{
\item{taxon}{A key to filter the taxon data.frame rows on}

\item{item}{A key to filter the item data.frame rows on}
}
\value{
\code{\link{classified}}
}
\description{
Create a subset of items classified by a taxonomy.
Only specified taxa and shared parent taxa of specified taxa with specified items are preserved.
Only specified items assigned to specified taxa are preserved.
}

