% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa--classified_class.R
\name{subset.classified}
\alias{subset.classified}
\title{Create a inclusive subset of \code{\link{classified}}}
\usage{
\method{subset}{classified}(x, taxon, item, subtaxa = TRUE,
  supertaxa = FALSE, itemless = TRUE, ...)
}
\arguments{
\item{x}{\code{\link{classified}}}

\item{taxon}{A key to filter the taxon data.frame rows on}

\item{item}{A key to filter the item data.frame rows on}

\item{subtaxa}{(\code{logical} of length 1) If \code{TRUE}, return subtaxa of specified taxa}

\item{supertaxa}{(\code{logical} of length 1) If \code{TRUE}, return supertaxa of specified taxa}

\item{itemless}{(\code{logical} of length 1) If \code{TRUE}, return taxa even if they have no items assigned to them}

\item{...}{not used}
}
\value{
\code{\link{classified}}
}
\description{
Create a subset of items classified by a taxonomy.
Only unspecified taxa with no items or children with items are discarded.
}

