% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classified--internal.R
\name{taxon_data_colnames}
\alias{taxon_data_colnames}
\title{Get column names of taxon_data}
\usage{
taxon_data_colnames(obj)
}
\arguments{
\item{obj}{a \code{classified} object}
}
\value{
\code{character}
}
\description{
Get column names of taxon_data without calculating columns
}

