% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxmap--column_generators.R
\name{taxon_levels}
\alias{taxon_levels}
\title{Get rank of taxa}
\usage{
taxon_levels(obj, subset = obj$taxon_data$taxon_ids)
}
\arguments{
\item{obj}{(\code{\link{taxmap}})}

\item{subset}{(\code{character}) The \code{taxon_ids}s to get ranks for.}
}
\value{
\code{numeric}
}
\description{
Get rank of taxa in an object of type \code{\link{taxmap}}
}

