% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_plotting.R
\name{transform_data}
\alias{transform_data}
\title{Transformation functions}
\usage{
transform_data(data = NULL, func = NULL, inverse = FALSE)
}
\arguments{
\item{data}{(\code{numeric}) Data to transform}

\item{func}{(\code{character}) Name of transformation to apply.}

\item{inverse}{If TRUE, return inverse}
}
\description{
Functions used by plotting funtions to transform data.
Calling the function with no parameters returns available function names.
}
\keyword{internal}

