% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heat_tree--internal.R
\name{transform_data}
\alias{transform_data}
\title{Transformation functions}
\usage{
transform_data(func = NULL, data = NULL, inverse = FALSE)
}
\arguments{
\item{func}{(\code{character}) Name of transformation to apply.}

\item{data}{(\code{numeric}) Data to transform}

\item{inverse}{(\code{logical} of length 1) If \code{TRUE}, return the inverse of the selected function.}
}
\description{
Functions used by plotting functions to transform data.
Calling the function with no parameters returns available function names.
Calling with just the function name returns the transformation function
}
\keyword{internal}
