% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_taxa--taxmap--docs.R
\name{transmute_obs}
\alias{transmute_obs}
\title{Replace columns in [taxmap()] objects}
\arguments{
\item{obj}{An object of type [taxmap()]}

\item{data}{Dataset name, index, or a logical vector that indicates which dataset in
`obj$data` to use.}

\item{...}{One or more named columns to add. Newly created columns can be
referenced in the same function call. Any variable name that appears in
[all_names()] can be used as if it was a vector on its own.}

\item{target}{DEPRECIATED. use "data" instead.}
}
\value{
An object of type [taxmap()]
}
\description{
Replace columns of tables in `obj$data` in [taxmap()] objects. See
[dplyr::transmute()] for the inspiration for this function and more
information. Calling the function using the `obj$transmute_obs(...)` style
edits "obj" in place, unlike most R functions. However, calling the function
using the `transmute_obs(obj, ...)` imitates R's traditional copy-on-modify
semantics, so "obj" would not be changed; instead a changed version would be
returned, like most R functions.
\preformatted{
obj$transmute_obs(data, ...)
transmute_obs(obj, data, ...)}
}
\examples{
# Replace columns in a table with new columns
transmute_obs(ex_taxmap, "info", new_col = paste0(name, "!!!"))

}
\seealso{
Other taxmap manipulation functions: 
\code{\link{arrange_obs}()},
\code{\link{arrange_taxa}()},
\code{\link{filter_obs}()},
\code{\link{filter_taxa}()},
\code{\link{mutate_obs}()},
\code{\link{sample_frac_obs}()},
\code{\link{sample_frac_taxa}()},
\code{\link{sample_n_obs}()},
\code{\link{sample_n_taxa}()},
\code{\link{select_obs}()}
}
\concept{taxmap manipulation functions}
