% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_taxonomy--internal.R
\name{validate_regex_key_pair_}
\alias{validate_regex_key_pair}
\alias{validate_regex_key_pair_}
\title{Check a regex-key pair}
\usage{
validate_regex_key_pair_(regex, key, key_options,
  multiple_allowed = key_options)

validate_regex_key_pair(regex, key, key_options,
  multiple_allowed = key_options)
}
\arguments{
\item{regex}{(\code{character})
A regex with capture groups}

\item{key}{(\code{character})
A key corresponding to \code{regex}}

\item{key_options}{(\code{character})
Values \code{key} can take on.}

\item{multiple_allowed}{(\code{character})
Values of \code{key_options} that can appear more than once.}
}
\value{
Returns the result of \code{\link{match.arg}} on the key.
}
\description{
Check that the number of capture groups in the regex matches the length of the key
and that the key consists of valid options.
Uses non-standard evaluation to get the name of input variables.
}
\keyword{internal}

