% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_taxa--taxmap--parsers.R
\name{validate_regex_key_pair}
\alias{validate_regex_key_pair}
\title{Check a regex-key pair}
\usage{
validate_regex_key_pair(regex, key, multiple_allowed)
}
\arguments{
\item{regex}{(\code{character})
A regex with capture groups}

\item{key}{(\code{character})
A key corresponding to \code{regex}}

\item{multiple_allowed}{(\code{character})
Values of \code{key_options} that can appear more than once.}
}
\value{
Returns the result of \code{\link{match.arg}} on the key.
}
\description{
Checks that the number of capture groups in the regex matches the length of the key.
Checks that only certain values of \code{key} can appear more that once.
Adds names to keys that will be used for column names in the output of \code{extract_taxonomy}.
Uses non-standard evaluation to get the name of input variables.
}
\keyword{internal}
