% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heat_tree--internal.R
\name{verify_size}
\alias{verify_size}
\title{Verify size parameters}
\usage{
verify_size(args)
}
\arguments{
\item{args}{(\code{character}) The names of arguments to verify.}
}
\description{
Verify size parameters
}
\keyword{internal}
