% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_taxonomy--internal.R
\name{vigilant_report}
\alias{vigilant_report}
\title{Report a error/warning if needed}
\usage{
vigilant_report(text)
}
\arguments{
\item{text}{The error to report}
}
\value{
\code{NULL}
}
\description{
Report a error/warning if needed
NOTE: This function is unusual in that it looks for a varaible names `vigilance` in a parent namespace.
}
\keyword{internal}
